Operating characteristics for array testing without master pooling

Calculate the expected number of tests and accuracy measures for each individual using array testing without master pooling

Array.Measures(p, se, sp)

Arguments

p

matrix of probabilities corresponding to each individual's risk of disease.

se

the sensitivity of the diagnostic test.

sp

the specificity of the diagnostic test.

Value

A list containing:

T

the expected number of tests for the array.

PSe

a matrix containing each individual's pooling sensitivity, corresponding to the input matrix of individual probabilities.

PSp

a matrix containing each individual's pooling specificity, corresponding to the input matrix of individual probabilities.

PPV

a matrix containing each individual's pooling positive predictive value, corresponding to the input matrix of individual probabilities.

NPV

a matrix containing each individual's pooling negative predictive value, corresponding to the input matrix of individual probabilities.

Details

This function calculates the operating characteristics for array testing without master pooling. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual.

Note

This function returns the pooling positive and negative predictive values for all individuals in the array even though these measures are diagnostic specific; i.e., PPV (NPV) should only be considered for those individuals who have tested positive (negative).

References

McMahan, C., Tebbs, J., Bilder, C. (2012). “Two-Dimensional Informative Array Testing.” Biometrics, 68(3), 793--804. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01726.x , http://www.ncbi.nlm.nih.gov/pubmed/22212007.

See also

Author

This function was originally written by Christopher S. McMahan for McMahan et al. (2012). The function was obtained from http://chrisbilder.com/grouptesting.

Examples

# Calculate the operating characteristics for 
#   non-informative array testing without master
#   pooling, with a 5x5 array and an overall disease 
#   risk of p = 0.02.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
p.mat <- matrix(data=0.02, ncol=5, nrow=5)
Array.Measures(p=p.mat, se=0.95, sp=0.95)
#> $T
#> [1] 12.12849
#> 
#> $PSe
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.9075577 0.9075577 0.9075577 0.9075577 0.9075577
#> [2,] 0.9075577 0.9075577 0.9075577 0.9075577 0.9075577
#> [3,] 0.9075577 0.9075577 0.9075577 0.9075577 0.9075577
#> [4,] 0.9075577 0.9075577 0.9075577 0.9075577 0.9075577
#> [5,] 0.9075577 0.9075577 0.9075577 0.9075577 0.9075577
#> 
#> $PSp
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,] 0.996631 0.996631 0.996631 0.996631 0.996631
#> [2,] 0.996631 0.996631 0.996631 0.996631 0.996631
#> [3,] 0.996631 0.996631 0.996631 0.996631 0.996631
#> [4,] 0.996631 0.996631 0.996631 0.996631 0.996631
#> [5,] 0.996631 0.996631 0.996631 0.996631 0.996631
#> 
#> $PPV
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.8460966 0.8460966 0.8460966 0.8460966 0.8460966
#> [2,] 0.8460966 0.8460966 0.8460966 0.8460966 0.8460966
#> [3,] 0.8460966 0.8460966 0.8460966 0.8460966 0.8460966
#> [4,] 0.8460966 0.8460966 0.8460966 0.8460966 0.8460966
#> [5,] 0.8460966 0.8460966 0.8460966 0.8460966 0.8460966
#> 
#> $NPV
#>           [,1]      [,2]      [,3]      [,4]      [,5]
#> [1,] 0.9981106 0.9981106 0.9981106 0.9981106 0.9981106
#> [2,] 0.9981106 0.9981106 0.9981106 0.9981106 0.9981106
#> [3,] 0.9981106 0.9981106 0.9981106 0.9981106 0.9981106
#> [4,] 0.9981106 0.9981106 0.9981106 0.9981106 0.9981106
#> [5,] 0.9981106 0.9981106 0.9981106 0.9981106 0.9981106
#> 

# Calculate the operating characteristics for 
#   informative array testing without master
#   pooling, with a 3x3 array and an overall disease
#   risk of p = 0.03 and alpha = 2.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(8791)
p.vec <- p.vec.func(p=0.03, alpha=2, grp.sz=9)
p.mat <- Informative.array.prob(prob.vec=p.vec, nr=3, 
nc=3, method="gd")
Array.Measures(p=p.mat, se=0.99, sp=0.99)
#> $T
#> [1] 6.446129
#> 
#> $PSe
#>           [,1]      [,2]      [,3]
#> [1,] 0.9870702 0.9864375 0.9861042
#> [2,] 0.9868357 0.9862030 0.9858697
#> [3,] 0.9866787 0.9860460 0.9857127
#> 
#> $PSp
#>           [,1]      [,2]      [,3]
#> [1,] 0.9997768 0.9997857 0.9998091
#> [2,] 0.9997801 0.9997962 0.9998144
#> [3,] 0.9997887 0.9998038 0.9998176
#> 
#> $PPV
#>           [,1]      [,2]      [,3]
#> [1,] 0.9968340 0.9934381 0.9887994
#> [2,] 0.9956564 0.9923802 0.9852056
#> [3,] 0.9947272 0.9909461 0.9762876
#> 
#> $NPV
#>          [,1]      [,2]      [,3]
#> [1,] 0.999080 0.9995540 0.9997625
#> [2,] 0.999328 0.9996288 0.9998229
#> [3,] 0.999462 0.9996962 0.9998911
#>