Find the optimal testing configuration for informative array testing without master pooling

Find the optimal testing configuration (OTC) for informative array testing without master pooling and calculate the associated operating characteristics.

Inf.Array(p, Se, Sp, group.sz, obj.fn, weights = NULL, alpha = 2)

Arguments

p

the probability of disease, which can be specified as an overall probability of disease, from which a heterogeneous vector of individual probabilities will be generated, or a heterogeneous vector of individual probabilities specified by the user.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single group size (representing the row/column size) for which to calculate the operating characteristics, or a range of group (row/column) sizes over which to find the OTC.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

alpha

a scale parameter for the beta distribution that specifies the degree of heterogeneity for the generated probability vector. If a heterogeneous vector of individual probabilities is specified by the user, alpha can be specified as NA or will be ignored.

Value

A list containing:

prob

the probability of disease, as specified by the user.

alpha

the level of heterogeneity used to generate the vector of individual probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which include:

Array.dim

the row/column size for the first stage of testing.

Array.sz

the overall array size (the square of the row/column size).

p.mat

the sorted matrix of individual probabilities, arranged using the gradient method described by McMahan et al. (2012).

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for informative array testing without master pooling, implemented using the gradient arrangement described in McMahan et al. (2012). Array testing without master pooling involves amalgamating specimens in rows and columns for the first stage of testing. This function uses only square arrays, which is the way array-based group testing is carried out in most real-world applications. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the algorithm. See Hitt et al. (2018) or McMahan et al. (2012) at http://chrisbilder.com/grouptesting for additional details on the implementaion of informative array testing without master pooling.

The value(s) specified by group.sz represent the initial group (row/column) size. If a single value is provided for group.sz, operating characteristics will be calculated and no optimization will be performed. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm, or within the entire array for an array-based algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

McMahan, C., Tebbs, J., Bilder, C. (2012). “Two-Dimensional Informative Array Testing.” Biometrics, 68(3), 793--804. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01726.x , http://www.ncbi.nlm.nih.gov/pubmed/22212007.

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for informative array testing without 
#   master pooling over a range of group (row/column) sizes.
# A vector of individual probabilities is generated using
#   the expected value of order statistics from a beta 
#   distribution with p = 0.03 and a heterogeneity level 
#   of alpha = 2. Depending on the specified probability, 
#   alpha level, and overall group size, simulation may 
#   be necessary in order to generate the vector of individual
#   probabilities. This is done using p.vec.func() and 
#   requires the user to set a seed in order to reproduce 
#   results.
# This examples takes approximately 30 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
set.seed(1002)
Inf.Array(p=0.03, Se=0.99, Sp=0.99, group.sz=3:20, 
obj.fn=c("ET", "MAR"), alpha=2)}

# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(1002)
Inf.Array(p=0.03, Se=0.99, Sp=0.99, group.sz=3:5, 
obj.fn=c("ET", "MAR"), alpha=2)
#> Row/Column Size=3, Array Size=9
#> Row/Column Size=4, Array Size=16
#> Row/Column Size=5, Array Size=25
#> 
#>  Number of minutes running: 0.001533333 
#>  
#> $prob
#> [1] 0.03
#> 
#> $alpha
#> [1] 2
#> 
#> $Se
#> [1] 0.99
#> 
#> $Sp
#> [1] 0.99
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Array.dim
#> [1] 5
#> 
#> $opt.ET$OTC$Array.sz
#> [1] 25
#> 
#> 
#> $opt.ET$p.mat
#>            [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,] 0.08345182 0.04320733 0.02973048 0.02037237 0.012159405
#> [2,] 0.06674511 0.03988263 0.02766518 0.01870106 0.010458408
#> [3,] 0.05792190 0.03695888 0.02571753 0.01706032 0.008658631
#> [4,] 0.05181638 0.03433472 0.02386497 0.01543513 0.006669938
#> [5,] 0.04709310 0.03194154 0.02208870 0.01380852 0.004255942
#> 
#> $opt.ET$ET
#> [1] 11.3266
#> 
#> $opt.ET$value
#> [1] 0.4530641
#> 
#> $opt.ET$PSe
#> [1] 0.980919
#> 
#> $opt.ET$PSp
#> [1] 0.9997594
#> 
#> $opt.ET$PPPV
#> [1] 0.9921313
#> 
#> $opt.ET$PNPV
#> [1] 0.9994101
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Array.dim
#> [1] 5
#> 
#> $opt.MAR$OTC$Array.sz
#> [1] 25
#> 
#> 
#> $opt.MAR$p.mat
#>            [,1]       [,2]       [,3]       [,4]        [,5]
#> [1,] 0.08345182 0.04320733 0.02973048 0.02037237 0.012159405
#> [2,] 0.06674511 0.03988263 0.02766518 0.01870106 0.010458408
#> [3,] 0.05792190 0.03695888 0.02571753 0.01706032 0.008658631
#> [4,] 0.05181638 0.03433472 0.02386497 0.01543513 0.006669938
#> [5,] 0.04709310 0.03194154 0.02208870 0.01380852 0.004255942
#> 
#> $opt.MAR$ET
#> [1] 11.3266
#> 
#> $opt.MAR$value
#> [1] 0.4534295
#> 
#> $opt.MAR$PSe
#> [1] 0.980919
#> 
#> $opt.MAR$PSp
#> [1] 0.9997594
#> 
#> $opt.MAR$PPPV
#> [1] 0.9921313
#> 
#> $opt.MAR$PNPV
#> [1] 0.9994101
#> 
#> 

# Find the OTC for a specified group (row/column) size 
#   for informative array testing without master pooling.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(14849)
Inf.Array(p=p.vec.func(p=0.05, alpha=0.5, grp.sz=100), 
Se=0.95, Sp=0.95, group.sz=10, obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1,1,10,10), nrow=2, ncol=2, byrow=TRUE),
alpha=NA)
#> Row/Column Size=10, Array Size=100
#> 
#>  Number of minutes running: 0.006283333 
#>  
#> $prob
#>   [1] 1.176984e-07 1.480772e-06 9.220764e-06 3.788785e-05 1.155564e-04
#>   [6] 2.790174e-04 5.555128e-04 9.379340e-04 1.370833e-03 1.761814e-03
#>  [11] 2.016332e-03 2.077093e-03 1.946376e-03 1.682031e-03 1.374401e-03
#>  [16] 1.119723e-03 1.002245e-03 1.087789e-03 1.423549e-03 2.036807e-03
#>  [21] 2.928594e-03 4.063807e-03 5.363562e-03 6.706530e-03 7.943269e-03
#>  [26] 8.922576e-03 9.523719e-03 9.685543e-03 9.424022e-03 8.833510e-03
#>  [31] 8.072029e-03 7.335138e-03 6.824866e-03 6.719514e-03 7.148136e-03
#>  [36] 8.171450e-03 9.770080e-03 1.184140e-02 1.420687e-02 1.471727e-02
#>  [41] 1.557426e-02 1.640471e-02 1.726414e-02 1.815772e-02 1.908122e-02
#>  [46] 2.003551e-02 2.102149e-02 2.204015e-02 2.309250e-02 2.417963e-02
#>  [51] 2.530270e-02 2.646294e-02 2.766167e-02 2.890027e-02 3.018023e-02
#>  [56] 3.150316e-02 3.287073e-02 3.428479e-02 3.574726e-02 3.726026e-02
#>  [61] 3.882602e-02 4.044698e-02 4.212574e-02 4.386515e-02 4.566825e-02
#>  [66] 4.753837e-02 4.947912e-02 5.149443e-02 5.358862e-02 5.576638e-02
#>  [71] 5.803290e-02 6.039388e-02 6.285562e-02 6.542512e-02 6.811017e-02
#>  [76] 7.091945e-02 7.386273e-02 7.695104e-02 8.019981e-02 8.361436e-02
#>  [81] 8.721992e-02 9.103232e-02 9.507345e-02 9.936894e-02 1.039491e-01
#>  [86] 1.088502e-01 1.141160e-01 1.198001e-01 1.259690e-01 1.327070e-01
#>  [91] 1.401227e-01 1.483599e-01 1.576142e-01 1.681623e-01 1.804137e-01
#>  [96] 1.950136e-01 2.130686e-01 2.367321e-01 2.711676e-01 3.353406e-01
#> 
#> $alpha
#> [1] NA
#> 
#> $Se
#> [1] 0.95
#> 
#> $Sp
#> [1] 0.95
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Array.dim
#> [1] 10
#> 
#> $opt.ET$OTC$Array.sz
#> [1] 100
#> 
#> 
#> $opt.ET$p.mat
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.3353406 0.13270702 0.08361436 0.05576638 0.03726026 0.02417963
#>  [2,] 0.2711676 0.12596903 0.08019981 0.05358862 0.03574726 0.02309250
#>  [3,] 0.2367321 0.11980010 0.07695104 0.05149443 0.03428479 0.02204015
#>  [4,] 0.2130686 0.11411603 0.07386273 0.04947912 0.03287073 0.02102149
#>  [5,] 0.1950136 0.10885024 0.07091945 0.04753837 0.03150316 0.02003551
#>  [6,] 0.1804137 0.10394913 0.06811017 0.04566825 0.03018023 0.01908122
#>  [7,] 0.1681623 0.09936894 0.06542512 0.04386515 0.02890027 0.01815772
#>  [8,] 0.1576142 0.09507345 0.06285562 0.04212574 0.02766167 0.01726414
#>  [9,] 0.1483599 0.09103232 0.06039388 0.04044698 0.02646294 0.01640471
#> [10,] 0.1401227 0.08721992 0.05803290 0.03882602 0.02530270 0.01557426
#>              [,7]        [,8]        [,9]        [,10]
#>  [1,] 0.014717270 0.008072029 0.002077093 1.087789e-03
#>  [2,] 0.014206867 0.007943269 0.002036807 1.002245e-03
#>  [3,] 0.011841396 0.007335138 0.002016332 9.379340e-04
#>  [4,] 0.009770080 0.007148136 0.001946376 5.555128e-04
#>  [5,] 0.009685543 0.006824866 0.001761814 2.790174e-04
#>  [6,] 0.009523719 0.006719514 0.001682031 1.155564e-04
#>  [7,] 0.009424022 0.006706530 0.001423549 3.788785e-05
#>  [8,] 0.008922576 0.005363562 0.001374401 9.220764e-06
#>  [9,] 0.008833510 0.004063807 0.001370833 1.480772e-06
#> [10,] 0.008171450 0.002928594 0.001119723 1.176984e-07
#> 
#> $opt.ET$ET
#> [1] 34.87659
#> 
#> $opt.ET$value
#> [1] 0.3487659
#> 
#> $opt.ET$PSe
#> [1] 0.8586945
#> 
#> $opt.ET$PSp
#> [1] 0.9945444
#> 
#> $opt.ET$PPPV
#> [1] 0.8920716
#> 
#> $opt.ET$PNPV
#> [1] 0.9925942
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Array.dim
#> [1] 10
#> 
#> $opt.MAR$OTC$Array.sz
#> [1] 100
#> 
#> 
#> $opt.MAR$p.mat
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.3353406 0.13270702 0.08361436 0.05576638 0.03726026 0.02417963
#>  [2,] 0.2711676 0.12596903 0.08019981 0.05358862 0.03574726 0.02309250
#>  [3,] 0.2367321 0.11980010 0.07695104 0.05149443 0.03428479 0.02204015
#>  [4,] 0.2130686 0.11411603 0.07386273 0.04947912 0.03287073 0.02102149
#>  [5,] 0.1950136 0.10885024 0.07091945 0.04753837 0.03150316 0.02003551
#>  [6,] 0.1804137 0.10394913 0.06811017 0.04566825 0.03018023 0.01908122
#>  [7,] 0.1681623 0.09936894 0.06542512 0.04386515 0.02890027 0.01815772
#>  [8,] 0.1576142 0.09507345 0.06285562 0.04212574 0.02766167 0.01726414
#>  [9,] 0.1483599 0.09103232 0.06039388 0.04044698 0.02646294 0.01640471
#> [10,] 0.1401227 0.08721992 0.05803290 0.03882602 0.02530270 0.01557426
#>              [,7]        [,8]        [,9]        [,10]
#>  [1,] 0.014717270 0.008072029 0.002077093 1.087789e-03
#>  [2,] 0.014206867 0.007943269 0.002036807 1.002245e-03
#>  [3,] 0.011841396 0.007335138 0.002016332 9.379340e-04
#>  [4,] 0.009770080 0.007148136 0.001946376 5.555128e-04
#>  [5,] 0.009685543 0.006824866 0.001761814 2.790174e-04
#>  [6,] 0.009523719 0.006719514 0.001682031 1.155564e-04
#>  [7,] 0.009424022 0.006706530 0.001423549 3.788785e-05
#>  [8,] 0.008922576 0.005363562 0.001374401 9.220764e-06
#>  [9,] 0.008833510 0.004063807 0.001370833 1.480772e-06
#> [10,] 0.008171450 0.002928594 0.001119723 1.176984e-07
#> 
#> $opt.MAR$ET
#> [1] 34.87659
#> 
#> $opt.MAR$value
#> [1] 0.3530854
#> 
#> $opt.MAR$PSe
#> [1] 0.8586945
#> 
#> $opt.MAR$PSp
#> [1] 0.9945444
#> 
#> $opt.MAR$PPPV
#> [1] 0.8920716
#> 
#> $opt.MAR$PNPV
#> [1] 0.9925942
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Array.dim
#> [1] 10
#> 
#> $opt.GR1$OTC$Array.sz
#> [1] 100
#> 
#> 
#> $opt.GR1$p.mat
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.3353406 0.13270702 0.08361436 0.05576638 0.03726026 0.02417963
#>  [2,] 0.2711676 0.12596903 0.08019981 0.05358862 0.03574726 0.02309250
#>  [3,] 0.2367321 0.11980010 0.07695104 0.05149443 0.03428479 0.02204015
#>  [4,] 0.2130686 0.11411603 0.07386273 0.04947912 0.03287073 0.02102149
#>  [5,] 0.1950136 0.10885024 0.07091945 0.04753837 0.03150316 0.02003551
#>  [6,] 0.1804137 0.10394913 0.06811017 0.04566825 0.03018023 0.01908122
#>  [7,] 0.1681623 0.09936894 0.06542512 0.04386515 0.02890027 0.01815772
#>  [8,] 0.1576142 0.09507345 0.06285562 0.04212574 0.02766167 0.01726414
#>  [9,] 0.1483599 0.09103232 0.06039388 0.04044698 0.02646294 0.01640471
#> [10,] 0.1401227 0.08721992 0.05803290 0.03882602 0.02530270 0.01557426
#>              [,7]        [,8]        [,9]        [,10]
#>  [1,] 0.014717270 0.008072029 0.002077093 1.087789e-03
#>  [2,] 0.014206867 0.007943269 0.002036807 1.002245e-03
#>  [3,] 0.011841396 0.007335138 0.002016332 9.379340e-04
#>  [4,] 0.009770080 0.007148136 0.001946376 5.555128e-04
#>  [5,] 0.009685543 0.006824866 0.001761814 2.790174e-04
#>  [6,] 0.009523719 0.006719514 0.001682031 1.155564e-04
#>  [7,] 0.009424022 0.006706530 0.001423549 3.788785e-05
#>  [8,] 0.008922576 0.005363562 0.001374401 9.220764e-06
#>  [9,] 0.008833510 0.004063807 0.001370833 1.480772e-06
#> [10,] 0.008171450 0.002928594 0.001119723 1.176984e-07
#> 
#> $opt.GR1$ET
#> [1] 34.87659
#> 
#> $opt.GR1$value
#> [1] 0.3609995
#> 
#> $opt.GR1$PSe
#> [1] 0.8586945
#> 
#> $opt.GR1$PSp
#> [1] 0.9945444
#> 
#> $opt.GR1$PPPV
#> [1] 0.8920716
#> 
#> $opt.GR1$PNPV
#> [1] 0.9925942
#> 
#> 
#> $opt.GR2
#> $opt.GR2$OTC
#> $opt.GR2$OTC$Array.dim
#> [1] 10
#> 
#> $opt.GR2$OTC$Array.sz
#> [1] 100
#> 
#> 
#> $opt.GR2$p.mat
#>            [,1]       [,2]       [,3]       [,4]       [,5]       [,6]
#>  [1,] 0.3353406 0.13270702 0.08361436 0.05576638 0.03726026 0.02417963
#>  [2,] 0.2711676 0.12596903 0.08019981 0.05358862 0.03574726 0.02309250
#>  [3,] 0.2367321 0.11980010 0.07695104 0.05149443 0.03428479 0.02204015
#>  [4,] 0.2130686 0.11411603 0.07386273 0.04947912 0.03287073 0.02102149
#>  [5,] 0.1950136 0.10885024 0.07091945 0.04753837 0.03150316 0.02003551
#>  [6,] 0.1804137 0.10394913 0.06811017 0.04566825 0.03018023 0.01908122
#>  [7,] 0.1681623 0.09936894 0.06542512 0.04386515 0.02890027 0.01815772
#>  [8,] 0.1576142 0.09507345 0.06285562 0.04212574 0.02766167 0.01726414
#>  [9,] 0.1483599 0.09103232 0.06039388 0.04044698 0.02646294 0.01640471
#> [10,] 0.1401227 0.08721992 0.05803290 0.03882602 0.02530270 0.01557426
#>              [,7]        [,8]        [,9]        [,10]
#>  [1,] 0.014717270 0.008072029 0.002077093 1.087789e-03
#>  [2,] 0.014206867 0.007943269 0.002036807 1.002245e-03
#>  [3,] 0.011841396 0.007335138 0.002016332 9.379340e-04
#>  [4,] 0.009770080 0.007148136 0.001946376 5.555128e-04
#>  [5,] 0.009685543 0.006824866 0.001761814 2.790174e-04
#>  [6,] 0.009523719 0.006719514 0.001682031 1.155564e-04
#>  [7,] 0.009424022 0.006706530 0.001423549 3.788785e-05
#>  [8,] 0.008922576 0.005363562 0.001374401 9.220764e-06
#>  [9,] 0.008833510 0.004063807 0.001370833 1.480772e-06
#> [10,] 0.008171450 0.002928594 0.001119723 1.176984e-07
#> 
#> $opt.GR2$ET
#> [1] 34.87659
#> 
#> $opt.GR2$value
#> [1] 0.4711016
#> 
#> $opt.GR2$PSe
#> [1] 0.8586945
#> 
#> $opt.GR2$PSp
#> [1] 0.9945444
#> 
#> $opt.GR2$PPPV
#> [1] 0.8920716
#> 
#> $opt.GR2$PNPV
#> [1] 0.9925942
#> 
#>