Find the optimal testing configuration for informative three-stage hierarchical testing

Find the optimal testing configuration (OTC) for informative three-stage hierarchical testing and calculate the associated operating charcteristics.

Inf.D3(p, Se, Sp, group.sz, obj.fn, weights = NULL, alpha = 2)

Arguments

p

the probability of disease, which can be specified as an overall probability of disease, from which a heterogeneous vector of individual probabilities will be generated, or a heterogeneous vector of individual probabilities specified by the user.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single group size over which to find the OTC out of all possible testing configurations, or a range of group sizes over which to find the OTC.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

alpha

a scale parameter for the beta distribution that specifies the degree of heterogeneity for the generated probability vector. If a heterogeneous vector of individual probabilities is specified by the user, alpha can be specified as NA or will be ignored.

Value

A list containing:

prob

the probability of disease, as specified by the user.

alpha

the level of heterogeneity used to generate the vector of individual probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which include:

Stage1

pool size for the first stage of testing, i.e. the initial group size.

Stage2

pool sizes for the second stage of testing.

p.vec

the sorted vector of individual probabilities.

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for informative three-stage hierarchical testing. This function finds the optimal testing configuration by considering all possible testing configurations. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the algorithm. See Hitt et al. (2018) or Black et al. (2015) at http://chrisbilder.com/grouptesting for additional details on the implementation of informative three-stage hierarchical testing.

The value(s) specified by group.sz represent the initial (stage 1) group size. If a single value is provided for group.sz, the OTC will be found over all possible testing configurations for that initial group size. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm, or within the entire array for an array-based algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Black, M., Bilder, C., Tebbs, J. (2015). “Optimal retesting configurations for hierarchical group testing.” Journal of the Royal Statistical Society. Series C: Applied Statistics, 64(4), 693--710. ISSN 14679876, doi:10.1111/rssc.12097 .

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for informative three-stage hierarchical 
#   testing over a range of group sizes.
# A vector of individual probabilities is generated using
#   the expected value of order statistics from a beta 
#   distribution with p = 0.05 and a heterogeneity level 
#   of alpha = 0.5. Depending on the specified probability, 
#   alpha level, and overall group size, simulation may 
#   be necessary in order to generate the vector of individual
#   probabilities. This is done using p.vec.func() and 
#   requires the user to set a seed in order to reproduce 
#   results.
# This example takes approximately 20 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
set.seed(8318)
Inf.D3(p=0.05, Se=0.99, Sp=0.99, group.sz=3:30, 
obj.fn=c("ET", "MAR"), alpha=0.5)}

# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(8318)
Inf.D3(p=0.05, Se=0.99, Sp=0.99, group.sz=10:15, 
obj.fn=c("ET", "MAR"), alpha=0.5)
#> Initial Group Size = 10 
#> Initial Group Size = 11 
#> Initial Group Size = 12 
#> Initial Group Size = 13 
#> Initial Group Size = 14 
#> Initial Group Size = 15 
#> 
#>  Number of minutes running: 0.007083333 
#>  
#> $prob
#> [1] 0.05
#> 
#> $alpha
#> [1] 0.5
#> 
#> $Se
#> [1] 0.99
#> 
#> $Sp
#> [1] 0.99
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 11
#> 
#> $opt.ET$OTC$Stage2
#> [1] 6 3 1 1
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.001147114 0.003443980 0.007046033 0.012171973 0.019070572 0.028196831
#>  [7] 0.040276923 0.056581560 0.079625310 0.115511702 0.186959325
#> 
#> $opt.ET$ET
#> [1] 3.701757
#> 
#> $opt.ET$value
#> [1] 0.3365234
#> 
#> $opt.ET$PSe
#> [1] 0.9756887
#> 
#> $opt.ET$PSp
#> [1] 0.9987819
#> 
#> $opt.ET$PPPV
#> [1] 0.9768308
#> 
#> $opt.ET$PNPV
#> [1] 0.9987205
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 11
#> 
#> $opt.MAR$OTC$Stage2
#> [1] 6 3 1 1
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.001147114 0.003443980 0.007046033 0.012171973 0.019070572 0.028196831
#>  [7] 0.040276923 0.056581560 0.079625310 0.115511702 0.186959325
#> 
#> $opt.MAR$ET
#> [1] 3.701757
#> 
#> $opt.MAR$value
#> [1] 0.3373238
#> 
#> $opt.MAR$PSe
#> [1] 0.9756887
#> 
#> $opt.MAR$PSp
#> [1] 0.9987819
#> 
#> $opt.MAR$PPPV
#> [1] 0.9768308
#> 
#> $opt.MAR$PNPV
#> [1] 0.9987205
#> 
#> 

# Find the OTC out of all possible testing configurations
#   for a specified group size and vector of individual 
#   probabilities.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(1216)
p.vec <- p.vec.func(p=0.10, alpha=2, grp.sz=12)
Inf.D3(p=p.vec, Se=0.99, Sp=0.99, group.sz=12,
obj.fn=c("ET", "MAR", "GR"), weights=matrix(data=c(1,1), 
nrow=1, ncol=2, byrow=TRUE), alpha=NA)
#> Initial Group Size = 12 
#> 
#>  Number of minutes running: 0.0008166667 
#>  
#> $prob
#>  [1] 0.02235316 0.03583773 0.04759180 0.05888261 0.07028747 0.08222831
#>  [7] 0.09513816 0.10958358 0.12645431 0.14741427 0.17639045 0.22783816
#> 
#> $alpha
#> [1] NA
#> 
#> $Se
#> [1] 0.99
#> 
#> $Sp
#> [1] 0.99
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 12
#> 
#> $opt.ET$OTC$Stage2
#> [1] 5 3 3 1
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.02235316 0.03583773 0.04759180 0.05888261 0.07028747 0.08222831
#>  [7] 0.09513816 0.10958358 0.12645431 0.14741427 0.17639045 0.22783816
#> 
#> $opt.ET$ET
#> [1] 6.884909
#> 
#> $opt.ET$value
#> [1] 0.5737424
#> 
#> $opt.ET$PSe
#> [1] 0.9721599
#> 
#> $opt.ET$PSp
#> [1] 0.9976432
#> 
#> $opt.ET$PPPV
#> [1] 0.9786471
#> 
#> $opt.ET$PNPV
#> [1] 0.9969089
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 12
#> 
#> $opt.MAR$OTC$Stage2
#> [1] 5 3 3 1
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.02235316 0.03583773 0.04759180 0.05888261 0.07028747 0.08222831
#>  [7] 0.09513816 0.10958358 0.12645431 0.14741427 0.17639045 0.22783816
#> 
#> $opt.MAR$ET
#> [1] 6.884909
#> 
#> $opt.MAR$value
#> [1] 0.5765705
#> 
#> $opt.MAR$PSe
#> [1] 0.9721599
#> 
#> $opt.MAR$PSp
#> [1] 0.9976432
#> 
#> $opt.MAR$PPPV
#> [1] 0.9786471
#> 
#> $opt.MAR$PNPV
#> [1] 0.9969089
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Stage1
#> [1] 12
#> 
#> $opt.GR1$OTC$Stage2
#> [1] 5 3 3 1
#> 
#> 
#> $opt.GR1$p.vec
#>  [1] 0.02235316 0.03583773 0.04759180 0.05888261 0.07028747 0.08222831
#>  [7] 0.09513816 0.10958358 0.12645431 0.14741427 0.17639045 0.22783816
#> 
#> $opt.GR1$ET
#> [1] 6.884909
#> 
#> $opt.GR1$value
#> [1] 0.5786475
#> 
#> $opt.GR1$PSe
#> [1] 0.9721599
#> 
#> $opt.GR1$PSp
#> [1] 0.9976432
#> 
#> $opt.GR1$PPPV
#> [1] 0.9786471
#> 
#> $opt.GR1$PNPV
#> [1] 0.9969089
#> 
#>