Find the optimal testing configuration for informative two-stage hierarchical (Dorfman) testing

Find the optimal testing configuration (OTC) for informative two-stage hierarchical (Dorfman) testing and calculate the associated operating characteristics.

Inf.Dorf(p, Se, Sp, group.sz, obj.fn, weights = NULL, alpha = 2)

Arguments

p

the probability of disease, which can be specified as an overall probability of disease, from which a heterogeneous vector of individual probabilities will be generated, or a heterogeneous vector of individual probabilities specified by the user.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single block size for which to find the OTC out of all possible configurations, or a range of block sizes over which to find the OTC.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

alpha

a scale parameter for the beta distribution that specifies the degree of heterogeneity for the generated probability vector. If a heterogeneous vector of individual probabilities is specified by the user, alpha can be specified as NA or will be ignored.

Value

A list containing:

prob

the probability of disease, as specified by the user.

alpha

the level of heterogeneity used to generate the vector of individual probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which include:

Block.sz

the block size/overall group size, which is not tested.

pool.szs

pool sizes for the first stage of testing.

p.vec

the sorted vector of individual probabilities.

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for informative two-stage hierarchical (Dorfman) testing, implemented via the pool-specific optimal Dorfman (PSOD) method described in McMahan et al. (2012). This function finds the optimal testing configuration by considering all possible testing configurations instead of using the greedy algorithm proposed for PSOD testing. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the algorithm. See Hitt et al. (2018) or McMahan et al. (2012) at http://chrisbilder.com/grouptesting for additional details on the implementation of informative two-stage hierarchical (Dorfman) testing.

The value(s) specified by group.sz represent the overall block size used in the pool-specific optimal Dorfman (PSOD) method, where the overall group size is not tested. Instead, multiple initial pool sizes within this block are found and tested in the first stage of testing. The second stage of testing consists of individual retesting. For more details on informative two-stage hierarchical testing implemented via the PSOD method, see Hitt et al. (2018) and McMahan et al. (2012).

If a single value is provided for group.sz, the OTC will be found over all possible testing configurations. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm, or within the entire array for an array-based algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Dorfman, R. (1943). “The Detection of Defective Members of Large Populations.” The Annals of Mathematical Statistics, 14(4), 436--440. ISSN 0003-4851, doi:10.1214/aoms/1177731363 , https://www.jstor.org/stable/2235930.

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

McMahan, C., Tebbs, J., Bilder, C. (2012). “Informative Dorfman Screening.” Biometrics, 68(1), 287--296. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01644.x , http://www.ncbi.nlm.nih.gov/pubmed/21762119.

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for informative two-stage hierarchical 
#   (Dorfman) testing.
# A vector of individual probabilities is generated using
#   the expected value of order statistics from a beta 
#   distribution with p = 0.01 and a heterogeneity level 
#   of alpha = 2. Depending on the specified probability, 
#   alpha level, and overall group size, simulation may 
#   be necessary in order to generate the vector of individual
#   probabilities. This is done using p.vec.func() and 
#   requires the user to set a seed in order to reproduce 
#   results.
# This example takes approximately 20 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
set.seed(9245)
Inf.Dorf(p=0.01, Se=0.95, Sp=0.95, group.sz=3:30, 
obj.fn=c("ET", "MAR"), alpha=2)}

# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(9245)
Inf.Dorf(p=0.01, Se=0.95, Sp=0.95, group.sz=5:10, 
obj.fn=c("ET", "MAR"), alpha=2)
#> Block Size = 5 
#> Block Size = 6 
#> Block Size = 7 
#> Block Size = 8 
#> Block Size = 9 
#> Block Size = 10 
#> 
#>  Number of minutes running: 0.001433333 
#>  
#> $prob
#> [1] 0.01
#> 
#> $alpha
#> [1] 2
#> 
#> $Se
#> [1] 0.95
#> 
#> $Sp
#> [1] 0.95
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Block.sz
#> [1] 10
#> 
#> $opt.ET$OTC$pool.szs
#> [1] 6 4
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.002344009 0.003817495 0.005151342 0.006486187 0.007900785 0.009473125
#>  [7] 0.011313952 0.013625938 0.016894213 0.022992953
#> 
#> $opt.ET$ET
#> [1] 2.915163
#> 
#> $opt.ET$value
#> [1] 0.2915163
#> 
#> $opt.ET$PSe
#> [1] 0.9025
#> 
#> $opt.ET$PSp
#> [1] 0.9958578
#> 
#> $opt.ET$PPPV
#> [1] 0.6875764
#> 
#> $opt.ET$PNPV
#> [1] 0.999012
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Block.sz
#> [1] 10
#> 
#> $opt.MAR$OTC$pool.szs
#> [1] 6 4
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.002344009 0.003817495 0.005151342 0.006486187 0.007900785 0.009473125
#>  [7] 0.011313952 0.013625938 0.016894213 0.022992953
#> 
#> $opt.MAR$ET
#> [1] 2.915163
#> 
#> $opt.MAR$value
#> [1] 0.2930035
#> 
#> $opt.MAR$PSe
#> [1] 0.9025
#> 
#> $opt.MAR$PSp
#> [1] 0.9958578
#> 
#> $opt.MAR$PPPV
#> [1] 0.6875764
#> 
#> $opt.MAR$PNPV
#> [1] 0.999012
#> 
#> 

# Find the OTC for informative two-stage hierarchical 
#   (Dorfman) testing, for a specified block size.
# This example uses rbeta() to generate random probabilities 
#   and requires the user to set a seed in order to reproduce 
#   results.
# This example takes approximately 2.5 minutes to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
set.seed(8791)
Inf.Dorf(p=p.vec.func(p=0.03, alpha=0.5, grp.sz=50), 
Se=0.90, Sp=0.90, group.sz=50, obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1,1,10,10), nrow=2, ncol=2, byrow=TRUE),
alpha=NA)}