Operating characteristics for array testing with master pooling

Calculate the expected number of tests and accuracy measures for each individual using array testing with master pooling.

MasterPool.Array.Measures(results, n, pmat, Se, Sp)

Arguments

results

an object containing results (expected number of tests and accuracy measures) from Array.Measures.

n

size of a row/column in the square array.

pmat

matrix of individual risk probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

Value

A list containing:

ET

the expected number of tests for the array.

PSe

a matrix containing each individual's pooling sensitivity, corresponding to the input matrix of individual probabilities.

PSp

a matrix containing each individual's pooling specificity, corresponding to the input matrix of individual probabilities.

PPV

a matrix containing each individual's pooling positive predictive value, corresponding to the input matrix of individual probabilities.

NPV

a matrix containing each individual's pooling negative predictive value, corresponding to the input matrix of individual probabilities.

Details

This function assumes that the array is square (i.e., the row and column size are equal) and utilizes the equations from Kim et al. (2007) for square array testing with master pooling. This function calculates the operating characteristics for array testing with master pooling. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual.

Note

This function returns the pooling positive and negative predictive values for all individuals in the array even though these measures are diagnostic specific; i.e., PPV (NPV) should only be considered for those individuals who have tested positive (negative).

References

Kim, H., Hudgens, M., Dreyfuss, J., Westreich, D., Pilcher, C. (2007). “Comparison of group testing algorithms for case identification in the presence of test error.” Biometrics, 63(4), 1152--1163. ISSN 0006341X, doi:10.1111/j.1541-0420.2007.00817.x , http://www.ncbi.nlm.nih.gov/pubmed/17501946.

See also

Author

Brianna D. Hitt

Examples

# Calculate the operating characteristics for 
#   non-informative array testing with master
#   pooling, with a 6x6 array and an overall 
#   disease risk of p = 0.10.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
p.mat <- matrix(data=0.10, ncol=6, nrow=6)
results <- Array.Measures(p=p.mat, se=0.90, sp=0.90)
MasterPool.Array.Measures(results=results, n=36, 
pmat=p.mat, Se=0.90, Sp=0.90)
#> $ET
#> [1] 19.01143
#> 
#> $PSe
#>        [,1]   [,2]   [,3]   [,4]   [,5]   [,6]
#> [1,] 0.6561 0.6561 0.6561 0.6561 0.6561 0.6561
#> [2,] 0.6561 0.6561 0.6561 0.6561 0.6561 0.6561
#> [3,] 0.6561 0.6561 0.6561 0.6561 0.6561 0.6561
#> [4,] 0.6561 0.6561 0.6561 0.6561 0.6561 0.6561
#> [5,] 0.6561 0.6561 0.6561 0.6561 0.6561 0.6561
#> [6,] 0.6561 0.6561 0.6561 0.6561 0.6561 0.6561
#> 
#> $PSp
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
#> [1,] 0.930308 0.930308 0.930308 0.930308 0.930308 0.930308
#> [2,] 0.930308 0.930308 0.930308 0.930308 0.930308 0.930308
#> [3,] 0.930308 0.930308 0.930308 0.930308 0.930308 0.930308
#> [4,] 0.930308 0.930308 0.930308 0.930308 0.930308 0.930308
#> [5,] 0.930308 0.930308 0.930308 0.930308 0.930308 0.930308
#> [6,] 0.930308 0.930308 0.930308 0.930308 0.930308 0.930308
#> 
#> $PPV
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489
#> [2,] 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489
#> [3,] 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489
#> [4,] 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489
#> [5,] 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489
#> [6,] 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489 0.5112489
#> 
#> $NPV
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469
#> [2,] 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469
#> [3,] 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469
#> [4,] 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469
#> [5,] 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469
#> [6,] 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469 0.9605469
#>