Find the optimal testing configuration for non-informative array testing without master pooling

Find the optimal testing configuration (OTC) for non-informative array testing without master pooling and calculate the associated operating characteristics.

NI.Array(p, Se, Sp, group.sz, obj.fn, weights = NULL)

Arguments

p

the probability of disease, which can be specified as an overall probability of disease or a homogeneous vector of individual probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single group size (representing the row/column size) for which to calculate the operating characteristics, or a range of group (row/column) sizes over which to find the OTC. The details of group size specification are given under 'Details'.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

Value

A list containing:

prob

the probability of disease, as specified by the user.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which include:

Array.dim

the row/column size for the first stage of testing.

Array.sz

the overall array size (the square of the row/column size).

p.mat

the matrix of individual probabilities.

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for non-informative array testing without master pooling. Array testing without master pooling involves amalgamating specimens in rows and columns for the first stage of testing. This function uses only square arrays, which is the way array-based group testing is carried out in most real-world applications. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the algorithm. See Hitt et al. (2018) at http://chrisbilder.com/grouptesting or Kim et al. (2007) for additional details on the implementation of non-informative array testing without master pooling.

The value(s) specified by group.sz represent the initial group (row/column) size. If a single value is provided for group.sz, operating characteristics will be calculated and no optimization will be performed. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Kim, H., Hudgens, M., Dreyfuss, J., Westreich, D., Pilcher, C. (2007). “Comparison of group testing algorithms for case identification in the presence of test error.” Biometrics, 63(4), 1152--1163. ISSN 0006341X, doi:10.1111/j.1541-0420.2007.00817.x , http://www.ncbi.nlm.nih.gov/pubmed/17501946.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for non-informative array testing 
#   without master pooling over a range of group
#   (row/column) sizes.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
NI.Array(p=0.04, Se=0.95, Sp=0.95, group.sz=3:10,
obj.fn=c("ET", "MAR"))
#> Row/Column Size=3, Array Size=9
#> Row/Column Size=4, Array Size=16
#> Row/Column Size=5, Array Size=25
#> Row/Column Size=6, Array Size=36
#> Row/Column Size=7, Array Size=49
#> Row/Column Size=8, Array Size=64
#> Row/Column Size=9, Array Size=81
#> Row/Column Size=10, Array Size=100
#> 
#>  Number of minutes running: 0.01608333 
#>  
#> $prob
#> [1] 0.04
#> 
#> $Se
#> [1] 0.95
#> 
#> $Sp
#> [1] 0.95
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Array.dim
#> [1] 10
#> 
#> $opt.ET$OTC$Array.sz
#> [1] 100
#> 
#> 
#> $opt.ET$p.mat
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [2,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [3,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [4,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [5,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [6,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [7,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [8,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [9,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#> [10,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#> 
#> $opt.ET$ET
#> [1] 34.06069
#> 
#> $opt.ET$value
#> [1] 0.3406069
#> 
#> $opt.ET$PSe
#> [1] 0.859202
#> 
#> $opt.ET$PSp
#> [1] 0.9945609
#> 
#> $opt.ET$PPPV
#> [1] 0.8681092
#> 
#> $opt.ET$PNPV
#> [1] 0.9941359
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Array.dim
#> [1] 10
#> 
#> $opt.MAR$OTC$Array.sz
#> [1] 100
#> 
#> 
#> $opt.MAR$p.mat
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#>  [1,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [2,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [3,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [4,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [5,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [6,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [7,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [8,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#>  [9,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#> [10,] 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04 0.04  0.04
#> 
#> $opt.MAR$ET
#> [1] 34.06069
#> 
#> $opt.MAR$value
#> [1] 0.3443443
#> 
#> $opt.MAR$PSe
#> [1] 0.859202
#> 
#> $opt.MAR$PSp
#> [1] 0.9945609
#> 
#> $opt.MAR$PPPV
#> [1] 0.8681092
#> 
#> $opt.MAR$PNPV
#> [1] 0.9941359
#> 
#> 

# Calculate the operating characteristics for a specified 
#   group (row/column) size for non-informative array 
#   testing without master pooling.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
NI.Array(p=rep(0.01, 64), Se=0.90, Sp=0.90, group.sz=8,
obj.fn=c("ET", "MAR", "GR"), 
weights=matrix(data=c(1,1,10,10,100,100), 
nrow=3, ncol=2, byrow=TRUE))
#> Row/Column Size=8, Array Size=64
#> 
#>  Number of minutes running: 0.00225 
#>  
#> $prob
#>  [1] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [16] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [31] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [46] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [61] 0.01 0.01 0.01 0.01
#> 
#> $Se
#> [1] 0.9
#> 
#> $Sp
#> [1] 0.9
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Array.dim
#> [1] 8
#> 
#> $opt.ET$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.ET$p.mat
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.ET$ET
#> [1] 21.38367
#> 
#> $opt.ET$value
#> [1] 0.3341199
#> 
#> $opt.ET$PSe
#> [1] 0.7760905
#> 
#> $opt.ET$PSp
#> [1] 0.9923741
#> 
#> $opt.ET$PPPV
#> [1] 0.5068984
#> 
#> $opt.ET$PNPV
#> [1] 0.9977261
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Array.dim
#> [1] 8
#> 
#> $opt.MAR$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.MAR$p.mat
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.MAR$ET
#> [1] 21.38367
#> 
#> $opt.MAR$value
#> [1] 0.3374229
#> 
#> $opt.MAR$PSe
#> [1] 0.7760905
#> 
#> $opt.MAR$PSp
#> [1] 0.9923741
#> 
#> $opt.MAR$PPPV
#> [1] 0.5068984
#> 
#> $opt.MAR$PNPV
#> [1] 0.9977261
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Array.dim
#> [1] 8
#> 
#> $opt.GR1$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.GR1$p.mat
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.GR1$ET
#> [1] 21.38367
#> 
#> $opt.GR1$value
#> [1] 0.3439087
#> 
#> $opt.GR1$PSe
#> [1] 0.7760905
#> 
#> $opt.GR1$PSp
#> [1] 0.9923741
#> 
#> $opt.GR1$PPPV
#> [1] 0.5068984
#> 
#> $opt.GR1$PNPV
#> [1] 0.9977261
#> 
#> 
#> $opt.GR2
#> $opt.GR2$OTC
#> $opt.GR2$OTC$Array.dim
#> [1] 8
#> 
#> $opt.GR2$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.GR2$p.mat
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.GR2$ET
#> [1] 21.38367
#> 
#> $opt.GR2$value
#> [1] 0.4320075
#> 
#> $opt.GR2$PSe
#> [1] 0.7760905
#> 
#> $opt.GR2$PSp
#> [1] 0.9923741
#> 
#> $opt.GR2$PPPV
#> [1] 0.5068984
#> 
#> $opt.GR2$PNPV
#> [1] 0.9977261
#> 
#> 
#> $opt.GR3
#> $opt.GR3$OTC
#> $opt.GR3$OTC$Array.dim
#> [1] 8
#> 
#> $opt.GR3$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.GR3$p.mat
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
#> [1,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [2,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [3,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [4,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [5,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [6,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [7,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> [8,] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.GR3$ET
#> [1] 21.38367
#> 
#> $opt.GR3$value
#> [1] 1.312996
#> 
#> $opt.GR3$PSe
#> [1] 0.7760905
#> 
#> $opt.GR3$PSp
#> [1] 0.9923741
#> 
#> $opt.GR3$PPPV
#> [1] 0.5068984
#> 
#> $opt.GR3$PNPV
#> [1] 0.9977261
#> 
#>