Find the optimal testing configuration for non-informative three-stage hierarchical testing

Find the optimal testing configuration (OTC) for non-informative three-stage hierarchical testing and calculate the associated operating characteristics.

NI.D3(p, Se, Sp, group.sz, obj.fn, weights = NULL)

Arguments

p

the probability of disease, which can be specified as an overall probability of disease or a homogeneous vector of individual probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single group size over which to find the OTC out of all possible testing configurations, or a range of group sizes over which to find the OTC.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

Value

A list containing:

prob

the probability of disease, as specified by the user.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which include:

Stage1

pool size for the first stage of testing, i.e. the initial group size.

Stage2

pool sizes for the second stage of testing.

p.vec

the vector of individual probabilities

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for non-informative three-stage hierarchical testing, by considering all possible testing configurations. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the algorithm. See Hitt et al. (2018) at http://chrisbilder.com/grouptesting or Kim et al. (2007) for additional details on the implementation of non-informative three-stage hierarchical testing.

The value(s) specified by group.sz represent the initial (stage 1) group size. If a single value is provided for group.sz, the OTC will be found over all possible testing configurations for that initial group size. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Kim, H., Hudgens, M., Dreyfuss, J., Westreich, D., Pilcher, C. (2007). “Comparison of group testing algorithms for case identification in the presence of test error.” Biometrics, 63(4), 1152--1163. ISSN 0006341X, doi:10.1111/j.1541-0420.2007.00817.x , http://www.ncbi.nlm.nih.gov/pubmed/17501946.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for non-informative three-stage 
#   hierarchical testing over a range of group sizes.
# This example takes approximately 20 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
NI.D3(p=0.02, Se=0.90, Sp=0.90, group.sz=3:30, 
obj.fn=c("ET", "MAR"))}

# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
NI.D3(p=0.02, Se=0.90, Sp=0.90, group.sz=3:12, 
obj.fn=c("ET", "MAR"))
#> Initial Group Size = 3 
#> Initial Group Size = 4 
#> Initial Group Size = 5 
#> Initial Group Size = 6 
#> Initial Group Size = 7 
#> Initial Group Size = 8 
#> Initial Group Size = 9 
#> Initial Group Size = 10 
#> Initial Group Size = 11 
#> Initial Group Size = 12 
#> 
#>  Number of minutes running: 0.00265 
#>  
#> $prob
#> [1] 0.02
#> 
#> $Se
#> [1] 0.9
#> 
#> $Sp
#> [1] 0.9
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 12
#> 
#> $opt.ET$OTC$Stage2
#> [1] 4 4 4
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02
#> 
#> $opt.ET$ET
#> [1] 2.814091
#> 
#> $opt.ET$value
#> [1] 0.2345076
#> 
#> $opt.ET$PSe
#> [1] 0.729
#> 
#> $opt.ET$PSp
#> [1] 0.9931717
#> 
#> $opt.ET$PPPV
#> [1] 0.6854158
#> 
#> $opt.ET$PNPV
#> [1] 0.9944622
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 12
#> 
#> $opt.MAR$OTC$Stage2
#> [1] 4 4 4
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02 0.02
#> 
#> $opt.MAR$ET
#> [1] 2.814091
#> 
#> $opt.MAR$value
#> [1] 0.2373827
#> 
#> $opt.MAR$PSe
#> [1] 0.729
#> 
#> $opt.MAR$PSp
#> [1] 0.9931717
#> 
#> $opt.MAR$PPPV
#> [1] 0.6854158
#> 
#> $opt.MAR$PNPV
#> [1] 0.9944622
#> 
#> 

# Find the OTC out of all possible configurations for 
#   a specified group size for non-informative
#   three-stage hierarchical testing.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
NI.D3(p=rep(0.005, 15), Se=0.99, Sp=0.99, group.sz=15,
obj.fn=c("ET", "MAR", "GR"), weights=matrix(data=c(1,1,10,10), 
nrow=2, ncol=2, byrow=TRUE))
#> Initial Group Size = 15 
#> 
#>  Number of minutes running: 0.002483333 
#>  
#> $prob
#>  [1] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [13] 0.005 0.005 0.005
#> 
#> $Se
#> [1] 0.99
#> 
#> $Sp
#> [1] 0.99
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 15
#> 
#> $opt.ET$OTC$Stage2
#> [1] 4 4 4 3
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [13] 0.005 0.005 0.005
#> 
#> $opt.ET$ET
#> [1] 1.610626
#> 
#> $opt.ET$value
#> [1] 0.107375
#> 
#> $opt.ET$PSe
#> [1] 0.970299
#> 
#> $opt.ET$PSp
#> [1] 0.9998572
#> 
#> $opt.ET$PPPV
#> [1] 0.9715386
#> 
#> $opt.ET$PNPV
#> [1] 0.9998507
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 15
#> 
#> $opt.MAR$OTC$Stage2
#> [1] 4 4 4 3
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [13] 0.005 0.005 0.005
#> 
#> $opt.MAR$ET
#> [1] 1.610626
#> 
#> $opt.MAR$value
#> [1] 0.1074063
#> 
#> $opt.MAR$PSe
#> [1] 0.970299
#> 
#> $opt.MAR$PSp
#> [1] 0.9998572
#> 
#> $opt.MAR$PPPV
#> [1] 0.9715386
#> 
#> $opt.MAR$PNPV
#> [1] 0.9998507
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Stage1
#> [1] 15
#> 
#> $opt.GR1$OTC$Stage2
#> [1] 4 4 4 3
#> 
#> 
#> $opt.GR1$p.vec
#>  [1] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [13] 0.005 0.005 0.005
#> 
#> $opt.GR1$ET
#> [1] 1.610626
#> 
#> $opt.GR1$value
#> [1] 0.1076657
#> 
#> $opt.GR1$PSe
#> [1] 0.970299
#> 
#> $opt.GR1$PSp
#> [1] 0.9998572
#> 
#> $opt.GR1$PPPV
#> [1] 0.9715386
#> 
#> $opt.GR1$PNPV
#> [1] 0.9998507
#> 
#> 
#> $opt.GR2
#> $opt.GR2$OTC
#> $opt.GR2$OTC$Stage1
#> [1] 15
#> 
#> $opt.GR2$OTC$Stage2
#> [1] 4 4 4 3
#> 
#> 
#> $opt.GR2$p.vec
#>  [1] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [13] 0.005 0.005 0.005
#> 
#> $opt.GR2$ET
#> [1] 1.610626
#> 
#> $opt.GR2$value
#> [1] 0.1102813
#> 
#> $opt.GR2$PSe
#> [1] 0.970299
#> 
#> $opt.GR2$PSp
#> [1] 0.9998572
#> 
#> $opt.GR2$PPPV
#> [1] 0.9715386
#> 
#> $opt.GR2$PNPV
#> [1] 0.9998507
#> 
#>