Find the optimal testing configuration for non-informative two-stage hierarchical testing

Find the optimal testing configuration (OTC) for non-informative two-stage hierarchical (Dorfman) testing and calculate the associated operating characteristics.

NI.Dorf(p, Se, Sp, group.sz, obj.fn, weights = NULL)

Arguments

p

the probability of disease, which can be specified as an overall probability of disease or a homogeneous vector of individual probabilities.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single group size for which to calculate operating characteristics, or a range of group sizes over which to find the OTC. The details of group size specification are given under 'Details'.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

Value

A list containing:

prob

the probability of disease, as specified by the user.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which include:

Stage1

pool size for the first stage of testing, i.e. the initial group size.

p.vec

the vector of individual probabilities.

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for non-informative two-stage hierarchical (Dorfman) testing. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for the algorithm. See Hitt et al. (2018) at http://chrisbilder.com/grouptesting, Dorfman (1943), or Kim et al. (2007) for additional details on the implementation of non-informative two-stage hierarchical testing.

The value(s) specified by group.sz represent the initial (stage 1) group size. If a single value is provided for group.sz, operating characteristics will be calculated and no optimization will be performed. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Dorfman, R. (1943). “The Detection of Defective Members of Large Populations.” The Annals of Mathematical Statistics, 14(4), 436--440. ISSN 0003-4851, doi:10.1214/aoms/1177731363 , https://www.jstor.org/stable/2235930.

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Kim, H., Hudgens, M., Dreyfuss, J., Westreich, D., Pilcher, C. (2007). “Comparison of group testing algorithms for case identification in the presence of test error.” Biometrics, 63(4), 1152--1163. ISSN 0006341X, doi:10.1111/j.1541-0420.2007.00817.x , http://www.ncbi.nlm.nih.gov/pubmed/17501946.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for non-informative two-stage 
#   hierarchical (Dorfman) testing over a range 
#   of group sizes.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
NI.Dorf(p=0.01, Se=0.95, Sp=0.95, group.sz=2:100, 
obj.fn=c("ET", "MAR"))
#> Initial Group Size = 2 
#> Initial Group Size = 3 
#> Initial Group Size = 4 
#> Initial Group Size = 5 
#> Initial Group Size = 6 
#> Initial Group Size = 7 
#> Initial Group Size = 8 
#> Initial Group Size = 9 
#> Initial Group Size = 10 
#> Initial Group Size = 11 
#> Initial Group Size = 12 
#> Initial Group Size = 13 
#> Initial Group Size = 14 
#> Initial Group Size = 15 
#> Initial Group Size = 16 
#> Initial Group Size = 17 
#> Initial Group Size = 18 
#> Initial Group Size = 19 
#> Initial Group Size = 20 
#> Initial Group Size = 21 
#> Initial Group Size = 22 
#> Initial Group Size = 23 
#> Initial Group Size = 24 
#> Initial Group Size = 25 
#> Initial Group Size = 26 
#> Initial Group Size = 27 
#> Initial Group Size = 28 
#> Initial Group Size = 29 
#> Initial Group Size = 30 
#> Initial Group Size = 31 
#> Initial Group Size = 32 
#> Initial Group Size = 33 
#> Initial Group Size = 34 
#> Initial Group Size = 35 
#> Initial Group Size = 36 
#> Initial Group Size = 37 
#> Initial Group Size = 38 
#> Initial Group Size = 39 
#> Initial Group Size = 40 
#> Initial Group Size = 41 
#> Initial Group Size = 42 
#> Initial Group Size = 43 
#> Initial Group Size = 44 
#> Initial Group Size = 45 
#> Initial Group Size = 46 
#> Initial Group Size = 47 
#> Initial Group Size = 48 
#> Initial Group Size = 49 
#> Initial Group Size = 50 
#> Initial Group Size = 51 
#> Initial Group Size = 52 
#> Initial Group Size = 53 
#> Initial Group Size = 54 
#> Initial Group Size = 55 
#> Initial Group Size = 56 
#> Initial Group Size = 57 
#> Initial Group Size = 58 
#> Initial Group Size = 59 
#> Initial Group Size = 60 
#> Initial Group Size = 61 
#> Initial Group Size = 62 
#> Initial Group Size = 63 
#> Initial Group Size = 64 
#> Initial Group Size = 65 
#> Initial Group Size = 66 
#> Initial Group Size = 67 
#> Initial Group Size = 68 
#> Initial Group Size = 69 
#> Initial Group Size = 70 
#> Initial Group Size = 71 
#> Initial Group Size = 72 
#> Initial Group Size = 73 
#> Initial Group Size = 74 
#> Initial Group Size = 75 
#> Initial Group Size = 76 
#> Initial Group Size = 77 
#> Initial Group Size = 78 
#> Initial Group Size = 79 
#> Initial Group Size = 80 
#> Initial Group Size = 81 
#> Initial Group Size = 82 
#> Initial Group Size = 83 
#> Initial Group Size = 84 
#> Initial Group Size = 85 
#> Initial Group Size = 86 
#> Initial Group Size = 87 
#> Initial Group Size = 88 
#> Initial Group Size = 89 
#> Initial Group Size = 90 
#> Initial Group Size = 91 
#> Initial Group Size = 92 
#> Initial Group Size = 93 
#> Initial Group Size = 94 
#> Initial Group Size = 95 
#> Initial Group Size = 96 
#> Initial Group Size = 97 
#> Initial Group Size = 98 
#> Initial Group Size = 99 
#> Initial Group Size = 100 
#> 
#>  Number of minutes running: 0.00175 
#>  
#> $prob
#> [1] 0.01
#> 
#> $Se
#> [1] 0.95
#> 
#> $Sp
#> [1] 0.95
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 11
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.ET$ET
#> [1] 2.586151
#> 
#> $opt.ET$value
#> [1] 0.2351047
#> 
#> $opt.ET$PSe
#> [1] 0.9025
#> 
#> $opt.ET$PSp
#> [1] 0.9931972
#> 
#> $opt.ET$PPPV
#> [1] 0.5726603
#> 
#> $opt.ET$PNPV
#> [1] 0.9990094
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 11
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01 0.01
#> 
#> $opt.MAR$ET
#> [1] 2.586151
#> 
#> $opt.MAR$value
#> [1] 0.2369314
#> 
#> $opt.MAR$PSe
#> [1] 0.9025
#> 
#> $opt.MAR$PSp
#> [1] 0.9931972
#> 
#> $opt.MAR$PPPV
#> [1] 0.5726603
#> 
#> $opt.MAR$PNPV
#> [1] 0.9990094
#> 
#> 

# Calculate the operating characteristics for a specified 
#   initial group size for non-informative two-stage 
#   hierarchical (Dorfman) testing.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
NI.Dorf(p=rep(0.025, 50), Se=0.90, Sp=0.90, group.sz=50,
obj.fn=c("ET", "MAR", "GR"), weights=matrix(data=c(1,1,10,10), 
nrow=2, ncol=2, byrow=TRUE))
#> Initial Group Size = 50 
#> 
#>  Number of minutes running: 1.666667e-05 
#>  
#> $prob
#>  [1] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [13] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [25] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [37] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [49] 0.025 0.025
#> 
#> $Se
#> [1] 0.9
#> 
#> $Sp
#> [1] 0.9
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 50
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [13] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [25] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [37] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [49] 0.025 0.025
#> 
#> $opt.ET$ET
#> [1] 34.72048
#> 
#> $opt.ET$value
#> [1] 0.6944095
#> 
#> $opt.ET$PSe
#> [1] 0.81
#> 
#> $opt.ET$PSp
#> [1] 0.9331375
#> 
#> $opt.ET$PPPV
#> [1] 0.2370058
#> 
#> $opt.ET$PNPV
#> [1] 0.9948062
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 50
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [13] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [25] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [37] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [49] 0.025 0.025
#> 
#> $opt.MAR$ET
#> [1] 34.72048
#> 
#> $opt.MAR$value
#> [1] 0.7466295
#> 
#> $opt.MAR$PSe
#> [1] 0.81
#> 
#> $opt.MAR$PSp
#> [1] 0.9331375
#> 
#> $opt.MAR$PPPV
#> [1] 0.2370058
#> 
#> $opt.MAR$PNPV
#> [1] 0.9948062
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Stage1
#> [1] 50
#> 
#> 
#> $opt.GR1$p.vec
#>  [1] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [13] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [25] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [37] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [49] 0.025 0.025
#> 
#> $opt.GR1$ET
#> [1] 34.72048
#> 
#> $opt.GR1$value
#> [1] 0.7643505
#> 
#> $opt.GR1$PSe
#> [1] 0.81
#> 
#> $opt.GR1$PSp
#> [1] 0.9331375
#> 
#> $opt.GR1$PPPV
#> [1] 0.2370058
#> 
#> $opt.GR1$PNPV
#> [1] 0.9948062
#> 
#> 
#> $opt.GR2
#> $opt.GR2$OTC
#> $opt.GR2$OTC$Stage1
#> [1] 50
#> 
#> 
#> $opt.GR2$p.vec
#>  [1] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [13] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [25] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [37] 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025 0.025
#> [49] 0.025 0.025
#> 
#> $opt.GR2$ET
#> [1] 34.72048
#> 
#> $opt.GR2$value
#> [1] 1.393819
#> 
#> $opt.GR2$PSe
#> [1] 0.81
#> 
#> $opt.GR2$PSp
#> [1] 0.9331375
#> 
#> $opt.GR2$PPPV
#> [1] 0.2370058
#> 
#> $opt.GR2$PNPV
#> [1] 0.9948062
#> 
#>