Find the optimal testing configuration

Find the optimal testing configuration (OTC) for standard group testing algorithms and calculate the associated operating characteristics.

OTC(algorithm, p = NULL, probabilities = NULL, Se = 0.99, Sp = 0.99,
  group.sz, obj.fn = c("ET", "MAR"), weights = NULL, alpha = 2)

Arguments

algorithm

character string defining the group testing algorithm to be used. Non-informative testing options include two-stage hierarchical ("D2"), three-stage hierarchical ("D3"), square array testing without master pooling ("A2"), and square array testing without master pooling ("A2M"). Informative testing options include two-stage hierarchical ("ID2"), three-stage hierarchical ("ID3"), and square array testing without master pooling ("IA2").

p

overall probability of disease that will be used to generate a vector/matrix of individual probabilities. For non-informative algorithms, a homogeneous set of probabilities will be used. For informative algorithms, the p.vec.func function will be used to generate a heterogeneous set of probabilities. Either p or probabilities should be specified, but not both.

probabilities

a vector of individual probabilities, which is homogeneous for non-informative testing algorithms and heterogeneous for informative testing algorithms. Either p or probabilities should be specified, but not both.

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

group.sz

a single group size or range of group sizes for which to calculate operating characteristics and/or find the OTC. The details of group size specification are given under 'Details'.

obj.fn

a list of objective functions which are minimized to find the OTC. The expected number of tests per individual, "ET", will always be calculated. Additional options include "MAR" (the expected number of tests divided by the expected number of correct classifications, described in Malinovsky et al. (2016)), and "GR" (a linear combination of the expected number of tests, the number of misclassified negatives, and the number of misclassified positives, described in Graff & Roeloffs (1972)). See Hitt et al. (2018) at http://chrisbilder.com/grouptesting for additional details.

weights

a matrix of up to six sets of weights for the GR function. Each set of weights is specified by a row of the matrix.

alpha

a shape parameter for the beta distribution that specifies the degree of heterogeneity for the generated probability vector (for informative testing only).

Value

A list containing:

prob

the probability of disease, as specified by the user.

alpha

level of heterogeneity for the generated probability vector (for informative testing only).

Se

the sensitivity of the diagnostic test.

Sp

the specificity of the diagnostic test.

opt.ET, opt.MAR, opt.GR

a list for each objective function specified by the user, containing:

OTC

a list specifying elements of the optimal testing configuration, which may include:

Stage1

pool size for the first stage of hierarchical testing, if applicable.

Stage2

pool sizes for the second stage of hierarchical testing, if applicable.

Block.sz

the block size/initial group size for informative Dorfman testing, which is not tested.

pool.szs

pool sizes for the first stage of testing for informative Dorfman testing.

Array.dim

the row/column size for array testing.

Array.sz

the overall array size for array testing (the square of the row/column size).

p.vec

the sorted vector of individual probabilities, if applicable.

p.mat

the sorted matrix of individual probabilities in gradient arrangement, if applicable.

ET

the expected testing expenditure for the OTC.

value

the value of the objective function per individual.

PSe

the overall pooling sensitivity for the algorithm. Further details are given under 'Details'.

PSp

the overall pooling specificity for the algorithm. Further details are given under 'Details'.

PPPV

the overall pooling positive predictive value for the algorithm. Further details are given under 'Details'.

PNPV

the overall pooling negative predictive value for the algorithm. Further details are given under 'Details'.

Details

This function finds the OTC and computes the associated operating characteristics for standard group testing algorithms, as described in Hitt et al. (2018) at http://chrisbilder.com/grouptesting.

Available algorithms include two- and three-stage hierarchical testing and array testing with and without master pooling. Both non-informative and informative group testing settings are allowed for each algorithm, except informative array testing with master pooling is unavailable because this method has not appeared in the group testing literature. Operating characteristics calculated are expected number of tests, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual.

The value(s) specified by group.sz represent the initial (stage 1) group size for three-stage hierarchical testing and non-informative two-stage hierarchical testing. For informative two-stage hierarchical testing, the group.sz specified represents the block size used in the pool-specific optimal Dorfman (PSOD) method, where the initial group (block) is not tested. For more details on informative two-stage hierarchical testing implemented via the PSOD method, see Hitt et al. (2018) and McMahan et al. (2012a). For array testing without master pooling, the group.sz specified represents the row/column size for initial (stage 1) testing. For array testing with master pooling, the group.sz specified represents the row/column size for stage 2 testing. The group size for initial testing is overall array size, given by the square of the row/column size.

If a single value is provided for group.sz with array testing or non-informative two-stage hierarchical testing, operating characteristics will be calculated and no optimization will be performed. If a single value is provided for group.sz with three-stage hierarchical or informative two-stage hierarchical, the OTC will be found over all possible configurations. If a range of group sizes is specified, the OTC will be found over all group sizes.

The displayed pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value are weighted averages of the corresponding individual accuracy measures for all individuals within the initial group for a hierarchical algorithm, or within the entire array for an array-based algorithm. Expressions for these averages are provided in the Supplementary Material for Hitt et al. (2018). These expressions are based on accuracy definitions given by Altman and Bland (1994a, 1994b).

References

Altman, D., Bland, J. (1994). “Diagnostic tests 1: sensitivity and specificity.” BMJ, 308, 1552.

Altman, D., Bland, J. (1994). “Diagnostic tests 2: predictive values.” BMJ, 309, 102.

Graff, L., Roeloffs, R. (1972). “Group testing in the presence of test error; an extension of the Dorfman procedure.” Technometrics, 14(1), 113--122. ISSN 15372723, doi:10.1080/00401706.1972.10488888 , https://www.tandfonline.com/doi/abs/10.1080/00401706.1972.10488888.

Hitt, B., Bilder, C., Tebbs, J., McMahan, C. (2018). “The Optimal Group Size Controversy for Infectious Disease Testing: Much Ado About Nothing?!” Manuscript submitted for publication.

Malinovsky, Y., Albert, P., Roy, A. (2016). “Reader reaction: A note on the evaluation of group testing algorithms in the presence of misclassification.” Biometrics, 72(1), 299--302. ISSN 15410420, doi:10.1111/biom.12385 .

McMahan, C., Tebbs, J., Bilder, C. (2012). “Informative Dorfman Screening.” Biometrics, 68(1), 287--296. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01644.x , http://www.ncbi.nlm.nih.gov/pubmed/21762119.

McMahan, C., Tebbs, J., Bilder, C. (2012). “Two-Dimensional Informative Array Testing.” Biometrics, 68(3), 793--804. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01726.x , http://www.ncbi.nlm.nih.gov/pubmed/22212007.

See also

Author

Brianna D. Hitt

Examples

# Find the OTC for non-informative
#   two-stage hierarchical (Dorfman) testing
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
OTC(algorithm="D2", p=0.05, Se=0.99, Sp=0.99, group.sz=2:100,
obj.fn=c("ET", "MAR"))
#> You have specified an overall probability of disease. 
#>  A probability vector will be generated based on the algorithm specified.
#> Non-informative Testing - alpha will be ignored. 
#> Algorithm: Non-informative two-stage hierarchical (Dorfman) testing 
#> Initial Group Size = 2 
#> Initial Group Size = 3 
#> Initial Group Size = 4 
#> Initial Group Size = 5 
#> Initial Group Size = 6 
#> Initial Group Size = 7 
#> Initial Group Size = 8 
#> Initial Group Size = 9 
#> Initial Group Size = 10 
#> Initial Group Size = 11 
#> Initial Group Size = 12 
#> Initial Group Size = 13 
#> Initial Group Size = 14 
#> Initial Group Size = 15 
#> Initial Group Size = 16 
#> Initial Group Size = 17 
#> Initial Group Size = 18 
#> Initial Group Size = 19 
#> Initial Group Size = 20 
#> Initial Group Size = 21 
#> Initial Group Size = 22 
#> Initial Group Size = 23 
#> Initial Group Size = 24 
#> Initial Group Size = 25 
#> Initial Group Size = 26 
#> Initial Group Size = 27 
#> Initial Group Size = 28 
#> Initial Group Size = 29 
#> Initial Group Size = 30 
#> Initial Group Size = 31 
#> Initial Group Size = 32 
#> Initial Group Size = 33 
#> Initial Group Size = 34 
#> Initial Group Size = 35 
#> Initial Group Size = 36 
#> Initial Group Size = 37 
#> Initial Group Size = 38 
#> Initial Group Size = 39 
#> Initial Group Size = 40 
#> Initial Group Size = 41 
#> Initial Group Size = 42 
#> Initial Group Size = 43 
#> Initial Group Size = 44 
#> Initial Group Size = 45 
#> Initial Group Size = 46 
#> Initial Group Size = 47 
#> Initial Group Size = 48 
#> Initial Group Size = 49 
#> Initial Group Size = 50 
#> Initial Group Size = 51 
#> Initial Group Size = 52 
#> Initial Group Size = 53 
#> Initial Group Size = 54 
#> Initial Group Size = 55 
#> Initial Group Size = 56 
#> Initial Group Size = 57 
#> Initial Group Size = 58 
#> Initial Group Size = 59 
#> Initial Group Size = 60 
#> Initial Group Size = 61 
#> Initial Group Size = 62 
#> Initial Group Size = 63 
#> Initial Group Size = 64 
#> Initial Group Size = 65 
#> Initial Group Size = 66 
#> Initial Group Size = 67 
#> Initial Group Size = 68 
#> Initial Group Size = 69 
#> Initial Group Size = 70 
#> Initial Group Size = 71 
#> Initial Group Size = 72 
#> Initial Group Size = 73 
#> Initial Group Size = 74 
#> Initial Group Size = 75 
#> Initial Group Size = 76 
#> Initial Group Size = 77 
#> Initial Group Size = 78 
#> Initial Group Size = 79 
#> Initial Group Size = 80 
#> Initial Group Size = 81 
#> Initial Group Size = 82 
#> Initial Group Size = 83 
#> Initial Group Size = 84 
#> Initial Group Size = 85 
#> Initial Group Size = 86 
#> Initial Group Size = 87 
#> Initial Group Size = 88 
#> Initial Group Size = 89 
#> Initial Group Size = 90 
#> Initial Group Size = 91 
#> Initial Group Size = 92 
#> Initial Group Size = 93 
#> Initial Group Size = 94 
#> Initial Group Size = 95 
#> Initial Group Size = 96 
#> Initial Group Size = 97 
#> Initial Group Size = 98 
#> Initial Group Size = 99 
#> Initial Group Size = 100 
#> 
#>  Number of minutes running: 0.0019 
#>  
#> $prob
#> [1] 0.05
#> 
#> $Se
#> [1] 0.99
#> 
#> $Sp
#> [1] 0.99
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 5
#> 
#> 
#> $opt.ET$p.vec
#> [1] 0.05 0.05 0.05 0.05 0.05
#> 
#> $opt.ET$ET
#> [1] 2.158473
#> 
#> $opt.ET$value
#> [1] 0.4316947
#> 
#> $opt.ET$PSe
#> [1] 0.9801
#> 
#> $opt.ET$PSp
#> [1] 0.9980822
#> 
#> $opt.ET$PPPV
#> [1] 0.9641539
#> 
#> $opt.ET$PNPV
#> [1] 0.9989517
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 5
#> 
#> 
#> $opt.MAR$p.vec
#> [1] 0.05 0.05 0.05 0.05 0.05
#> 
#> $opt.MAR$ET
#> [1] 2.158473
#> 
#> $opt.MAR$value
#> [1] 0.4329142
#> 
#> $opt.MAR$PSe
#> [1] 0.9801
#> 
#> $opt.MAR$PSp
#> [1] 0.9980822
#> 
#> $opt.MAR$PPPV
#> [1] 0.9641539
#> 
#> $opt.MAR$PNPV
#> [1] 0.9989517
#> 
#> 

# Find the OTC for informative
#   two-stage hierarchical (Dorfman) testing, implemented
#   via the pool-specific optimal Dorfman (PSOD) method
#   described in McMahan et al. (2012a), where the greedy
#   algorithm proposed for PSOD is replaced by considering
#   all possible testing configurations.
# A vector of individual probabilities is generated using
#   the expected value of order statistics from a beta 
#   distribution with p = 0.01 and a heterogeneity level 
#   of alpha = 0.5. Depending on the specified probability, 
#   alpha level, and overall group size, simulation may 
#   be necessary in order to generate the vector of individual
#   probabilities. This is done using p.vec.func() and 
#   requires the user to set a seed in order to reproduce 
#   results.
# This example takes approximately 2.5 minutes to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
set.seed(52613)
OTC(algorithm="ID2", p=0.01, Se=0.95, Sp=0.95, group.sz=50,
obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1, 1, 10, 10, 0.5, 0.5), 
nrow=3, ncol=2, byrow=TRUE), alpha=0.5)}

# Find the OTC over all possible
#   testing configurations for a specified group size for
#   non-informative three-stage hierarchical testing
# This example takes approximately 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
OTC(algorithm="D3", p=0.001, Se=0.95, Sp=0.95, group.sz=18,
obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1, 1), nrow=1, ncol=2, byrow=TRUE))
#> You have specified an overall probability of disease. 
#>  A probability vector will be generated based on the algorithm specified.
#> A single group size was provided. The optimal testing configuration will be found 
#>  over all possible testing configurations for the specified group size.
#> Non-informative Testing - alpha will be ignored. 
#> Algorithm: Non-informative three-stage hierarchical testing 
#> Initial Group Size = 18 
#> 
#>  Number of minutes running: 0.00425 
#>  
#> $prob
#> [1] 0.001
#> 
#> $Se
#> [1] 0.95
#> 
#> $Sp
#> [1] 0.95
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 18
#> 
#> $opt.ET$OTC$Stage2
#> [1] 9 9
#> 
#> 
#> $opt.ET$p.vec
#>  [1] 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001
#> [13] 0.001 0.001 0.001 0.001 0.001 0.001
#> 
#> $opt.ET$ET
#> [1] 1.32954
#> 
#> $opt.ET$value
#> [1] 0.07386334
#> 
#> $opt.ET$PSe
#> [1] 0.857375
#> 
#> $opt.ET$PSp
#> [1] 0.9994962
#> 
#> $opt.ET$PPPV
#> [1] 0.6301345
#> 
#> $opt.ET$PNPV
#> [1] 0.9998572
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 18
#> 
#> $opt.MAR$OTC$Stage2
#> [1] 9 9
#> 
#> 
#> $opt.MAR$p.vec
#>  [1] 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001
#> [13] 0.001 0.001 0.001 0.001 0.001 0.001
#> 
#> $opt.MAR$ET
#> [1] 1.32954
#> 
#> $opt.MAR$value
#> [1] 0.07391107
#> 
#> $opt.MAR$PSe
#> [1] 0.857375
#> 
#> $opt.MAR$PSp
#> [1] 0.9994962
#> 
#> $opt.MAR$PPPV
#> [1] 0.6301345
#> 
#> $opt.MAR$PNPV
#> [1] 0.9998572
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Stage1
#> [1] 18
#> 
#> $opt.GR1$OTC$Stage2
#> [1] 9 9
#> 
#> 
#> $opt.GR1$p.vec
#>  [1] 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001 0.001
#> [13] 0.001 0.001 0.001 0.001 0.001 0.001
#> 
#> $opt.GR1$ET
#> [1] 1.32954
#> 
#> $opt.GR1$value
#> [1] 0.07450921
#> 
#> $opt.GR1$PSe
#> [1] 0.857375
#> 
#> $opt.GR1$PSp
#> [1] 0.9994962
#> 
#> $opt.GR1$PPPV
#> [1] 0.6301345
#> 
#> $opt.GR1$PNPV
#> [1] 0.9998572
#> 
#> 

# Find the OTC for non-informative
#   three-stage hierarchical testing
# This example takes approximately 20 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
OTC(algorithm="D3", p=0.06, Se=0.90, Sp=0.90, 
group.sz=3:30, obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1, 1, 10, 10, 100, 100), 
nrow=3, ncol=2, byrow=TRUE))}

# Find the OTC over all possible configurations 
#   for a specified group size, given a 
#   heterogeneous vector of probabilities.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
OTC(algorithm="ID3", probabilities=c(0.012, 0.014, 0.011, 
0.012, 0.010, 0.015), Se=0.99, Sp=0.99, group.sz=6, 
obj.fn=c("ET","MAR","GR"), weights=matrix(data=c(1, 1), 
nrow=1, ncol=2, byrow=TRUE), alpha=0.5)
#> You have specified a vector containing individual probabilities of disease.
#> A single group size was provided. The optimal testing configuration will be found 
#>  over all possible testing configurations for the specified group size.
#> Algorithm: Informative three-stage hierarchical testing 
#> Initial Group Size = 6 
#> 
#>  Number of minutes running: 8.333333e-05 
#>  
#> $prob
#> [1] 0.012 0.014 0.011 0.012 0.010 0.015
#> 
#> $alpha
#> [1] NA
#> 
#> $Se
#> [1] 0.99
#> 
#> $Sp
#> [1] 0.99
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Stage1
#> [1] 6
#> 
#> $opt.ET$OTC$Stage2
#> [1] 3 3
#> 
#> 
#> $opt.ET$p.vec
#> [1] 0.010 0.011 0.012 0.012 0.014 0.015
#> 
#> $opt.ET$ET
#> [1] 1.378195
#> 
#> $opt.ET$value
#> [1] 0.2296991
#> 
#> $opt.ET$PSe
#> [1] 0.970299
#> 
#> $opt.ET$PSp
#> [1] 0.9997553
#> 
#> $opt.ET$PPPV
#> [1] 0.980204
#> 
#> $opt.ET$PNPV
#> [1] 0.9996292
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Stage1
#> [1] 6
#> 
#> $opt.MAR$OTC$Stage2
#> [1] 3 3
#> 
#> 
#> $opt.MAR$p.vec
#> [1] 0.010 0.011 0.012 0.012 0.014 0.015
#> 
#> $opt.MAR$ET
#> [1] 1.378195
#> 
#> $opt.MAR$value
#> [1] 0.2298388
#> 
#> $opt.MAR$PSe
#> [1] 0.970299
#> 
#> $opt.MAR$PSp
#> [1] 0.9997553
#> 
#> $opt.MAR$PPPV
#> [1] 0.980204
#> 
#> $opt.MAR$PNPV
#> [1] 0.9996292
#> 
#> 
#> $opt.GR1
#> $opt.GR1$OTC
#> $opt.GR1$OTC$Stage1
#> [1] 6
#> 
#> $opt.GR1$OTC$Stage2
#> [1] 3 3
#> 
#> 
#> $opt.GR1$p.vec
#> [1] 0.010 0.011 0.012 0.012 0.014 0.015
#> 
#> $opt.GR1$ET
#> [1] 1.378195
#> 
#> $opt.GR1$value
#> [1] 0.2303071
#> 
#> $opt.GR1$PSe
#> [1] 0.970299
#> 
#> $opt.GR1$PSp
#> [1] 0.9997553
#> 
#> $opt.GR1$PPPV
#> [1] 0.980204
#> 
#> $opt.GR1$PNPV
#> [1] 0.9996292
#> 
#> 

# Calculate the operating characteristics for a specified array size
#   for non-informative array testing without master pooling
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
OTC(algorithm="A2", p=0.005, Se=0.95, Sp=0.95, group.sz=8, 
obj.fn=c("ET", "MAR"))
#> You have specified an overall probability of disease. 
#>  A probability vector will be generated based on the algorithm specified.
#> A single group size was provided. No optimization will be performed.
#> Non-informative Testing - alpha will be ignored. 
#> Algorithm: Non-informative square array testing without master pooling 
#> Row/Column Size=8, Array Size=64
#> 
#>  Number of minutes running: 0.002016667 
#>  
#> $prob
#> [1] 0.005
#> 
#> $Se
#> [1] 0.95
#> 
#> $Sp
#> [1] 0.95
#> 
#> $opt.ET
#> $opt.ET$OTC
#> $opt.ET$OTC$Array.dim
#> [1] 8
#> 
#> $opt.ET$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.ET$p.mat
#>       [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]
#> [1,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [2,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [3,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [4,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [5,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [6,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [7,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [8,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> 
#> $opt.ET$ET
#> [1] 19.96432
#> 
#> $opt.ET$value
#> [1] 0.3119425
#> 
#> $opt.ET$PSe
#> [1] 0.905697
#> 
#> $opt.ET$PSp
#> [1] 0.9971268
#> 
#> $opt.ET$PPPV
#> [1] 0.613012
#> 
#> $opt.ET$PNPV
#> [1] 0.999525
#> 
#> 
#> $opt.MAR
#> $opt.MAR$OTC
#> $opt.MAR$OTC$Array.dim
#> [1] 8
#> 
#> $opt.MAR$OTC$Array.sz
#> [1] 64
#> 
#> 
#> $opt.MAR$p.mat
#>       [,1]  [,2]  [,3]  [,4]  [,5]  [,6]  [,7]  [,8]
#> [1,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [2,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [3,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [4,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [5,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [6,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [7,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> [8,] 0.005 0.005 0.005 0.005 0.005 0.005 0.005 0.005
#> 
#> $opt.MAR$ET
#> [1] 19.96432
#> 
#> $opt.MAR$value
#> [1] 0.3129849
#> 
#> $opt.MAR$PSe
#> [1] 0.905697
#> 
#> $opt.MAR$PSp
#> [1] 0.9971268
#> 
#> $opt.MAR$PPPV
#> [1] 0.613012
#> 
#> $opt.MAR$PNPV
#> [1] 0.999525
#> 
#> 

# Find the OTC for informative array testing without 
#   master pooling
# A vector of individual probabilities is generated using
#   the expected value of order statistics from a beta 
#   distribution with p = 0.03 and a heterogeneity level 
#   of alpha = 2. The probabilities are then arranged in 
#   a matrix using the gradient method described in 
#   McMahan et al. (2012b). Depending on the specified 
#   probability, alpha level, and overall group size, 
#   simulation may be necessary in order to generate the 
#   vector of individual probabilities. This is done using 
#   p.vec.func() and requires the user to set a 
#   seed in order to reproduce results.
# This example takes approximately 30 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
set.seed(1002)
OTC(algorithm="IA2", p=0.03, Se=0.95, Sp=0.95, 
group.sz=3:20, obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1, 1, 10, 10, 100, 100), 
nrow=3, ncol=2, byrow=TRUE), alpha=2)}

# Find the OTC for non-informative array testing 
#   with master pooling
# This example takes approximately 20 seconds to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
if (FALSE) {
OTC(algorithm="A2M", p=0.02, Se=0.90, Sp=0.90, 
group.sz=3:20, obj.fn=c("ET", "MAR", "GR"),
weights=matrix(data=c(1, 1, 10, 10, 0.5, 0.5, 2, 2, 
100, 100, 10, 100), nrow=6, ncol=2, byrow=TRUE))}