Power to Reject a Hypothesis in Binomial Group Testing for One Proportion

Closed calculation of the Power to reject a hypothesis in a binomial group testing experiment using confidence intervals for decision. Closed calculation of bias of the point estimator for a given experimental design n, s and the true, unknown proportion.

bgtPower(n, s, delta, p.hyp, conf.level = 0.95,
 method = "CP", alternative = "two.sided")

Arguments

n

integer, giving the number of groups i.e. assays i.e. observations, a vector of integers is also allowed

s

integer, giving the common size of groups i.e. the number of individual units in each group, a vector of integers is also allowed

delta

absolute difference between the threshold and the true proportion which shall be detectable with specified power, a vector is also allowed

p.hyp

number between 0 and 1, specifying the threshold proportion in the hypotheses

conf.level

confidence level required for the decision on the hypotheses

method

character string, specifying the Confidence interval method (see bgtCI) to be used

alternative

character string, defining the alternative hypothesis, either 'two.sided', 'less' or 'greater' where 'less' calculates the probability that p.hyp is excluded by an upper confidence limit for a true proportion p.hyp-delta, 'greater' calculates the probability that p.hyp is excluded by a lower confidence limit for a true proportion p.hyp+delta, 'two.sided' calculates min(power(p.hyp-delta, p.hyp+delta)) for a two-sided CI, thus can result in much lower power.

Details

The power of a confidence interval here is defined as the probability that a confidence interval or limit excludes the threshold parameter (p.hyp) of the null hypothesis, as described in Schaarschmidt(2007). I.e., the null hypothesis H0: p >= p.hyp might be rejected, if an upper confidence limit for p does not contain p.hyp. Due to discreteness, the power does not increase monotone for increasing number of groups n or group size s, but exhibits local maxima and minima, depending on n,s, conf.level, p.hyp. The power can be identical for different methods, depending on the particular combination of n, s, p.hyp, conf.level. Note that coverage probability and power are not necessarily symmetric for upper and lower bounds of binomial CI, especially for Wald, Wilson Score and Agresti-Coull CI.

Additional to the power, bias of the point estimator is calculated according to Swallow (1985).

If vectors are specified for n, s, and (or) delta, a matrix will be constructed and power and bias are calculated for each line in this matrix.

Value

A matrix containing the columns

ns

a vector of total sample size n*s resulting from the latter

n

a vector of number of groups

s

a vector of group sizes

delta

a vector of delta

power

the power to reject the given null hypothesis, with the specified method and parameters of the first 4 columns

bias

the bias of the estimator for the specified n, s, and the true proportion

References

Schaarschmidt F (2007). Experimental design for one-sided confidence intervals or hypothesis tests in binomial group testing. Communications in Biometry and Crop Science 2 (1), 32-40. http://agrobiol.sggw.waw.pl/cbcs/

Swallow WH (1985). Group testing for estimating infection rates and probabilities of disease transmission. Phytopathology Vol.75, N.8, 882-889.

See also

Examples

# Calculate the power for the design
# in the example given in Tebbs and Bilder(2004):
# n=24 groups each containing 7 insects
# if the true proportion of virus vectors
# in the population would be 0.04 (4 percent),
# the power to reject H0: p>=0.1 using an
# upper Clopper-Pearson ("CP") confidence interval
# can be calculated using the following call:

bgtPower(n=24, s=7, delta=0.06, p.hyp=0.1,
 conf.level=0.95, alternative="less", method="CP")
#>    ns  n s delta     power         bias
#> n 168 24 7  0.06 0.8821248 0.0008405368

# c(), seq() or rep() might be used to explore development
# of power and bias for varying n, s, delta. How much can
# we decrease the number of groups (costly assays to be performed)
# by pooling the same number of 320 individuals to groups of
# increasing size without largely decreasing power?

bgtPower(n=c(320,160,80,64,40,32,20,10,5),
 s=c(1,2,4,5,8,10,16,32,64),
 delta=0.01, p.hyp=0.02)
#>        ns   n  s delta       power         bias
#>  [1,] 320 320  1  0.01 0.169575200 2.428613e-17
#>  [2,] 320 160  2  0.01 0.153984506 4.775749e-05
#>  [3,] 320  80  4  0.01 0.171809348 1.485202e-04
#>  [4,] 320  64  5  0.01 0.172465973 2.017053e-04
#>  [5,] 320  40  8  0.01 0.138488643 3.737105e-04
#>  [6,] 320  32 10  0.01 0.175933578 4.999685e-04
#>  [7,] 320  20 16  0.01 0.102452816 9.486069e-04
#>  [8,] 320  10 32  0.01 0.062895913 1.090916e-02
#>  [9,] 320   5 64  0.01 0.001819794 4.455608e-01

# How does power develop for increasing differences
# delta between the true proportion and the threshold proportion?

bgtPower(n=50, s=10, delta=seq(from=0, to=0.01, by=0.001),
 p.hyp=0.01, method="CP")
#>        ns  n  s delta      power         bias
#>  [1,] 500 50 10 0.000 0.02509474 9.556995e-05
#>  [2,] 500 50 10 0.001 0.02045901 1.056338e-04
#>  [3,] 500 50 10 0.002 0.02240288 1.157943e-04
#>  [4,] 500 50 10 0.003 0.03154284 1.260525e-04
#>  [5,] 500 50 10 0.004 0.04988657 1.364095e-04
#>  [6,] 500 50 10 0.005 0.08170390 1.468667e-04
#>  [7,] 500 50 10 0.006 0.13481479 1.574252e-04
#>  [8,] 500 50 10 0.007 0.22262947 1.680863e-04
#>  [9,] 500 50 10 0.008 0.31313453 1.788510e-04
#> [10,] 500 50 10 0.009 0.37280062 1.897208e-04
#> [11,] 500 50 10 0.010 0.43351536 2.006969e-04

# use a more liberal method:

bgtPower(n=50, s=10, delta=seq(from=0, to=0.01, by=0.001),
 p.hyp=0.01, method="SOC")
#>        ns  n  s delta      power         bias
#>  [1,] 500 50 10 0.000 0.05982873 9.556995e-05
#>  [2,] 500 50 10 0.001 0.05963478 1.056338e-04
#>  [3,] 500 50 10 0.002 0.07029882 1.157943e-04
#>  [4,] 500 50 10 0.003 0.09106470 1.260525e-04
#>  [5,] 500 50 10 0.004 0.12120569 1.364095e-04
#>  [6,] 500 50 10 0.005 0.15983730 1.468667e-04
#>  [7,] 500 50 10 0.006 0.20583301 1.574252e-04
#>  [8,] 500 50 10 0.007 0.25782539 1.680863e-04
#>  [9,] 500 50 10 0.008 0.31426684 1.788510e-04
#> [10,] 500 50 10 0.009 0.37352291 1.897208e-04
#> [11,] 500 50 10 0.010 0.43397458 2.006969e-04