Hypothesis Test for One Proportion in Binomial Group Testing

Calculates p values for hypotheses tests of binomial proportions estimated from binomial group testing experiments against a threshold proportion in the hypotheses. Exact test, Score test and Wald test are available methods. Assumes equal group sizes, 100 percent sensitivity and specificity of the assays to test the groups, and individuals units randomly assigned to the groups with identical true probability of success.

bgtTest(n, y, s, p.hyp, alternative = "two.sided",
 method = "Exact")

Arguments

n

integer, number of groups (i.e. assays i.e. observations)

y

integer, number of positive groups

s

integer, common size of groups i.e. the number of individual units in each group

p.hyp

number between 0 and 1, specifying the hypothetical threshold proportion to test against

alternative

character string defining the alternative hypothesis, either 'two.sided', 'less' or 'greater'

method

character string defining the test method to be used: can be one of 'Exact' for an exact test corresponding to the Clopper-Pearson confidence interval 'Score' for a Score test, corresponding to the Wilson confidence interval 'Wald' for a Wald test corresponding to the Wald confidence interval, not recommended

Value

A list containing:

p.value

the p value of the test

estimate

the estimated proportion

p.hyp

as input

alternative

as input

method

as input

References

Swallow WH (1985) Group testing for estimating infection rates and probabilities of disease transmission. Phytopathology 75 (8), 882-889.

Blyth, C and Still, H. (1983) Binomial confidence intervals. Journal of the American Statistical Association 78, 108-116.

Santner TJ and Duffy DE (1989) The statistical analysis of discrete data. Springer New York.

Remund KM, Dixon DA, Wright DL, Holden LR (2001) Statistical considerations on seed purity testing on transgenic traits. Seed Science Research (11), 101-119.

See also

Examples

# Assume the experiment: Assays are performed on
# n=10 groups, each group is a bulk sample
# of s=100 individuals, aim is to show that 
# less than 0.5 percent ('p<0.005') of the units
# of the population show a detrimental trait (positive assay).
# The assay is senstive to show a positive result if only 1
# unit in the bulk sample of 100 units is positive.
# y=1 positive assay and 9 negative assays are observed.

bgtTest(n=10,y=1,s=100,alternative="less",method="Exact",p.hyp=0.005)
#> 
#> Exact test for one proportion in group testing
#> Alternative hypothesis: true proportion is less than 0.005 
#> p-value = 0.04996 
#> point estimate = 0.001053 

# The exact test corresponds to the 
# limits of the Clopper-Pearson confidence interval
# in the example of Tebbs and Bilder(2004):

bgtTest(n=24, y=3, s=7, alternative="two.sided",
 method="Exact", p.hyp=0.0543)
#> 
#> Exact test for one proportion in group testing
#> Alternative hypothesis: true proportion is not equal to 0.0543 
#> p-value = 0.05014 
#> point estimate = 0.0189 

bgtTest(n=24, y=3, s=7, alternative="two.sided",
 method="Exact", p.hyp=0.0038)
#> 
#> Exact test for one proportion in group testing
#> Alternative hypothesis: true proportion is not equal to 0.0038 
#> p-value = 0.04874 
#> point estimate = 0.0189 

# Further methods:

bgtTest(n=24, y=3, s=7, alternative="two.sided",
 method="Score", p.hyp=0.0516)
#> 
#> Score test for one proportion in group testing
#> Alternative hypothesis: true proportion is not equal to 0.0516 
#> p-value = 0.05019 
#> point estimate = 0.0189 

bgtTest(n=24, y=3, s=7, alternative="two.sided",
 method="Wald", p.hyp=0.0401)
#> 
#> Wald test for one proportion in group testing
#> Alternative hypothesis: true proportion is not equal to 0.0401 
#> p-value = 0.04988 
#> point estimate = 0.0189