Expected Width of Confidence Intervals in Binomial Group Testing

Calculation of expected value of the width of confidence intervals for one proportion in binomial group testing, in dependence of the number of groups, group size, confidence level and an assumed true proportion. Available for the confidence interval methods in bgtCI(binGroup).

bgtWidth(n, s, p, conf.level = 0.95, alternative = "two.sided",
 method = "CP")

Arguments

n

integer, giving the number of groups i.e. assays i.e. observations, vector of integers is also allowed

s

integer, giving the common size of groups i.e. the number of individual units in each group, vector of integers is also allowed

p

assumed true proportion of individuals showing the trait to be estimated, vector is also allowed

conf.level

required confidence level of the interval

alternative

character string, defining the alternative hypothesis, either 'two.sided', 'less' or 'greater' where 'less' calculates the expected width between the assumed true proportion p and the upper conf.level*100 percent-bound of a one-sided CI, 'greater' calculates the expected width between the true assumed proportion p and the lower conf.level*100 percent-bound of a one-sided CI, 'two.sided' calculates the expected width between the lower and the upper bound of a two-sided conf.level*100 percent-CI.

method

character string as in the method argumnet in bgtCI

Details

For calculation of expected interval width in the standard binomial estimation see, e.g., Brown et al. (2001). The calculation in case of binomial group testing is simply done by replacing the binomial probabilities by the probabilities P(Y=y) for group testing (see Tebbs and Bilder, 2004)

Value

A matrix containing the columns

ns

the resulting total number of units n*s

n

number of groups

s

group size

p

the assumed true proportion

and the calculated

expCIWidth

expected value of CI width as defined under argument alternative

References

Brown LD, Cai TT, DasGupta A (2001) Interval estimation for a binomial proportion. Statistical Science 16 (2), 101-133.

Schaarschmidt F (2007) Experimental design for one-sided confidence intervals or hypothesis tests in binomial group testing. Communications in Biometry and Crop Science 2 (1), 32-40. http://agrobiol.sggw.waw.pl/cbcs/

Tebbs JM & Bilder CR (2004) Confidence interval procedures for the probability of disease transmission in multiple-vector-transfer designs. Journal of Agricultural, Biological and Environmental Statistics 9 (1), 75-90.

Author

Frank Schaarschmidt

Examples

# There is a minimal expected CI length, if 
# group size s is increased (fixed other parameters)
# the corresponding group size might be chosen:

bgtWidth(n=20, s=seq(from=1, to=200, by=10),
 p=0.01, alternative="less", method="CP" )
#>         ns  n   s    p  expCIwidth
#>  [1,]   20 20   1 0.01 0.144314249
#>  [2,]  220 20  11 0.01 0.020459877
#>  [3,]  420 20  21 0.01 0.013520195
#>  [4,]  620 20  31 0.01 0.010890547
#>  [5,]  820 20  41 0.01 0.009503989
#>  [6,] 1020 20  51 0.01 0.008665159
#>  [7,] 1220 20  61 0.01 0.008124800
#>  [8,] 1420 20  71 0.01 0.007771843
#>  [9,] 1620 20  81 0.01 0.007553342
#> [10,] 1820 20  91 0.01 0.007452126
#> [11,] 2020 20 101 0.01 0.007487536
#> [12,] 2220 20 111 0.01 0.007728727
#> [13,] 2420 20 121 0.01 0.008313282
#> [14,] 2620 20 131 0.01 0.009463048
#> [15,] 2820 20 141 0.01 0.011489335
#> [16,] 3020 20 151 0.01 0.014782427
#> [17,] 3220 20 161 0.01 0.019784901
#> [18,] 3420 20 171 0.01 0.026952691
#> [19,] 3620 20 181 0.01 0.036710802
#> [20,] 3820 20 191 0.01 0.049411254

# and this depends largely on the assumed proportion p: 

bgtWidth(n=20, s=seq(from=1, to=200, by=10),
 p=0.05, alternative="less", method="CP" )
#>         ns  n   s    p expCIwidth
#>  [1,]   20 20   1 0.05 0.16175304
#>  [2,]  220 20  11 0.05 0.03995691
#>  [3,]  420 20  21 0.05 0.03547371
#>  [4,]  620 20  31 0.05 0.04733383
#>  [5,]  820 20  41 0.05 0.10685661
#>  [6,] 1020 20  51 0.05 0.23686121
#>  [7,] 1220 20  61 0.05 0.40647232
#>  [8,] 1420 20  71 0.05 0.56855328
#>  [9,] 1620 20  81 0.05 0.69719521
#> [10,] 1820 20  91 0.05 0.78858536
#> [11,] 2020 20 101 0.05 0.84933760
#> [12,] 2220 20 111 0.05 0.88813935
#> [13,] 2420 20 121 0.05 0.91232908
#> [14,] 2620 20 131 0.05 0.92718916
#> [15,] 2820 20 141 0.05 0.93623612
#> [16,] 3020 20 151 0.05 0.94171351
#> [17,] 3220 20 161 0.05 0.94501829
#> [18,] 3420 20 171 0.05 0.94700788
#> [19,] 3620 20 181 0.05 0.94820400
#> [20,] 3820 20 191 0.05 0.94892244

bgtWidth(n=20, s=seq(from=1, to=200, by=10),
 p=0.005, alternative="less", method="CP" )
#>         ns  n   s     p  expCIwidth
#>  [1,]   20 20   1 0.005 0.141759023
#>  [2,]  220 20  11 0.005 0.017293919
#>  [3,]  420 20  21 0.005 0.010683921
#>  [4,]  620 20  31 0.005 0.008225941
#>  [5,]  820 20  41 0.005 0.006917132
#>  [6,] 1020 20  51 0.005 0.006097076
#>  [7,] 1220 20  61 0.005 0.005533639
#>  [8,] 1420 20  71 0.005 0.005123244
#>  [9,] 1620 20  81 0.005 0.004812326
#> [10,] 1820 20  91 0.005 0.004570254
#> [11,] 2020 20 101 0.005 0.004378185
#> [12,] 2220 20 111 0.005 0.004223889
#> [13,] 2420 20 121 0.005 0.004099106
#> [14,] 2620 20 131 0.005 0.003998143
#> [15,] 2820 20 141 0.005 0.003917102
#> [16,] 3020 20 151 0.005 0.003853513
#> [17,] 3220 20 161 0.005 0.003806240
#> [18,] 3420 20 171 0.005 0.003775597
#> [19,] 3620 20 181 0.005 0.003763654
#> [20,] 3820 20 191 0.005 0.003774700