Confidence Interval for One Proportion in Group Testing with Variable Group Sizes

Calculates confidence intervals for a single proportion in binomial group testing if groups of different size are evaluated

bgtvs(n, s, y, conf.level = 0.95, alternative = "two.sided",
 maxiter = 100)

Arguments

n

vector of integer values, specifying the number of groups of the different sizes

s

vector of integer values, specifying the group sizes, must be of same length as n

y

vector of integer values, specifying the number of positive groups among the n groups tested

conf.level

a single numeric value, the confidence level of the interval

alternative

a character string, with options "two.sided", "less", "greater"

maxiter

maximal number steps in iteration of confidence limits

Details

Hepworth (1996) describes methods for constructing confidence intervals in binomial group testing, if groups of different size are used. Currently, only the exact method (Hepworth, 1996, equation5, Table.5) is implemented. Note, that the algorithm becomes computationally very expensive if the number of different groups becomes larger than 3.

Value

A list containing

conf.int

a numeric vector, the lower and upper limits of the confidence interval

estimate

the point estimate

conf.level

as input

alternative

as imput

input

a matrix containing the input values of n (number of groups), s (group size), and y (number of positive pools)

moreover, some of the input arguments.

References

Hepworth, G (1996): Exact confidence intervals for proportions estimated by group testing. Biometrics 52, 1134-1146.

Author

Frank Schaarschmidt

See also

Examples

# Consider a very simple example,
# given in Hepworth (1996), table 5:
# 2 groups each containing 5 units,
# and 3 groups, each containing 2 units

# In the first setting (n=2, s=5) y=1 positive group 
# has been observed, in the second setting (n=3, s=2),
# y=2 positive have been observed.

bgtvs(n=c(2,3), s=c(5,2), y=c(1,2)) 
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    1    2
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.05172 0.55330
#> 
#> Point estimate: 
#> 
#> [1] 0.2435

###############################################

# Recalculate the example given in
# Hepworth (1996), table 5:

 bgtvs(n=c(2,3), s=c(5,2), y=c(0,0))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    0    0
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.0000 0.2059
#> 
#> Point estimate: 
#> 
#> [1] 0
 bgtvs(n=c(2,3), s=c(5,2), y=c(0,1))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    0    1
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.001581 0.266743
#> 
#> Point estimate: 
#> 
#> [1] 0.06459
 bgtvs(n=c(2,3), s=c(5,2), y=c(0,2))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    0    2
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.01726 0.35318
#> 
#> Point estimate: 
#> 
#> [1] 0.134
 bgtvs(n=c(2,3), s=c(5,2), y=c(0,3))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    0    3
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.05125 0.50029
#> 
#> Point estimate: 
#> 
#> [1] 0.2094
 bgtvs(n=c(2,3), s=c(5,2), y=c(1,0))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    1    0
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.002521 0.338654
#> 
#> Point estimate: 
#> 
#> [1] 0.0722
 bgtvs(n=c(2,3), s=c(5,2), y=c(1,1))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    1    1
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.01842 0.44346
#> 
#> Point estimate: 
#> 
#> [1] 0.1523
 bgtvs(n=c(2,3), s=c(5,2), y=c(1,2))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    1    2
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.05172 0.55330
#> 
#> Point estimate: 
#> 
#> [1] 0.2435
 bgtvs(n=c(2,3), s=c(5,2), y=c(1,3))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    1    3
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.1084 0.7072
#> 
#> Point estimate: 
#> 
#> [1] 0.3516
 bgtvs(n=c(2,3), s=c(5,2), y=c(2,0))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    2    0
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.03161 0.49909
#> 
#> Point estimate: 
#> 
#> [1] 0.1781
 bgtvs(n=c(2,3), s=c(5,2), y=c(2,1))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    2    1
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.06237 0.69670
#> 
#> Point estimate: 
#> 
#> [1] 0.2979
 bgtvs(n=c(2,3), s=c(5,2), y=c(2,2))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    2    2
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.1147 0.9084
#> 
#> Point estimate: 
#> 
#> [1] 0.4699
 bgtvs(n=c(2,3), s=c(5,2), y=c(2,3))
#> Input 
#>                           [,1] [,2]
#> number of groups             2    3
#> group size                   5    2
#> number of positive groups    2    3
#> 
#> Exact 95 percent confidence interval 
#> 
#> [1] 0.209 1.000
#> 
#> Point estimate: 
#> 
#> [1] 1