Sample Size Iteration Depending on Minimal MSE in One-Parameter Group Testing

Find the group size s for a fixed number of assays n and an assumed true proportion p.tr for which the mean squared error (mse) of the point estimator is minimal and bias is within a restriction. For experimental design in binomial group testing as recommended by Swallow (1985), if main objective is estimation.

estDesign(n, smax, p.tr, biasrest = 0.05)

Arguments

n

integer, fixed sample size (number of assays)

smax

integer, maximal group size allowed in planning of the design

p.tr

assumed true proportion of the 'positive' trait in the population to be tested, specify as a value between 0 and 1

biasrest

value between 0 and 1 specifying the absolute bias maximally allowed

Details

Swallow (1985) recommends to use the upper bound of the expected range of true proportion p.tr for optimization of tzhe design. For further details see the reference. Up to now, specify n<1020.

Value

the group size s, for which the mse of the estimator is minimal for the given n, p.tr or the group size s for which bias restriction biasrest is just not violated, and for this particular group size s: a list containing:

varp

the variance of the estimator

mse

the mean square error of the estimator

bias

the bias of the estimator

exp

the expected value of the estimator

References

Swallow WH (1985) Group testing for estimating infection rates and probabilities of disease transmission. Phytopathology Vol.75, N.8, 882-889.

Author

Frank Schaarschmidt

See also

Examples

### Compare table 1 in Swallow(1985),885:

estDesign(n=10, smax=100, p.tr=0.001)
#>  minimal mse(p) is achieved with group size s >= smax 
#> $varp
#> [1] 1.198715e-06
#> 
#> $mse
#> [1] 1.201888e-06
#> 
#> $bias
#> [1] 5.63273e-05
#> 
#> $exp
#> [1] 0.001056327
#> 

estDesign(n=10, smax=100, p.tr=0.01)
#> group size s with minimal mse(p) = 33 
#> $varp
#> [1] 4.539192e-05
#> 
#> $mse
#> [1] 4.580377e-05
#> 
#> $bias
#> [1] 0.0006417578
#> 
#> $exp
#> [1] 0.01064176
#> 

estDesign(n=25, smax=100, p.tr=0.05)
#> group size s with minimal mse(p) = 18 
#> $varp
#> [1] 0.0001971096
#> 
#> $mse
#> [1] 0.0001998226
#> 
#> $bias
#> [1] 0.001647126
#> 
#> $exp
#> [1] 0.05164713
#> 

estDesign(n=40, smax=100, p.tr=0.25)
#> group size s with minimal mse(p) = 4 
#> $varp
#> [1] 0.00203969
#> 
#> $mse
#> [1] 0.002055717
#> 
#> $bias
#> [1] 0.004003408
#> 
#> $exp
#> [1] 0.2540034
#> 

estDesign(n=200, smax=100, p.tr=0.3)
#> group size s with minimal mse(p) = 4 
#> $varp
#> [1] 0.0004932118
#> 
#> $mse
#> [1] 0.0004943187
#> 
#> $bias
#> [1] 0.001052103
#> 
#> $exp
#> [1] 0.3010521
#>