Auxiliary for Controlling Group Testing Regression

Auxiliary function to control fitting parameters of EM algorithm used internally in gtreg.mp and EM.mp or gtreg, EM, and EM.ret.

gt.control(tol = 0.0001, n.gibbs = 1000, n.burnin = 20,
 maxit = 500, trace = FALSE, time = TRUE)

Arguments

tol

convergence criterion

n.gibbs

the Gibbs sample size to be used in each E step for the EM algorithm (default is 1000), for matrix pooling

n.burnin

the number of samples in the burn-in period (default is 20), for matrix pooling

maxit

maximal number of iterations in the EM algorithm

trace

logical indicating if output should be printed for each iteration, defaults to FALSE

time

logical indicating if the length of time for the model fitting should be printed, defaults to TRUE

Value

A list with components named as the arguments

Examples

# The default settings:
gt.control()
#> $tol
#> [1] 1e-04
#> 
#> $n.gibbs
#> [1] 1000
#> 
#> $n.burnin
#> [1] 20
#> 
#> $maxit
#> [1] 500
#> 
#> $trace
#> [1] FALSE
#> 
#> $time
#> [1] TRUE
#>