Operating characteristics for informative two-stage hierarchical (Dorfman) testing

Calculate the expected number of tests and accuracy measures for each individual using informative two-stage hierarchical (Dorfman) testing, given a vector of individual probabilities and a testing configuration.

inf.dorf.measures(prob, se, sp, N, pool.sizes)

Arguments

prob

vector of probabilities corresponding to each individual's risk of disease.

se

the sensitivity of the diagnostic test.

sp

the specificity of the diagnostic test.

N

block size/initial group size that is not tested. This is the total number of individuals being tested. The details of block.sz are given under 'Details'.

pool.sizes

a vector of pool sizes for the first stage of testing.

Value

A list containing:

e

the expected number of tests needed to decode all N individuals.

v

the variance of the total number of tests needed to decode all N individuals.

summary

a matrix containing the pool, probability of disease, pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual. The pool column identifies which pool each individual is contained in for the first stage of testing.

Details

This function utilizes the equations given by McMahan et al. (2012) for informative two-stage hierarchical (Dorfman) testing. It also repurposes functions written by Christopher S. McMahan (see http://chrisbilder.com/grouptesting) for the implementation of informative Dorfman testing and directly uses functions written for the calculation of the associated operating characteristics. This function calculates the operating characteristics for informative two-stage hierarchical (Dorfman) testing. Operating characteristics calculated are expected number of tests, and pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual.

The specified N represents the block size used in the pool-specific optimal Dorfman (PSOD) method. This is the total number of individuals being tested by the algorithm. This block is not initially tested. Instead, multiple initial pool sizes within this block are found and tested in the first stage of testing. The second stage of testing consists of individual retesting. For more information on block size specification, see McMahan et al. (2012).

Note

This function returns the pooling positive and negative predictive values for all individuals even though these measures are diagnostic specific; i.e., PPPV (PNPV) should only be considered for those individuals who have tested positive (negative).

References

McMahan, C., Tebbs, J., Bilder, C. (2012). “Informative Dorfman Screening.” Biometrics, 68(1), 287--296. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01644.x , http://www.ncbi.nlm.nih.gov/pubmed/21762119.

See also

Author

The majority of this function was originally written by Christopher S. McMahan for McMahan et al. (2012). The function was obtained from http://chrisbilder.com/grouptesting. Minor modifications were made to the function for inclusion in the binGroup package.

Examples

# Calculate the operating characteristics for 
#   informative two-stage hierarchical (Dorfman) testing
#   with an overall disease prevalence of E(p(i)) = 0.01,
#   where a block size of 50 is split into initial pools
#   of 18, 13, 11, and 8 individuals.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(8791)
inf.dorf.measures(prob=beta.dist(p=0.01, alpha=2, 
grp.sz=50, simul=TRUE), se=0.95, sp=0.95, N=50, 
pool.sizes=c(18, 13, 11, 8))
#> [1] "Using simulation"
#> $e
#> [1] 11.32286
#> 
#> $v
#> [1] 79.92548
#> 
#> $summary
#>    pool probability    PSe       PSp       PPV       NPV
#> 1     1 0.000964839 0.9025 0.9945235 0.1373025 0.9999053
#> 2     1 0.001491870 0.9025 0.9945457 0.1982183 0.9998535
#> 3     1 0.001906296 0.9025 0.9945632 0.2407240 0.9998128
#> 4     1 0.002272982 0.9025 0.9945786 0.2749664 0.9997767
#> 5     1 0.002609000 0.9025 0.9945928 0.3039116 0.9997436
#> 6     1 0.002925607 0.9025 0.9946062 0.3292869 0.9997124
#> 7     1 0.003228595 0.9025 0.9946190 0.3520164 0.9996826
#> 8     1 0.003519830 0.9025 0.9946313 0.3725625 0.9996539
#> 9     1 0.003803796 0.9025 0.9946433 0.3914722 0.9996258
#> 10    1 0.004083119 0.9025 0.9946551 0.4090783 0.9995983
#> 11    1 0.004360304 0.9025 0.9946668 0.4256504 0.9995709
#> 12    1 0.004628075 0.9025 0.9946782 0.4408724 0.9995444
#> 13    1 0.004899461 0.9025 0.9946897 0.4555671 0.9995176
#> 14    1 0.005165369 0.9025 0.9947010 0.4692993 0.9994913
#> 15    1 0.005434036 0.9025 0.9947124 0.4825501 0.9994647
#> 16    1 0.005708476 0.9025 0.9947240 0.4954811 0.9994376
#> 17    1 0.005982808 0.9025 0.9947357 0.5078366 0.9994104
#> 18    1 0.006258616 0.9025 0.9947474 0.5197209 0.9993831
#> 19    2 0.006535886 0.9025 0.9931101 0.4628735 0.9993545
#> 20    2 0.006818977 0.9025 0.9931217 0.4739207 0.9993264
#> 21    2 0.007105802 0.9025 0.9931334 0.4847014 0.9992979
#> 22    2 0.007401755 0.9025 0.9931455 0.4954141 0.9992685
#> 23    2 0.007698052 0.9025 0.9931576 0.5057440 0.9992390
#> 24    2 0.007997275 0.9025 0.9931699 0.5157962 0.9992092
#> 25    2 0.008306605 0.9025 0.9931826 0.5258088 0.9991784
#> 26    2 0.008631119 0.9025 0.9931959 0.5359212 0.9991461
#> 27    2 0.008953608 0.9025 0.9932092 0.5455952 0.9991139
#> 28    2 0.009290883 0.9025 0.9932230 0.5553347 0.9990803
#> 29    2 0.009635682 0.9025 0.9932372 0.5649143 0.9990458
#> 30    2 0.009994855 0.9025 0.9932520 0.5745107 0.9990100
#> 31    2 0.010366220 0.9025 0.9932673 0.5840459 0.9989728
#> 32    3 0.010750141 0.9025 0.9918248 0.5453828 0.9989329
#> 33    3 0.011147978 0.9025 0.9918406 0.5549548 0.9988930
#> 34    3 0.011575954 0.9025 0.9918576 0.5648589 0.9988501
#> 35    3 0.012009098 0.9025 0.9918749 0.5744913 0.9988066
#> 36    3 0.012478998 0.9025 0.9918936 0.5845224 0.9987594
#> 37    3 0.012972341 0.9025 0.9919133 0.5946135 0.9987098
#> 38    3 0.013503307 0.9025 0.9919345 0.6050019 0.9986564
#> 39    3 0.014073845 0.9025 0.9919573 0.6156549 0.9985989
#> 40    3 0.014690843 0.9025 0.9919821 0.6266227 0.9985367
#> 41    3 0.015354022 0.9025 0.9920086 0.6378174 0.9984697
#> 42    3 0.016084387 0.9025 0.9920380 0.6494890 0.9983959
#> 43    4 0.016900403 0.9025 0.9907414 0.6262681 0.9983111
#> 44    4 0.017805662 0.9025 0.9907766 0.6394907 0.9982192
#> 45    4 0.018852616 0.9025 0.9908175 0.6538021 0.9981128
#> 46    4 0.020097982 0.9025 0.9908662 0.6695940 0.9979859
#> 47    4 0.021610914 0.9025 0.9909255 0.6871847 0.9978314
#> 48    4 0.023636716 0.9025 0.9910052 0.7083720 0.9976239
#> 49    4 0.026567983 0.9025 0.9911212 0.7350461 0.9973223
#> 50    4 0.032217050 0.9025 0.9913466 0.7763818 0.9967366
#>