Find the characteristics of an informative two-stage hierarchical (Dorfman) algorithm

Find the characteristics of an informative two-stage hierarchical (Dorfman) decoding process using Optimal Dorfman (OD), Thresholded Optimal Dorfman (TOD), or Pool-Specific Optimal Dorfman (PSOD) algorithms.

opt.info.dorf(prob, se = 1, sp = 1, method = "OD", max.pool = 15,
  thresh.pool = 8, threshold = NULL)

Arguments

prob

a vector of all subjects' infection probabilities.

se

the sensitivity of the diagnostic test.

sp

the specificity of the diagnostic test.

method

character string defining the specific screening procedure for implementation of Dorfman retesting in a heterogeneous population. Options include Optimal Dorfman ("OD"), Thresholded Optimal Dorfman ("TOD"), and Pool-Specific Optimal Dorfman ("PSOD"). Further details are given under 'Details'.

max.pool

the maximum allowable pool size. Further details are given under 'Details'.

thresh.pool

the initial pool size used for TOD, if threshold is not specified. Further details are given under 'Details'.

threshold

the threshold value for TOD. If a threshold value is not specified, one is found algorithmically. Further details are given under 'Details'.

Value

A list containing:

tv

the threshold value used for TOD, if applicable.

e

the expected expenditure of the decoding process.

v

the variance of the expenditure of the decoding process.

summary

a matrix of summary measures that includes each individual's infection probability, pool (pool to which they belong), pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value.

Details

This function finds the characteristics of an informative two-stage hierarchical (Dorfman) decoding process. Characteristics found include the expected expenditure of the decoding process, the variance of the expenditure of the decoding process, and the pooling sensitivity, pooling specificity, pooling positive predictive value, and pooling negative predictive value for each individual. Calculations of these characteristics are done using equations presented in McMahan et al. (2012).

Optimal Dorfman (OD) is an informative Dorfman algorithm in which the common pool size \(c=c_{opt}\) minimizes \(E(T^(c))\), the expected number of tests needed to decode all \(N\) individuals when pools of size \(c\) are used.

Thresholded Optimal Dorfman (TOD) is an informative Dorfman algorithm in which all \(N\) individuals are partitioned into two classes, low-risk and high-risk individuals, based on whether their risk probability falls below or above a particular threshold value. The threshold can be specified using the threshold argument or the TOD algorithm can identify the optimal threshold value. The low-risk individuals are tested using a optimal common pool size, and high-risk individuals are tested individually.

Pool-Specific Optimal Dorfman (PSOD) is an informative Dorfman algorithm in which optimal sizes are determined for each pool. A total of \(N\) individuals are tested in pools that minimize the expected number of tests per individual, on a pool-by-pool basis. If desired, the user can add the constraint of a maximum allowable pool size, so that each pool will contain no more than the maximum allowable number of individuals.

All three informative Dorfman procedures described above require individuals to be ordered from smallest to largest probability of infection. See McMahan et al. (2012) for additional details on the implementation of informative two-stage hierarchical (Dorfman) testing algorithms.

References

Dorfman, R. (1943). “The Detection of Defective Members of Large Populations.” The Annals of Mathematical Statistics, 14(4), 436--440. ISSN 0003-4851, doi:10.1214/aoms/1177731363 , https://www.jstor.org/stable/2235930.

McMahan, C., Tebbs, J., Bilder, C. (2012). “Informative Dorfman Screening.” Biometrics, 68(1), 287--296. ISSN 0006341X, doi:10.1111/j.1541-0420.2011.01644.x , http://www.ncbi.nlm.nih.gov/pubmed/21762119.

See also

Author

This function was originally written by Christopher S. McMahan for McMahan et al. (2012). The function was obtained from http://chrisbilder.com/grouptesting.

Examples

# Find the characteristics of an informative
#   Dorfman algorithm, using the OD procedure.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
opt.info.dorf(prob=rbeta(1000,1,10), se=1, sp=1, 
method ="OD", max.pool=15, thresh.pool=8, threshold=NULL)
#> $tv
#> NULL
#> 
#> $e
#> [1] 529.2846
#> 
#> $v
#> [1] 831.3632
#> 
#> $summary
#>      pool  probability PSe PSp PPV NPV
#> 1      47 2.468769e-02   1   1   1   1
#> 2      43 2.308007e-02   1   1   1   1
#> 3     147 1.210288e-01   1   1   1   1
#> 4     137 1.065673e-01   1   1   1   1
#> 5     103 7.249004e-02   1   1   1   1
#> 6      40 2.140679e-02   1   1   1   1
#> 7      83 5.189986e-02   1   1   1   1
#> 8     173 1.707023e-01   1   1   1   1
#> 9      47 2.460822e-02   1   1   1   1
#> 10    137 1.066939e-01   1   1   1   1
#> 11     53 2.813306e-02   1   1   1   1
#> 12     35 1.803130e-02   1   1   1   1
#> 13    198 3.796655e-01   1   1   1   1
#> 14     20 8.472046e-03   1   1   1   1
#> 15     62 3.527498e-02   1   1   1   1
#> 16    180 1.953743e-01   1   1   1   1
#> 17     35 1.821333e-02   1   1   1   1
#> 18     90 5.940466e-02   1   1   1   1
#> 19     82 5.006110e-02   1   1   1   1
#> 20     83 5.042604e-02   1   1   1   1
#> 21     80 4.780597e-02   1   1   1   1
#> 22    126 9.267940e-02   1   1   1   1
#> 23    140 1.130882e-01   1   1   1   1
#> 24     35 1.871687e-02   1   1   1   1
#> 25     76 4.521193e-02   1   1   1   1
#> 26     46 2.411494e-02   1   1   1   1
#> 27     43 2.318404e-02   1   1   1   1
#> 28    120 8.817146e-02   1   1   1   1
#> 29    187 2.274207e-01   1   1   1   1
#> 30    176 1.829458e-01   1   1   1   1
#> 31    163 1.505835e-01   1   1   1   1
#> 32     80 4.743691e-02   1   1   1   1
#> 33    159 1.436886e-01   1   1   1   1
#> 34     51 2.681023e-02   1   1   1   1
#> 35     14 5.767000e-03   1   1   1   1
#> 36    188 2.373017e-01   1   1   1   1
#> 37    107 7.554120e-02   1   1   1   1
#> 38    166 1.553598e-01   1   1   1   1
#> 39    131 9.865375e-02   1   1   1   1
#> 40     52 2.780978e-02   1   1   1   1
#> 41    178 1.895933e-01   1   1   1   1
#> 42    125 9.212848e-02   1   1   1   1
#> 43     66 3.842708e-02   1   1   1   1
#> 44    109 7.792316e-02   1   1   1   1
#> 45     39 2.071691e-02   1   1   1   1
#> 46     36 1.919161e-02   1   1   1   1
#> 47     86 5.445738e-02   1   1   1   1
#> 48     83 5.191830e-02   1   1   1   1
#> 49    179 1.903068e-01   1   1   1   1
#> 50    169 1.614679e-01   1   1   1   1
#> 51    171 1.669861e-01   1   1   1   1
#> 52    110 7.833458e-02   1   1   1   1
#> 53    179 1.904131e-01   1   1   1   1
#> 54     77 4.544390e-02   1   1   1   1
#> 55    193 2.706342e-01   1   1   1   1
#> 56      6 1.824861e-03   1   1   1   1
#> 57     77 4.537933e-02   1   1   1   1
#> 58    176 1.796318e-01   1   1   1   1
#> 59     34 1.759606e-02   1   1   1   1
#> 60    167 1.577480e-01   1   1   1   1
#> 61     31 1.636308e-02   1   1   1   1
#> 62     96 6.401792e-02   1   1   1   1
#> 63    105 7.333307e-02   1   1   1   1
#> 64     56 3.094369e-02   1   1   1   1
#> 65    163 1.515135e-01   1   1   1   1
#> 66    136 1.051702e-01   1   1   1   1
#> 67     19 8.394278e-03   1   1   1   1
#> 68     88 5.751485e-02   1   1   1   1
#> 69    103 7.215063e-02   1   1   1   1
#> 70     19 8.111303e-03   1   1   1   1
#> 71     58 3.283584e-02   1   1   1   1
#> 72    106 7.375733e-02   1   1   1   1
#> 73    139 1.091872e-01   1   1   1   1
#> 74     65 3.834581e-02   1   1   1   1
#> 75    103 7.259906e-02   1   1   1   1
#> 76     81 4.844209e-02   1   1   1   1
#> 77     90 5.946833e-02   1   1   1   1
#> 78      7 1.932005e-03   1   1   1   1
#> 79    155 1.366817e-01   1   1   1   1
#> 80    172 1.681431e-01   1   1   1   1
#> 81     41 2.199962e-02   1   1   1   1
#> 82     86 5.500709e-02   1   1   1   1
#> 83    150 1.260270e-01   1   1   1   1
#> 84     52 2.734232e-02   1   1   1   1
#> 85    139 1.106931e-01   1   1   1   1
#> 86    194 2.956426e-01   1   1   1   1
#> 87    161 1.470924e-01   1   1   1   1
#> 88    187 2.284023e-01   1   1   1   1
#> 89    109 7.731824e-02   1   1   1   1
#> 90    105 7.328279e-02   1   1   1   1
#> 91     58 3.268523e-02   1   1   1   1
#> 92     35 1.840357e-02   1   1   1   1
#> 93     14 5.654581e-03   1   1   1   1
#> 94    111 7.906837e-02   1   1   1   1
#> 95      9 2.707795e-03   1   1   1   1
#> 96    160 1.452237e-01   1   1   1   1
#> 97     67 3.969714e-02   1   1   1   1
#> 98      8 2.421793e-03   1   1   1   1
#> 99     49 2.555360e-02   1   1   1   1
#> 100    37 1.972199e-02   1   1   1   1
#> 101   149 1.243414e-01   1   1   1   1
#> 102   100 6.795082e-02   1   1   1   1
#> 103    63 3.583892e-02   1   1   1   1
#> 104    69 4.104877e-02   1   1   1   1
#> 105   132 1.002221e-01   1   1   1   1
#> 106    76 4.516087e-02   1   1   1   1
#> 107   158 1.413477e-01   1   1   1   1
#> 108    67 3.941871e-02   1   1   1   1
#> 109   153 1.318864e-01   1   1   1   1
#> 110   153 1.324427e-01   1   1   1   1
#> 111     3 1.083870e-03   1   1   1   1
#> 112   192 2.631389e-01   1   1   1   1
#> 113   197 3.220077e-01   1   1   1   1
#> 114   124 9.110625e-02   1   1   1   1
#> 115    57 3.173309e-02   1   1   1   1
#> 116    64 3.732524e-02   1   1   1   1
#> 117     8 2.323653e-03   1   1   1   1
#> 118    80 4.766856e-02   1   1   1   1
#> 119   138 1.076881e-01   1   1   1   1
#> 120    79 4.742086e-02   1   1   1   1
#> 121   147 1.202411e-01   1   1   1   1
#> 122   133 1.005570e-01   1   1   1   1
#> 123   165 1.542989e-01   1   1   1   1
#> 124   105 7.375160e-02   1   1   1   1
#> 125   101 6.911024e-02   1   1   1   1
#> 126   171 1.672019e-01   1   1   1   1
#> 127    21 8.567552e-03   1   1   1   1
#> 128    88 5.610213e-02   1   1   1   1
#> 129   140 1.113274e-01   1   1   1   1
#> 130    77 4.574071e-02   1   1   1   1
#> 131    44 2.349100e-02   1   1   1   1
#> 132   181 1.988308e-01   1   1   1   1
#> 133   170 1.648505e-01   1   1   1   1
#> 134    32 1.665808e-02   1   1   1   1
#> 135   183 2.066391e-01   1   1   1   1
#> 136    60 3.405011e-02   1   1   1   1
#> 137    52 2.799056e-02   1   1   1   1
#> 138    94 6.227987e-02   1   1   1   1
#> 139   118 8.524360e-02   1   1   1   1
#> 140   187 2.271304e-01   1   1   1   1
#> 141   185 2.190534e-01   1   1   1   1
#> 142     4 1.383172e-03   1   1   1   1
#> 143    17 6.551080e-03   1   1   1   1
#> 144   104 7.317646e-02   1   1   1   1
#> 145    54 3.033816e-02   1   1   1   1
#> 146    34 1.757853e-02   1   1   1   1
#> 147   128 9.605331e-02   1   1   1   1
#> 148   148 1.232991e-01   1   1   1   1
#> 149    15 5.916322e-03   1   1   1   1
#> 150   119 8.711167e-02   1   1   1   1
#> 151    58 3.272420e-02   1   1   1   1
#> 152    26 1.198981e-02   1   1   1   1
#> 153    27 1.260492e-02   1   1   1   1
#> 154    12 4.691209e-03   1   1   1   1
#> 155     1 2.719472e-04   1   1   1   1
#> 156    48 2.512785e-02   1   1   1   1
#> 157    70 4.153769e-02   1   1   1   1
#> 158   140 1.117865e-01   1   1   1   1
#> 159   154 1.335376e-01   1   1   1   1
#> 160    34 1.768238e-02   1   1   1   1
#> 161     2 3.902761e-04   1   1   1   1
#> 162   115 8.226119e-02   1   1   1   1
#> 163   165 1.539993e-01   1   1   1   1
#> 164   126 9.306573e-02   1   1   1   1
#> 165    61 3.475463e-02   1   1   1   1
#> 166    54 3.019993e-02   1   1   1   1
#> 167   157 1.391008e-01   1   1   1   1
#> 168     2 8.557898e-04   1   1   1   1
#> 169   188 2.314761e-01   1   1   1   1
#> 170   110 7.820326e-02   1   1   1   1
#> 171   121 8.967008e-02   1   1   1   1
#> 172   111 7.943874e-02   1   1   1   1
#> 173   124 9.156110e-02   1   1   1   1
#> 174    50 2.602921e-02   1   1   1   1
#> 175   125 9.219646e-02   1   1   1   1
#> 176    57 3.188957e-02   1   1   1   1
#> 177   172 1.687012e-01   1   1   1   1
#> 178    68 4.011778e-02   1   1   1   1
#> 179   167 1.571936e-01   1   1   1   1
#> 180    12 4.845704e-03   1   1   1   1
#> 181    57 3.171106e-02   1   1   1   1
#> 182    81 4.895741e-02   1   1   1   1
#> 183    56 3.152005e-02   1   1   1   1
#> 184   130 9.734476e-02   1   1   1   1
#> 185   189 2.419042e-01   1   1   1   1
#> 186   154 1.332935e-01   1   1   1   1
#> 187    96 6.356869e-02   1   1   1   1
#> 188   154 1.331902e-01   1   1   1   1
#> 189   152 1.315972e-01   1   1   1   1
#> 190    30 1.556812e-02   1   1   1   1
#> 191   138 1.078578e-01   1   1   1   1
#> 192    72 4.305669e-02   1   1   1   1
#> 193   140 1.130844e-01   1   1   1   1
#> 194   200 4.653359e-01   1   1   1   1
#> 195    19 7.578608e-03   1   1   1   1
#> 196    45 2.377319e-02   1   1   1   1
#> 197    16 6.529255e-03   1   1   1   1
#> 198   146 1.198014e-01   1   1   1   1
#> 199   149 1.253443e-01   1   1   1   1
#> 200   185 2.160892e-01   1   1   1   1
#> 201    50 2.606631e-02   1   1   1   1
#> 202   116 8.345971e-02   1   1   1   1
#> 203   144 1.161704e-01   1   1   1   1
#> 204   164 1.521726e-01   1   1   1   1
#> 205   114 8.134522e-02   1   1   1   1
#> 206    76 4.524474e-02   1   1   1   1
#> 207   147 1.207568e-01   1   1   1   1
#> 208   192 2.551032e-01   1   1   1   1
#> 209   197 3.219664e-01   1   1   1   1
#> 210   133 1.029930e-01   1   1   1   1
#> 211   168 1.604134e-01   1   1   1   1
#> 212     5 1.466880e-03   1   1   1   1
#> 213   121 8.945490e-02   1   1   1   1
#> 214   159 1.422192e-01   1   1   1   1
#> 215   134 1.031878e-01   1   1   1   1
#> 216    57 3.180812e-02   1   1   1   1
#> 217   107 7.623700e-02   1   1   1   1
#> 218   130 9.757765e-02   1   1   1   1
#> 219    54 3.033611e-02   1   1   1   1
#> 220   102 7.036804e-02   1   1   1   1
#> 221    97 6.414566e-02   1   1   1   1
#> 222    14 5.844900e-03   1   1   1   1
#> 223   171 1.665115e-01   1   1   1   1
#> 224    69 4.105393e-02   1   1   1   1
#> 225    28 1.361360e-02   1   1   1   1
#> 226   151 1.292131e-01   1   1   1   1
#> 227   189 2.394865e-01   1   1   1   1
#> 228    64 3.749278e-02   1   1   1   1
#> 229    97 6.470560e-02   1   1   1   1
#> 230   145 1.184734e-01   1   1   1   1
#> 231   133 1.010097e-01   1   1   1   1
#> 232   172 1.695418e-01   1   1   1   1
#> 233   190 2.435737e-01   1   1   1   1
#> 234    36 1.892359e-02   1   1   1   1
#> 235   126 9.421107e-02   1   1   1   1
#> 236   129 9.669133e-02   1   1   1   1
#> 237   125 9.173977e-02   1   1   1   1
#> 238   176 1.795419e-01   1   1   1   1
#> 239    69 4.122738e-02   1   1   1   1
#> 240    55 3.044889e-02   1   1   1   1
#> 241    16 6.520064e-03   1   1   1   1
#> 242   133 1.029734e-01   1   1   1   1
#> 243   136 1.057397e-01   1   1   1   1
#> 244   117 8.408415e-02   1   1   1   1
#> 245    97 6.469850e-02   1   1   1   1
#> 246   116 8.264913e-02   1   1   1   1
#> 247   180 1.974586e-01   1   1   1   1
#> 248    63 3.650958e-02   1   1   1   1
#> 249    11 4.536665e-03   1   1   1   1
#> 250    25 1.108100e-02   1   1   1   1
#> 251    72 4.282135e-02   1   1   1   1
#> 252    44 2.319902e-02   1   1   1   1
#> 253   142 1.147644e-01   1   1   1   1
#> 254    22 9.004856e-03   1   1   1   1
#> 255   170 1.658895e-01   1   1   1   1
#> 256    93 6.113927e-02   1   1   1   1
#> 257   117 8.417093e-02   1   1   1   1
#> 258   200 4.621832e-01   1   1   1   1
#> 259    38 2.053143e-02   1   1   1   1
#> 260    93 6.141575e-02   1   1   1   1
#> 261   167 1.591893e-01   1   1   1   1
#> 262   191 2.530230e-01   1   1   1   1
#> 263    21 8.982494e-03   1   1   1   1
#> 264   122 9.035067e-02   1   1   1   1
#> 265    18 7.390545e-03   1   1   1   1
#> 266   115 8.249099e-02   1   1   1   1
#> 267   186 2.212673e-01   1   1   1   1
#> 268   126 9.240343e-02   1   1   1   1
#> 269    71 4.211720e-02   1   1   1   1
#> 270    89 5.838994e-02   1   1   1   1
#> 271    65 3.816719e-02   1   1   1   1
#> 272   128 9.625691e-02   1   1   1   1
#> 273    19 7.466517e-03   1   1   1   1
#> 274    20 8.395199e-03   1   1   1   1
#> 275    79 4.724489e-02   1   1   1   1
#> 276   156 1.378421e-01   1   1   1   1
#> 277   143 1.151405e-01   1   1   1   1
#> 278    91 5.976017e-02   1   1   1   1
#> 279   122 9.008286e-02   1   1   1   1
#> 280   155 1.359957e-01   1   1   1   1
#> 281    89 5.780055e-02   1   1   1   1
#> 282   120 8.895118e-02   1   1   1   1
#> 283   125 9.231286e-02   1   1   1   1
#> 284    36 1.914908e-02   1   1   1   1
#> 285   171 1.659989e-01   1   1   1   1
#> 286    89 5.792556e-02   1   1   1   1
#> 287   129 9.672650e-02   1   1   1   1
#> 288   135 1.038501e-01   1   1   1   1
#> 289   119 8.757467e-02   1   1   1   1
#> 290    46 2.413965e-02   1   1   1   1
#> 291    97 6.557448e-02   1   1   1   1
#> 292    96 6.369966e-02   1   1   1   1
#> 293    94 6.209577e-02   1   1   1   1
#> 294    34 1.753436e-02   1   1   1   1
#> 295     7 2.161838e-03   1   1   1   1
#> 296   155 1.355087e-01   1   1   1   1
#> 297   167 1.589213e-01   1   1   1   1
#> 298   112 8.005102e-02   1   1   1   1
#> 299    18 6.962803e-03   1   1   1   1
#> 300    84 5.269500e-02   1   1   1   1
#> 301    30 1.504259e-02   1   1   1   1
#> 302    78 4.637522e-02   1   1   1   1
#> 303   173 1.698463e-01   1   1   1   1
#> 304    73 4.324910e-02   1   1   1   1
#> 305   153 1.320381e-01   1   1   1   1
#> 306   194 2.791201e-01   1   1   1   1
#> 307   191 2.472087e-01   1   1   1   1
#> 308   175 1.749988e-01   1   1   1   1
#> 309    16 6.249147e-03   1   1   1   1
#> 310    14 5.619806e-03   1   1   1   1
#> 311    85 5.296776e-02   1   1   1   1
#> 312    87 5.603197e-02   1   1   1   1
#> 313   103 7.275965e-02   1   1   1   1
#> 314   110 7.834095e-02   1   1   1   1
#> 315    15 5.889875e-03   1   1   1   1
#> 316   164 1.528535e-01   1   1   1   1
#> 317   196 3.073358e-01   1   1   1   1
#> 318    69 4.115294e-02   1   1   1   1
#> 319    46 2.447101e-02   1   1   1   1
#> 320    46 2.408359e-02   1   1   1   1
#> 321    32 1.642557e-02   1   1   1   1
#> 322   101 6.932920e-02   1   1   1   1
#> 323   158 1.401495e-01   1   1   1   1
#> 324   182 2.040471e-01   1   1   1   1
#> 325   132 9.985108e-02   1   1   1   1
#> 326   189 2.404306e-01   1   1   1   1
#> 327   169 1.610909e-01   1   1   1   1
#> 328   154 1.331370e-01   1   1   1   1
#> 329   196 3.202404e-01   1   1   1   1
#> 330   163 1.501012e-01   1   1   1   1
#> 331    95 6.292185e-02   1   1   1   1
#> 332    86 5.427262e-02   1   1   1   1
#> 333    79 4.695285e-02   1   1   1   1
#> 334    55 3.082042e-02   1   1   1   1
#> 335    32 1.668581e-02   1   1   1   1
#> 336   141 1.138386e-01   1   1   1   1
#> 337   127 9.522762e-02   1   1   1   1
#> 338    86 5.476057e-02   1   1   1   1
#> 339    13 5.414013e-03   1   1   1   1
#> 340    26 1.182310e-02   1   1   1   1
#> 341    88 5.728363e-02   1   1   1   1
#> 342   148 1.225183e-01   1   1   1   1
#> 343    31 1.619332e-02   1   1   1   1
#> 344     5 1.498456e-03   1   1   1   1
#> 345   178 1.884705e-01   1   1   1   1
#> 346    11 3.703197e-03   1   1   1   1
#> 347    53 2.828618e-02   1   1   1   1
#> 348   184 2.156725e-01   1   1   1   1
#> 349     5 1.485301e-03   1   1   1   1
#> 350   199 4.135695e-01   1   1   1   1
#> 351   174 1.719843e-01   1   1   1   1
#> 352   117 8.464460e-02   1   1   1   1
#> 353   186 2.209714e-01   1   1   1   1
#> 354   144 1.179909e-01   1   1   1   1
#> 355    63 3.596230e-02   1   1   1   1
#> 356    25 1.173298e-02   1   1   1   1
#> 357   146 1.196068e-01   1   1   1   1
#> 358   191 2.530039e-01   1   1   1   1
#> 359   181 1.991457e-01   1   1   1   1
#> 360   117 8.461633e-02   1   1   1   1
#> 361    95 6.260667e-02   1   1   1   1
#> 362    40 2.173234e-02   1   1   1   1
#> 363    96 6.359418e-02   1   1   1   1
#> 364    91 5.973220e-02   1   1   1   1
#> 365   128 9.577601e-02   1   1   1   1
#> 366   200 4.710833e-01   1   1   1   1
#> 367   194 2.952878e-01   1   1   1   1
#> 368   139 1.108160e-01   1   1   1   1
#> 369   156 1.380031e-01   1   1   1   1
#> 370   100 6.769528e-02   1   1   1   1
#> 371   109 7.773913e-02   1   1   1   1
#> 372     6 1.725931e-03   1   1   1   1
#> 373   104 7.306864e-02   1   1   1   1
#> 374   141 1.131099e-01   1   1   1   1
#> 375     6 1.775324e-03   1   1   1   1
#> 376    15 5.930275e-03   1   1   1   1
#> 377   118 8.587146e-02   1   1   1   1
#> 378    95 6.260789e-02   1   1   1   1
#> 379    18 7.330009e-03   1   1   1   1
#> 380     8 2.350835e-03   1   1   1   1
#> 381   195 2.985048e-01   1   1   1   1
#> 382    61 3.482779e-02   1   1   1   1
#> 383   157 1.388171e-01   1   1   1   1
#> 384   145 1.183338e-01   1   1   1   1
#> 385   143 1.156241e-01   1   1   1   1
#> 386    12 4.631122e-03   1   1   1   1
#> 387    60 3.380446e-02   1   1   1   1
#> 388   113 8.082659e-02   1   1   1   1
#> 389    30 1.515260e-02   1   1   1   1
#> 390   174 1.748125e-01   1   1   1   1
#> 391    54 3.005735e-02   1   1   1   1
#> 392    87 5.563830e-02   1   1   1   1
#> 393   144 1.169692e-01   1   1   1   1
#> 394     2 5.899862e-04   1   1   1   1
#> 395    92 6.021884e-02   1   1   1   1
#> 396    90 5.870687e-02   1   1   1   1
#> 397   101 6.858544e-02   1   1   1   1
#> 398    41 2.207964e-02   1   1   1   1
#> 399     7 2.155819e-03   1   1   1   1
#> 400   190 2.439241e-01   1   1   1   1
#> 401   134 1.033405e-01   1   1   1   1
#> 402    36 1.925796e-02   1   1   1   1
#> 403   137 1.076368e-01   1   1   1   1
#> 404   109 7.728465e-02   1   1   1   1
#> 405   138 1.088160e-01   1   1   1   1
#> 406    37 2.004901e-02   1   1   1   1
#> 407    20 8.485060e-03   1   1   1   1
#> 408   127 9.480347e-02   1   1   1   1
#> 409   181 1.990112e-01   1   1   1   1
#> 410   126 9.295709e-02   1   1   1   1
#> 411   143 1.153787e-01   1   1   1   1
#> 412   168 1.605180e-01   1   1   1   1
#> 413    63 3.580913e-02   1   1   1   1
#> 414   109 7.724502e-02   1   1   1   1
#> 415    23 9.918856e-03   1   1   1   1
#> 416   123 9.090975e-02   1   1   1   1
#> 417   102 7.165329e-02   1   1   1   1
#> 418    42 2.266921e-02   1   1   1   1
#> 419    88 5.680322e-02   1   1   1   1
#> 420   100 6.727381e-02   1   1   1   1
#> 421    38 2.068891e-02   1   1   1   1
#> 422    91 5.948481e-02   1   1   1   1
#> 423   145 1.183525e-01   1   1   1   1
#> 424   199 4.404020e-01   1   1   1   1
#> 425   200 5.101531e-01   1   1   1   1
#> 426   107 7.638827e-02   1   1   1   1
#> 427    18 7.387583e-03   1   1   1   1
#> 428   187 2.307867e-01   1   1   1   1
#> 429    86 5.464415e-02   1   1   1   1
#> 430    63 3.617837e-02   1   1   1   1
#> 431    62 3.508252e-02   1   1   1   1
#> 432   129 9.702797e-02   1   1   1   1
#> 433   179 1.946303e-01   1   1   1   1
#> 434   174 1.749438e-01   1   1   1   1
#> 435    81 4.926899e-02   1   1   1   1
#> 436   181 1.986476e-01   1   1   1   1
#> 437    64 3.713662e-02   1   1   1   1
#> 438    61 3.499397e-02   1   1   1   1
#> 439    39 2.121466e-02   1   1   1   1
#> 440    47 2.471946e-02   1   1   1   1
#> 441     4 1.450229e-03   1   1   1   1
#> 442    68 4.011977e-02   1   1   1   1
#> 443    75 4.473770e-02   1   1   1   1
#> 444    26 1.173526e-02   1   1   1   1
#> 445   117 8.402500e-02   1   1   1   1
#> 446   147 1.209092e-01   1   1   1   1
#> 447    98 6.632733e-02   1   1   1   1
#> 448   130 9.766614e-02   1   1   1   1
#> 449    30 1.511525e-02   1   1   1   1
#> 450   116 8.251560e-02   1   1   1   1
#> 451    39 2.079316e-02   1   1   1   1
#> 452    84 5.234937e-02   1   1   1   1
#> 453   111 7.956825e-02   1   1   1   1
#> 454   123 9.055254e-02   1   1   1   1
#> 455   164 1.521099e-01   1   1   1   1
#> 456    84 5.277594e-02   1   1   1   1
#> 457    24 1.062222e-02   1   1   1   1
#> 458     7 2.116180e-03   1   1   1   1
#> 459     7 2.095253e-03   1   1   1   1
#> 460   111 7.889396e-02   1   1   1   1
#> 461    78 4.668070e-02   1   1   1   1
#> 462    15 5.954406e-03   1   1   1   1
#> 463   112 8.005369e-02   1   1   1   1
#> 464   104 7.285158e-02   1   1   1   1
#> 465   165 1.545566e-01   1   1   1   1
#> 466   195 2.997772e-01   1   1   1   1
#> 467   115 8.248672e-02   1   1   1   1
#> 468    25 1.110174e-02   1   1   1   1
#> 469   115 8.236668e-02   1   1   1   1
#> 470   156 1.372007e-01   1   1   1   1
#> 471    76 4.486096e-02   1   1   1   1
#> 472    98 6.594928e-02   1   1   1   1
#> 473   189 2.429267e-01   1   1   1   1
#> 474   155 1.340828e-01   1   1   1   1
#> 475   132 1.002618e-01   1   1   1   1
#> 476   168 1.592800e-01   1   1   1   1
#> 477    22 9.315641e-03   1   1   1   1
#> 478    18 7.235473e-03   1   1   1   1
#> 479    32 1.662031e-02   1   1   1   1
#> 480   163 1.505893e-01   1   1   1   1
#> 481   149 1.243299e-01   1   1   1   1
#> 482    28 1.324201e-02   1   1   1   1
#> 483    41 2.235030e-02   1   1   1   1
#> 484   100 6.800266e-02   1   1   1   1
#> 485   102 7.089079e-02   1   1   1   1
#> 486   134 1.031936e-01   1   1   1   1
#> 487    23 9.565051e-03   1   1   1   1
#> 488     9 2.547186e-03   1   1   1   1
#> 489   179 1.907963e-01   1   1   1   1
#> 490    38 2.054290e-02   1   1   1   1
#> 491   188 2.310393e-01   1   1   1   1
#> 492   112 8.019449e-02   1   1   1   1
#> 493    89 5.805033e-02   1   1   1   1
#> 494   129 9.699138e-02   1   1   1   1
#> 495    81 4.869719e-02   1   1   1   1
#> 496   127 9.484631e-02   1   1   1   1
#> 497    22 9.259819e-03   1   1   1   1
#> 498   108 7.677239e-02   1   1   1   1
#> 499     2 7.134360e-04   1   1   1   1
#> 500   165 1.552620e-01   1   1   1   1
#> 501   182 2.009366e-01   1   1   1   1
#> 502   108 7.701168e-02   1   1   1   1
#> 503   156 1.377543e-01   1   1   1   1
#> 504   122 9.005908e-02   1   1   1   1
#> 505    11 4.282036e-03   1   1   1   1
#> 506   132 9.997725e-02   1   1   1   1
#> 507    44 2.355545e-02   1   1   1   1
#> 508   196 3.081291e-01   1   1   1   1
#> 509   155 1.339248e-01   1   1   1   1
#> 510    27 1.318951e-02   1   1   1   1
#> 511   135 1.042965e-01   1   1   1   1
#> 512    85 5.374147e-02   1   1   1   1
#> 513   184 2.126107e-01   1   1   1   1
#> 514   170 1.655036e-01   1   1   1   1
#> 515   106 7.530474e-02   1   1   1   1
#> 516    37 2.020722e-02   1   1   1   1
#> 517   108 7.655643e-02   1   1   1   1
#> 518    42 2.235572e-02   1   1   1   1
#> 519    13 5.238045e-03   1   1   1   1
#> 520   194 2.842854e-01   1   1   1   1
#> 521    84 5.259363e-02   1   1   1   1
#> 522    42 2.287941e-02   1   1   1   1
#> 523    73 4.335163e-02   1   1   1   1
#> 524    73 4.322114e-02   1   1   1   1
#> 525    54 3.028556e-02   1   1   1   1
#> 526   192 2.560511e-01   1   1   1   1
#> 527    94 6.228134e-02   1   1   1   1
#> 528    60 3.448903e-02   1   1   1   1
#> 529    87 5.570381e-02   1   1   1   1
#> 530    88 5.648375e-02   1   1   1   1
#> 531   113 8.026391e-02   1   1   1   1
#> 532   111 7.893834e-02   1   1   1   1
#> 533    22 9.352524e-03   1   1   1   1
#> 534   142 1.143386e-01   1   1   1   1
#> 535    68 3.996379e-02   1   1   1   1
#> 536   153 1.317977e-01   1   1   1   1
#> 537   174 1.725316e-01   1   1   1   1
#> 538    87 5.584400e-02   1   1   1   1
#> 539    12 4.622436e-03   1   1   1   1
#> 540    44 2.346891e-02   1   1   1   1
#> 541     6 1.693299e-03   1   1   1   1
#> 542    51 2.694157e-02   1   1   1   1
#> 543   103 7.247008e-02   1   1   1   1
#> 544    93 6.199027e-02   1   1   1   1
#> 545    95 6.272635e-02   1   1   1   1
#> 546    26 1.240757e-02   1   1   1   1
#> 547   162 1.490214e-01   1   1   1   1
#> 548     5 1.528298e-03   1   1   1   1
#> 549   133 1.008210e-01   1   1   1   1
#> 550   159 1.419519e-01   1   1   1   1
#> 551    31 1.573548e-02   1   1   1   1
#> 552    11 3.776806e-03   1   1   1   1
#> 553   151 1.281230e-01   1   1   1   1
#> 554    74 4.401221e-02   1   1   1   1
#> 555    33 1.741033e-02   1   1   1   1
#> 556   181 1.984053e-01   1   1   1   1
#> 557   119 8.672138e-02   1   1   1   1
#> 558   120 8.824469e-02   1   1   1   1
#> 559    50 2.633903e-02   1   1   1   1
#> 560   152 1.311791e-01   1   1   1   1
#> 561    20 8.466600e-03   1   1   1   1
#> 562   116 8.400911e-02   1   1   1   1
#> 563   182 2.042709e-01   1   1   1   1
#> 564    10 3.222850e-03   1   1   1   1
#> 565    31 1.572282e-02   1   1   1   1
#> 566    13 5.218188e-03   1   1   1   1
#> 567   114 8.127058e-02   1   1   1   1
#> 568    56 3.107970e-02   1   1   1   1
#> 569   137 1.075466e-01   1   1   1   1
#> 570    70 4.157278e-02   1   1   1   1
#> 571   118 8.596101e-02   1   1   1   1
#> 572   177 1.868458e-01   1   1   1   1
#> 573   112 7.982562e-02   1   1   1   1
#> 574    53 2.853526e-02   1   1   1   1
#> 575   108 7.647471e-02   1   1   1   1
#> 576   124 9.105894e-02   1   1   1   1
#> 577    94 6.260551e-02   1   1   1   1
#> 578   116 8.344556e-02   1   1   1   1
#> 579    92 6.072814e-02   1   1   1   1
#> 580   121 8.982311e-02   1   1   1   1
#> 581    38 2.064821e-02   1   1   1   1
#> 582   140 1.122944e-01   1   1   1   1
#> 583   146 1.189999e-01   1   1   1   1
#> 584   190 2.442423e-01   1   1   1   1
#> 585    52 2.811035e-02   1   1   1   1
#> 586   171 1.666447e-01   1   1   1   1
#> 587   104 7.309899e-02   1   1   1   1
#> 588   183 2.095078e-01   1   1   1   1
#> 589     9 3.057984e-03   1   1   1   1
#> 590   102 7.117879e-02   1   1   1   1
#> 591    16 6.073095e-03   1   1   1   1
#> 592    14 5.515950e-03   1   1   1   1
#> 593   200 4.656495e-01   1   1   1   1
#> 594   198 3.807425e-01   1   1   1   1
#> 595    99 6.683034e-02   1   1   1   1
#> 596     8 2.329639e-03   1   1   1   1
#> 597   175 1.765340e-01   1   1   1   1
#> 598   101 6.945759e-02   1   1   1   1
#> 599    59 3.316883e-02   1   1   1   1
#> 600    17 6.944962e-03   1   1   1   1
#> 601   153 1.331084e-01   1   1   1   1
#> 602    97 6.506327e-02   1   1   1   1
#> 603   161 1.469267e-01   1   1   1   1
#> 604   119 8.723606e-02   1   1   1   1
#> 605   184 2.117104e-01   1   1   1   1
#> 606    98 6.666113e-02   1   1   1   1
#> 607    85 5.331278e-02   1   1   1   1
#> 608   148 1.234627e-01   1   1   1   1
#> 609    40 2.145958e-02   1   1   1   1
#> 610    92 6.016820e-02   1   1   1   1
#> 611   173 1.714167e-01   1   1   1   1
#> 612   142 1.147913e-01   1   1   1   1
#> 613   147 1.210655e-01   1   1   1   1
#> 614   143 1.154349e-01   1   1   1   1
#> 615    81 4.914571e-02   1   1   1   1
#> 616   114 8.137365e-02   1   1   1   1
#> 617   160 1.452273e-01   1   1   1   1
#> 618    28 1.334750e-02   1   1   1   1
#> 619    43 2.319328e-02   1   1   1   1
#> 620   120 8.852371e-02   1   1   1   1
#> 621    33 1.748432e-02   1   1   1   1
#> 622     4 1.312621e-03   1   1   1   1
#> 623    10 3.643247e-03   1   1   1   1
#> 624    84 5.254038e-02   1   1   1   1
#> 625    92 6.022260e-02   1   1   1   1
#> 626   168 1.607501e-01   1   1   1   1
#> 627   194 2.886144e-01   1   1   1   1
#> 628    59 3.320669e-02   1   1   1   1
#> 629    64 3.675514e-02   1   1   1   1
#> 630    13 5.310548e-03   1   1   1   1
#> 631   188 2.312207e-01   1   1   1   1
#> 632    45 2.398736e-02   1   1   1   1
#> 633   106 7.522154e-02   1   1   1   1
#> 634    91 5.978785e-02   1   1   1   1
#> 635    37 2.018514e-02   1   1   1   1
#> 636    47 2.465120e-02   1   1   1   1
#> 637    91 5.994832e-02   1   1   1   1
#> 638    71 4.252217e-02   1   1   1   1
#> 639    95 6.305816e-02   1   1   1   1
#> 640   160 1.456739e-01   1   1   1   1
#> 641   195 3.000699e-01   1   1   1   1
#> 642    26 1.180412e-02   1   1   1   1
#> 643    55 3.035979e-02   1   1   1   1
#> 644   113 8.125941e-02   1   1   1   1
#> 645   130 9.762309e-02   1   1   1   1
#> 646   113 8.114996e-02   1   1   1   1
#> 647    61 3.492856e-02   1   1   1   1
#> 648    48 2.480025e-02   1   1   1   1
#> 649   162 1.477301e-01   1   1   1   1
#> 650   141 1.138380e-01   1   1   1   1
#> 651   183 2.107392e-01   1   1   1   1
#> 652   193 2.723607e-01   1   1   1   1
#> 653    83 5.215252e-02   1   1   1   1
#> 654   161 1.474773e-01   1   1   1   1
#> 655   148 1.217957e-01   1   1   1   1
#> 656   166 1.570020e-01   1   1   1   1
#> 657   195 3.036822e-01   1   1   1   1
#> 658    43 2.304297e-02   1   1   1   1
#> 659    85 5.291945e-02   1   1   1   1
#> 660   185 2.167886e-01   1   1   1   1
#> 661   172 1.672148e-01   1   1   1   1
#> 662   177 1.869233e-01   1   1   1   1
#> 663    38 2.022587e-02   1   1   1   1
#> 664    23 9.741368e-03   1   1   1   1
#> 665   123 9.102363e-02   1   1   1   1
#> 666   130 9.728989e-02   1   1   1   1
#> 667   107 7.598377e-02   1   1   1   1
#> 668    98 6.564051e-02   1   1   1   1
#> 669   159 1.414826e-01   1   1   1   1
#> 670    62 3.510369e-02   1   1   1   1
#> 671    45 2.377697e-02   1   1   1   1
#> 672    99 6.723730e-02   1   1   1   1
#> 673    55 3.080898e-02   1   1   1   1
#> 674   188 2.379490e-01   1   1   1   1
#> 675    93 6.192950e-02   1   1   1   1
#> 676    49 2.560415e-02   1   1   1   1
#> 677   173 1.709057e-01   1   1   1   1
#> 678     1 1.364646e-04   1   1   1   1
#> 679    82 4.932765e-02   1   1   1   1
#> 680   166 1.560664e-01   1   1   1   1
#> 681   157 1.383147e-01   1   1   1   1
#> 682    80 4.789662e-02   1   1   1   1
#> 683   160 1.450389e-01   1   1   1   1
#> 684   124 9.128101e-02   1   1   1   1
#> 685    12 4.660944e-03   1   1   1   1
#> 686    58 3.218246e-02   1   1   1   1
#> 687   187 2.300256e-01   1   1   1   1
#> 688   178 1.899379e-01   1   1   1   1
#> 689   170 1.659762e-01   1   1   1   1
#> 690   141 1.136242e-01   1   1   1   1
#> 691   190 2.461961e-01   1   1   1   1
#> 692    67 3.926007e-02   1   1   1   1
#> 693   136 1.057649e-01   1   1   1   1
#> 694   149 1.243115e-01   1   1   1   1
#> 695   159 1.425329e-01   1   1   1   1
#> 696   158 1.392374e-01   1   1   1   1
#> 697   165 1.546063e-01   1   1   1   1
#> 698    37 1.994956e-02   1   1   1   1
#> 699    90 5.885417e-02   1   1   1   1
#> 700   139 1.108659e-01   1   1   1   1
#> 701    66 3.916157e-02   1   1   1   1
#> 702    59 3.284474e-02   1   1   1   1
#> 703   158 1.410922e-01   1   1   1   1
#> 704   164 1.528427e-01   1   1   1   1
#> 705    24 9.983083e-03   1   1   1   1
#> 706   193 2.678231e-01   1   1   1   1
#> 707    96 6.320499e-02   1   1   1   1
#> 708   107 7.575461e-02   1   1   1   1
#> 709    60 3.387579e-02   1   1   1   1
#> 710   106 7.394803e-02   1   1   1   1
#> 711   184 2.112336e-01   1   1   1   1
#> 712     8 2.213610e-03   1   1   1   1
#> 713   131 9.806927e-02   1   1   1   1
#> 714   193 2.632207e-01   1   1   1   1
#> 715    28 1.360737e-02   1   1   1   1
#> 716   135 1.043145e-01   1   1   1   1
#> 717   167 1.570901e-01   1   1   1   1
#> 718    56 3.097215e-02   1   1   1   1
#> 719   157 1.385023e-01   1   1   1   1
#> 720   192 2.548984e-01   1   1   1   1
#> 721   129 9.704900e-02   1   1   1   1
#> 722    80 4.805936e-02   1   1   1   1
#> 723    29 1.427153e-02   1   1   1   1
#> 724    30 1.562034e-02   1   1   1   1
#> 725    27 1.309500e-02   1   1   1   1
#> 726   146 1.202022e-01   1   1   1   1
#> 727   100 6.796620e-02   1   1   1   1
#> 728   151 1.295560e-01   1   1   1   1
#> 729    47 2.465525e-02   1   1   1   1
#> 730   144 1.178940e-01   1   1   1   1
#> 731   183 2.071167e-01   1   1   1   1
#> 732   163 1.513454e-01   1   1   1   1
#> 733   127 9.532163e-02   1   1   1   1
#> 734    21 8.654327e-03   1   1   1   1
#> 735    25 1.085014e-02   1   1   1   1
#> 736   136 1.059582e-01   1   1   1   1
#> 737    66 3.889393e-02   1   1   1   1
#> 738   180 1.961779e-01   1   1   1   1
#> 739    24 1.052235e-02   1   1   1   1
#> 740    71 4.229291e-02   1   1   1   1
#> 741   150 1.264536e-01   1   1   1   1
#> 742    74 4.387701e-02   1   1   1   1
#> 743    40 2.174460e-02   1   1   1   1
#> 744    70 4.138149e-02   1   1   1   1
#> 745   169 1.634455e-01   1   1   1   1
#> 746   199 4.309761e-01   1   1   1   1
#> 747    67 3.955708e-02   1   1   1   1
#> 748    42 2.285123e-02   1   1   1   1
#> 749     4 1.349071e-03   1   1   1   1
#> 750   193 2.719844e-01   1   1   1   1
#> 751   141 1.133881e-01   1   1   1   1
#> 752    72 4.254453e-02   1   1   1   1
#> 753     4 1.316801e-03   1   1   1   1
#> 754   172 1.679226e-01   1   1   1   1
#> 755   105 7.323930e-02   1   1   1   1
#> 756    10 3.115755e-03   1   1   1   1
#> 757   199 3.823631e-01   1   1   1   1
#> 758    82 4.935304e-02   1   1   1   1
#> 759   184 2.135240e-01   1   1   1   1
#> 760   196 3.148478e-01   1   1   1   1
#> 761   152 1.313863e-01   1   1   1   1
#> 762    87 5.557920e-02   1   1   1   1
#> 763    79 4.698428e-02   1   1   1   1
#> 764   156 1.368349e-01   1   1   1   1
#> 765     1 3.389406e-05   1   1   1   1
#> 766    34 1.763281e-02   1   1   1   1
#> 767     3 1.259068e-03   1   1   1   1
#> 768    73 4.349930e-02   1   1   1   1
#> 769     9 2.846905e-03   1   1   1   1
#> 770    48 2.474514e-02   1   1   1   1
#> 771    21 8.799329e-03   1   1   1   1
#> 772    70 4.130199e-02   1   1   1   1
#> 773    59 3.364301e-02   1   1   1   1
#> 774   169 1.614634e-01   1   1   1   1
#> 775    92 6.089811e-02   1   1   1   1
#> 776   104 7.289362e-02   1   1   1   1
#> 777   150 1.262107e-01   1   1   1   1
#> 778    45 2.405671e-02   1   1   1   1
#> 779   183 2.063502e-01   1   1   1   1
#> 780    99 6.725072e-02   1   1   1   1
#> 781    51 2.702427e-02   1   1   1   1
#> 782    75 4.481365e-02   1   1   1   1
#> 783   148 1.229281e-01   1   1   1   1
#> 784    21 8.622818e-03   1   1   1   1
#> 785   162 1.480547e-01   1   1   1   1
#> 786    68 4.037219e-02   1   1   1   1
#> 787    69 4.056982e-02   1   1   1   1
#> 788    40 2.175135e-02   1   1   1   1
#> 789    82 4.980420e-02   1   1   1   1
#> 790    24 1.043304e-02   1   1   1   1
#> 791   121 8.987303e-02   1   1   1   1
#> 792    77 4.573012e-02   1   1   1   1
#> 793   131 9.813239e-02   1   1   1   1
#> 794    74 4.383538e-02   1   1   1   1
#> 795   154 1.338652e-01   1   1   1   1
#> 796   123 9.061056e-02   1   1   1   1
#> 797    39 2.133466e-02   1   1   1   1
#> 798    31 1.623681e-02   1   1   1   1
#> 799   143 1.159780e-01   1   1   1   1
#> 800    75 4.451270e-02   1   1   1   1
#> 801    55 3.088325e-02   1   1   1   1
#> 802    62 3.510687e-02   1   1   1   1
#> 803    42 2.280504e-02   1   1   1   1
#> 804   162 1.490155e-01   1   1   1   1
#> 805    10 3.509659e-03   1   1   1   1
#> 806     1 2.171188e-04   1   1   1   1
#> 807    39 2.089998e-02   1   1   1   1
#> 808   128 9.622346e-02   1   1   1   1
#> 809    72 4.265030e-02   1   1   1   1
#> 810   190 2.439099e-01   1   1   1   1
#> 811   177 1.863220e-01   1   1   1   1
#> 812    52 2.732130e-02   1   1   1   1
#> 813    65 3.802048e-02   1   1   1   1
#> 814    78 4.597096e-02   1   1   1   1
#> 815   178 1.893080e-01   1   1   1   1
#> 816   118 8.569690e-02   1   1   1   1
#> 817   122 9.000850e-02   1   1   1   1
#> 818   150 1.258637e-01   1   1   1   1
#> 819   166 1.558242e-01   1   1   1   1
#> 820   170 1.640214e-01   1   1   1   1
#> 821   161 1.471806e-01   1   1   1   1
#> 822   110 7.797167e-02   1   1   1   1
#> 823   175 1.788430e-01   1   1   1   1
#> 824    79 4.693108e-02   1   1   1   1
#> 825   134 1.035918e-01   1   1   1   1
#> 826   186 2.247229e-01   1   1   1   1
#> 827   139 1.094240e-01   1   1   1   1
#> 828    29 1.399519e-02   1   1   1   1
#> 829   162 1.479402e-01   1   1   1   1
#> 830   186 2.240320e-01   1   1   1   1
#> 831    33 1.689481e-02   1   1   1   1
#> 832   145 1.187356e-01   1   1   1   1
#> 833    43 2.299900e-02   1   1   1   1
#> 834    59 3.344979e-02   1   1   1   1
#> 835   174 1.718383e-01   1   1   1   1
#> 836   106 7.534274e-02   1   1   1   1
#> 837   114 8.193880e-02   1   1   1   1
#> 838   191 2.516721e-01   1   1   1   1
#> 839   197 3.361264e-01   1   1   1   1
#> 840    49 2.556120e-02   1   1   1   1
#> 841    56 3.096468e-02   1   1   1   1
#> 842   134 1.036468e-01   1   1   1   1
#> 843    32 1.672379e-02   1   1   1   1
#> 844   198 3.437399e-01   1   1   1   1
#> 845   157 1.392149e-01   1   1   1   1
#> 846    70 4.151086e-02   1   1   1   1
#> 847   151 1.281067e-01   1   1   1   1
#> 848   112 7.958111e-02   1   1   1   1
#> 849   197 3.312954e-01   1   1   1   1
#> 850   161 1.464644e-01   1   1   1   1
#> 851    33 1.687473e-02   1   1   1   1
#> 852    46 2.451922e-02   1   1   1   1
#> 853     6 1.628737e-03   1   1   1   1
#> 854   142 1.149350e-01   1   1   1   1
#> 855    10 3.696022e-03   1   1   1   1
#> 856    73 4.310026e-02   1   1   1   1
#> 857   180 1.969559e-01   1   1   1   1
#> 858    66 3.899685e-02   1   1   1   1
#> 859    71 4.170017e-02   1   1   1   1
#> 860     2 9.687285e-04   1   1   1   1
#> 861    62 3.516490e-02   1   1   1   1
#> 862    11 4.078865e-03   1   1   1   1
#> 863    15 6.043964e-03   1   1   1   1
#> 864    74 4.367154e-02   1   1   1   1
#> 865    94 6.209175e-02   1   1   1   1
#> 866    29 1.390903e-02   1   1   1   1
#> 867   113 8.115411e-02   1   1   1   1
#> 868    44 2.374652e-02   1   1   1   1
#> 869   119 8.773491e-02   1   1   1   1
#> 870   166 1.570350e-01   1   1   1   1
#> 871     3 1.103350e-03   1   1   1   1
#> 872     9 2.829678e-03   1   1   1   1
#> 873    48 2.502603e-02   1   1   1   1
#> 874   164 1.528302e-01   1   1   1   1
#> 875    29 1.499265e-02   1   1   1   1
#> 876   185 2.159665e-01   1   1   1   1
#> 877   182 2.040193e-01   1   1   1   1
#> 878    13 4.995397e-03   1   1   1   1
#> 879   138 1.091853e-01   1   1   1   1
#> 880   101 6.863156e-02   1   1   1   1
#> 881   176 1.821711e-01   1   1   1   1
#> 882    66 3.840733e-02   1   1   1   1
#> 883   132 9.996644e-02   1   1   1   1
#> 884    35 1.779888e-02   1   1   1   1
#> 885   195 3.038537e-01   1   1   1   1
#> 886    99 6.670772e-02   1   1   1   1
#> 887    72 4.278119e-02   1   1   1   1
#> 888   121 8.985328e-02   1   1   1   1
#> 889   196 3.176361e-01   1   1   1   1
#> 890   182 1.995357e-01   1   1   1   1
#> 891   144 1.165813e-01   1   1   1   1
#> 892    89 5.775236e-02   1   1   1   1
#> 893   175 1.749698e-01   1   1   1   1
#> 894    16 6.298558e-03   1   1   1   1
#> 895    17 6.818651e-03   1   1   1   1
#> 896   136 1.048334e-01   1   1   1   1
#> 897   115 8.249227e-02   1   1   1   1
#> 898    41 2.209278e-02   1   1   1   1
#> 899    33 1.750438e-02   1   1   1   1
#> 900    90 5.846030e-02   1   1   1   1
#> 901   145 1.180535e-01   1   1   1   1
#> 902     3 1.103862e-03   1   1   1   1
#> 903   131 9.882505e-02   1   1   1   1
#> 904    99 6.716364e-02   1   1   1   1
#> 905   168 1.608223e-01   1   1   1   1
#> 906    61 3.493253e-02   1   1   1   1
#> 907   180 1.976580e-01   1   1   1   1
#> 908    48 2.541822e-02   1   1   1   1
#> 909    23 9.402970e-03   1   1   1   1
#> 910    50 2.642575e-02   1   1   1   1
#> 911   125 9.177515e-02   1   1   1   1
#> 912    98 6.560176e-02   1   1   1   1
#> 913    60 3.388266e-02   1   1   1   1
#> 914     1 3.000654e-04   1   1   1   1
#> 915    82 5.014590e-02   1   1   1   1
#> 916    29 1.501655e-02   1   1   1   1
#> 917    74 4.415018e-02   1   1   1   1
#> 918    64 3.734760e-02   1   1   1   1
#> 919    25 1.075513e-02   1   1   1   1
#> 920   122 9.052280e-02   1   1   1   1
#> 921    41 2.208740e-02   1   1   1   1
#> 922   177 1.869558e-01   1   1   1   1
#> 923    17 6.867851e-03   1   1   1   1
#> 924    75 4.468555e-02   1   1   1   1
#> 925    36 1.956656e-02   1   1   1   1
#> 926    27 1.243275e-02   1   1   1   1
#> 927    23 9.620753e-03   1   1   1   1
#> 928   120 8.902911e-02   1   1   1   1
#> 929   179 1.908787e-01   1   1   1   1
#> 930   160 1.440870e-01   1   1   1   1
#> 931   138 1.083789e-01   1   1   1   1
#> 932   175 1.763251e-01   1   1   1   1
#> 933   110 7.874352e-02   1   1   1   1
#> 934   131 9.871718e-02   1   1   1   1
#> 935   186 2.233122e-01   1   1   1   1
#> 936    58 3.211512e-02   1   1   1   1
#> 937    75 4.419294e-02   1   1   1   1
#> 938    50 2.615726e-02   1   1   1   1
#> 939   173 1.709258e-01   1   1   1   1
#> 940    51 2.730749e-02   1   1   1   1
#> 941     5 1.467328e-03   1   1   1   1
#> 942   124 9.158838e-02   1   1   1   1
#> 943   198 3.406856e-01   1   1   1   1
#> 944    27 1.245430e-02   1   1   1   1
#> 945    57 3.153600e-02   1   1   1   1
#> 946    53 2.951848e-02   1   1   1   1
#> 947    85 5.319292e-02   1   1   1   1
#> 948   108 7.707985e-02   1   1   1   1
#> 949   150 1.259586e-01   1   1   1   1
#> 950    28 1.327315e-02   1   1   1   1
#> 951    68 3.974672e-02   1   1   1   1
#> 952    22 9.331203e-03   1   1   1   1
#> 953   127 9.465291e-02   1   1   1   1
#> 954   118 8.530473e-02   1   1   1   1
#> 955   128 9.640127e-02   1   1   1   1
#> 956   102 7.204676e-02   1   1   1   1
#> 957    17 6.727814e-03   1   1   1   1
#> 958    78 4.618069e-02   1   1   1   1
#> 959   123 9.102693e-02   1   1   1   1
#> 960   176 1.837241e-01   1   1   1   1
#> 961   169 1.630017e-01   1   1   1   1
#> 962   142 1.147188e-01   1   1   1   1
#> 963    71 4.252261e-02   1   1   1   1
#> 964    67 3.960642e-02   1   1   1   1
#> 965    49 2.596532e-02   1   1   1   1
#> 966    76 4.519629e-02   1   1   1   1
#> 967   178 1.898731e-01   1   1   1   1
#> 968    19 8.133066e-03   1   1   1   1
#> 969    20 8.551364e-03   1   1   1   1
#> 970    78 4.648692e-02   1   1   1   1
#> 971   151 1.275687e-01   1   1   1   1
#> 972    24 1.069936e-02   1   1   1   1
#> 973    77 4.549157e-02   1   1   1   1
#> 974   199 4.112641e-01   1   1   1   1
#> 975     3 1.026085e-03   1   1   1   1
#> 976    45 2.381264e-02   1   1   1   1
#> 977   152 1.316728e-01   1   1   1   1
#> 978   152 1.300202e-01   1   1   1   1
#> 979    65 3.767945e-02   1   1   1   1
#> 980    51 2.728816e-02   1   1   1   1
#> 981   137 1.063413e-01   1   1   1   1
#> 982   185 2.207328e-01   1   1   1   1
#> 983    93 6.178613e-02   1   1   1   1
#> 984   197 3.234835e-01   1   1   1   1
#> 985    83 5.049310e-02   1   1   1   1
#> 986   198 3.650226e-01   1   1   1   1
#> 987   158 1.400027e-01   1   1   1   1
#> 988   135 1.039861e-01   1   1   1   1
#> 989   149 1.256708e-01   1   1   1   1
#> 990   191 2.504642e-01   1   1   1   1
#> 991    53 2.953820e-02   1   1   1   1
#> 992   114 8.159989e-02   1   1   1   1
#> 993   192 2.614774e-01   1   1   1   1
#> 994   146 1.196550e-01   1   1   1   1
#> 995    65 3.834298e-02   1   1   1   1
#> 996   135 1.044772e-01   1   1   1   1
#> 997    49 2.591039e-02   1   1   1   1
#> 998   177 1.846020e-01   1   1   1   1
#> 999   189 2.434457e-01   1   1   1   1
#> 1000  105 7.359500e-02   1   1   1   1
#> 

# Find the characteristics of an informative 
#   Dorfman algorithm, using the TOD procedure.
# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(1002)
p.vec <- p.vec.func(p=0.01, alpha=2, grp.sz=20)
opt.info.dorf(prob=p.vec, se=0.95, sp=0.95, 
method="TOD", max.pool=5, threshold=0.015)
#> $tv
#> [1] 0.015
#> 
#> $e
#> [1] 9.218941
#> 
#> $v
#> [1] 4.494338
#> 
#> $summary
#>    pool probability    PSe       PSp       PPV       NPV
#> 1     1 0.001583625 0.9025 0.9970628 0.3276716 0.9998449
#> 2     1 0.002500582 0.9025 0.9971038 0.4385710 0.9997549
#> 3     1 0.003269132 0.9025 0.9971382 0.5084377 0.9996794
#> 4     1 0.003976450 0.9025 0.9971699 0.5600752 0.9996098
#> 5     2 0.004656581 0.9025 0.9966937 0.5608303 0.9995426
#> 6     2 0.005328128 0.9025 0.9967235 0.5960398 0.9994763
#> 7     2 0.006003726 0.9025 0.9967536 0.6267425 0.9994095
#> 8     2 0.006693557 0.9025 0.9967843 0.6541289 0.9993413
#> 9     3 0.007407063 0.9025 0.9963005 0.6454465 0.9992703
#> 10    3 0.008154056 0.9025 0.9963335 0.6692676 0.9991961
#> 11    3 0.008945696 0.9025 0.9963685 0.6916672 0.9991175
#> 12    3 0.009795627 0.9025 0.9964062 0.7129940 0.9990329
#> 13    4 0.010721593 0.9025 0.9957710 0.6981442 0.9989400
#> 14    4 0.011748029 0.9025 0.9958159 0.7194277 0.9988374
#> 15    4 0.012910694 0.9025 0.9958669 0.7406675 0.9987211
#> 16    4 0.014265788 0.9025 0.9959265 0.7622676 0.9985852
#> 17    5 0.015910262 0.9500 0.9500000 0.2349958 0.9991498
#> 18    6 0.018035443 0.9500 0.9500000 0.2586921 0.9990343
#> 19    7 0.021112395 0.9500 0.9500000 0.2906730 0.9988661
#> 20    8 0.026981576 0.9500 0.9500000 0.3450635 0.9985427
#>