Generate a vector of probabilities for informative group testing algorithms.

Generate a vector of individual risk probabilities using an overall probability of disease (i.e., the expected value of order statistics from a beta distribution) for use with informative group testing algorithms.

p.vec.func(p, alpha, grp.sz)

Arguments

p

overall probability of disease that will be used to generate a vector of individual risk probabilities.

alpha

a shape parameter for the beta distribution that specifies the degree of heterogeneity for the generated probability vector.

grp.sz

the number of total individuals for which to generate risk probabilities.

Value

A vector of individual risk probabilities.

Details

This function uses Michael Black's beta.dist function to generate a vector of individual risk probabilities, ordered from least to greatest. Depending on the specified probability, alpha level, and overall group size, simulation may be necessary in order to generate the vector of individual probabilities. For this reason, the user should set a seed in order to reproduce results. The p.vec.func function augments the beta.dist function by checking whether simulation is needed before attempting to generate the vector of individual risk probabilities. See Black et al. (2015) for additional details on Michael Black's beta.dist function.

References

Black, M., Bilder, C., Tebbs, J. (2015). “Optimal retesting configurations for hierarchical group testing.” Journal of the Royal Statistical Society. Series C: Applied Statistics, 64(4), 693--710. ISSN 14679876, doi:10.1111/rssc.12097 .

See also

Author

Brianna D. Hitt

Examples

# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(8791)
p.vec.func(p=0.03, alpha=0.5, grp.sz=100)
#>   [1] 2.925177e-09 5.011808e-08 4.250085e-07 2.378237e-06 9.878101e-06
#>   [6] 3.248136e-05 8.806795e-05 2.024929e-04 4.030091e-04 7.052139e-04
#>  [11] 1.098426e-03 1.538070e-03 1.952034e-03 2.260888e-03 2.403719e-03
#>  [16] 2.357783e-03 2.143571e-03 1.814735e-03 1.439410e-03 1.082062e-03
#>  [21] 7.926536e-04 6.051135e-04 5.427700e-04 6.262520e-04 8.794012e-04
#>  [26] 1.330185e-03 2.005561e-03 2.921243e-03 4.069071e-03 5.405976e-03
#>  [31] 6.848836e-03 8.278571e-03 9.554485e-03 1.053674e-02 1.111207e-02
#>  [36] 1.121626e-02 1.084774e-02 1.006900e-02 8.996372e-03 7.781692e-03
#>  [41] 6.591360e-03 5.587707e-03 4.915683e-03 4.694973e-03 5.015300e-03
#>  [46] 5.931677e-03 7.457191e-03 9.553149e-03 1.211922e-02 1.362259e-02
#>  [51] 1.486163e-02 1.550586e-02 1.617942e-02 1.690843e-02 1.766220e-02
#>  [56] 1.844176e-02 1.924816e-02 2.008250e-02 2.094594e-02 2.183977e-02
#>  [61] 2.276542e-02 2.372439e-02 2.471829e-02 2.574888e-02 2.681807e-02
#>  [66] 2.792791e-02 2.908066e-02 3.027878e-02 3.152494e-02 3.282210e-02
#>  [71] 3.417350e-02 3.558271e-02 3.705369e-02 3.859084e-02 4.019906e-02
#>  [76] 4.188384e-02 4.365135e-02 4.550856e-02 4.746339e-02 4.952488e-02
#>  [81] 5.170341e-02 5.401099e-02 5.646161e-02 5.907173e-02 6.186082e-02
#>  [86] 6.485225e-02 6.807431e-02 7.156176e-02 7.535794e-02 7.951786e-02
#>  [91] 8.411279e-02 8.923744e-02 9.502148e-02 1.016496e-01 1.093955e-01
#>  [96] 1.186968e-01 1.303094e-01 1.457257e-01 1.685973e-01 2.128764e-01

# This example takes less than 1 second to run.
# Estimated running time was calculated using a 
#   computer with 16 GB of RAM and one core of an 
#   Intel i7-6500U processor.
set.seed(52613)
p.vec.func(p=0.005, alpha=2, grp.sz=40)
#>  [1] 5.399042e-04 8.374973e-04 1.077003e-03 1.289020e-03 2.425413e-05
#>  [6] 4.080066e-05 1.849789e-03 2.024207e-03 5.573256e-05 1.457842e-05
#> [11] 3.321438e-06 6.656775e-07 1.224036e-07 4.481376e-08 1.371505e-07
#> [16] 6.195728e-07 2.565376e-06 9.585560e-06 3.235647e-05 4.135834e-03
#> [21] 4.334546e-03 4.540106e-03 4.753482e-03 4.975787e-03 5.208315e-03
#> [26] 5.452595e-03 5.710456e-03 5.984130e-03 6.276375e-03 6.590679e-03
#> [31] 6.931544e-03 7.304934e-03 7.719005e-03 8.185344e-03 8.721232e-03
#> [36] 9.354154e-03 1.013189e-02 1.114922e-02 1.264062e-02 1.552771e-02