Predict Method for Group Testing Model Fits

Obtains predictions for individual observations and optionally estimates standard errors of those predictions from objects of class "gt" or "gt.mp" returned by gtreg and gtreg.mp, respectively.

# S3 method for gt
predict(object, newdata, type = c("link", "response"),
 se.fit = FALSE, conf.level = NULL, na.action = na.pass, ...)

Arguments

object

a fitted object of class "gt" or "gt.mp".

newdata

optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.

type

the type of prediction required. The option "link" is on the scale of the linear predictors; the alternative "response" is on the scale of the response variable. Thus for the binomial model the "link" predictions are of log-odds (probabilities on logit scale) and type = "response" gives the predicted probabilities.

se.fit

logical switch indicating if standard errors are required.

conf.level

confidence level of the interval of the predicted values.

na.action

function determining what should be done with missing values in newdata. The default is to predict NA.

...

currently not used

Details

If newdata is omitted the predictions are based on the data used for the fit. When newdata is present and contains missing values, how the missing values will be dealt with is determined by the na.action argument. In this case, if na.action = na.omit omitted cases will not appear, whereas if na.action = na.exclude they will appear (in predictions and standard errors), with value NA. See also napredict.

Value

If se = FALSE, a vector or matrix of predictions. If se = TRUE, a list with components

fit

Predictions

se.fit

Estimated standard errors

lower

Lower bound of the confidence interval if calculated

upper

Upper bound of the confidence interval if calculated

Author

Boan Zhang

Examples

data(hivsurv)

fit1 <- gtreg(formula = groupres ~ AGE + EDUC., data = hivsurv,  
  groupn = gnum, sens = 0.9, spec = 0.9, linkf = "logit", method = "V")
pred.data <- data.frame(AGE = c(15, 25, 30), EDUC. = c(1, 3, 2))
predict(object = fit1, newdata = pred.data, type = "link", se.fit = TRUE)
#> $fit
#>         1         2         3 
#> -3.144613 -2.058004 -3.171061 
#> 
#> $se.fit
#>         1         2         3 
#> 0.8906438 0.3167530 0.7505289 
#> 
predict(object = fit1, newdata = pred.data, type = "response", 
  se.fit = TRUE, conf.level = 0.9)
#> $fit
#>          1          2          3 
#> 0.04130408 0.11324612 0.04026939 
#> 
#> $se.fit
#>          1          2          3 
#> 0.03526776 0.03180879 0.02900627 
#> 
#> $lower
#>           1           2           3 
#> 0.009857743 0.070501164 0.012061868 
#> 
#> $upper
#>         1         2         3 
#> 0.1571439 0.1769730 0.1260273 
#> 
predict(object = fit1, type = "response", se.fit = TRUE, conf.level = 0.9)
#> $fit
#>           1           2           3           4           5           6 
#> 0.268563644 0.085223130 0.037020397 0.080885022 0.035041958 0.069054266 
#>           7           8           9          10          11          12 
#> 0.050060792 0.030592292 0.080885022 0.076749210 0.125203217 0.113246119 
#>          13          14          15          16          17          18 
#> 0.113246119 0.052840747 0.107649422 0.097182503 0.092297050 0.013598314 
#>          19          20          21          22          23          24 
#> 0.030592292 0.012151600 0.060556716 0.031386464 0.145153802 0.131577971 
#>          25          26          27          28          29          30 
#> 0.131577971 0.131577971 0.119094956 0.113246119 0.052840747 0.268563644 
#>          31          32          33          34          35          36 
#> 0.131577971 0.125203217 0.050060792 0.085223130 0.085223130 0.076749210 
#>          37          38          39          40          41          42 
#> 0.035041958 0.062078948 0.023791095 0.102297408 0.092297050 0.042529960 
#>          43          44          45          46          47          48 
#> 0.019613298 0.041304079 0.175776758 0.080885022 0.035041958 0.031386464 
#>          49          50          51          52          53          54 
#> 0.145153802 0.065480287 0.062078948 0.125203217 0.055766015 0.023791095 
#>          55          56          57          58          59          60 
#> 0.080885022 0.159872483 0.031386464 0.031386464 0.062078948 0.062078948 
#>          61          62          63          64          65          66 
#> 0.017999611 0.016092286 0.076749210 0.031386464 0.069054266 0.058843165 
#>          67          68          69          70          71          72 
#> 0.023791095 0.017999611 0.159872483 0.065480287 0.065480287 0.119094956 
#>          73          74          75          76          77          78 
#> 0.023791095 0.138226011 0.131577971 0.058843165 0.113246119 0.107649422 
#>          79          80          81          82          83          84 
#> 0.047419768 0.092297050 0.060556716 0.054389605 0.072808118 0.125203217 
#>          85          86          87          88          89          90 
#> 0.119094956 0.047419768 0.080885022 0.078940461 0.039106009 0.080885022 
#>          91          92          93          94          95          96 
#> 0.076749210 0.072808118 0.069054266 0.058843165 0.058843165 0.055766015 
#>          97          98          99         100         101         102 
#> 0.107649422 0.102297408 0.044911488 0.044911488 0.087633353 0.063880331 
#>         103         104         105         106         107         108 
#> 0.037020397 0.069054266 0.131577971 0.023791095 0.113246119 0.050060792 
#>         109         110         111         112         113         114 
#> 0.097182503 0.040269394 0.013598314 0.010261983 0.069054266 0.268563644 
#>         115         116         117         118         119         120 
#> 0.052840747 0.102297408 0.097182503 0.089771179 0.119094956 0.052840747 
#>         121         122         123         124         125         126 
#> 0.016092286 0.167674093 0.033165610 0.152367420 0.069054266 0.055766015 
#>         127         128         129         130         131         132 
#> 0.236340569 0.107649422 0.107649422 0.050060792 0.087633353 0.089771179 
#>         133         134         135         136         137         138 
#> 0.085223130 0.076749210 0.069054266 0.055766015 0.113246119 0.022503133 
#>         139         140         141         142         143         144 
#> 0.047419768 0.017999611 0.010857107 0.145153802 0.019034834 0.017999611 
#>         145         146         147         148         149         150 
#> 0.069054266 0.145153802 0.125203217 0.058843165 0.246777638 0.119094956 
#>         151         152         153         154         155         156 
#> 0.216395177 0.230875349 0.080885022 0.031386464 0.062078948 0.058843165 
#>         157         158         159         160         161         162 
#> 0.119094956 0.113246119 0.113246119 0.092297050 0.092297050 0.083183715 
#>         163         164         165         166         167         168 
#> 0.016092286 0.074895975 0.074895975 0.067373279 0.039106009 0.076749210 
#>         169         170         171         172         173         174 
#> 0.031386464 0.062078948 0.107649422 0.044911488 0.020128375 0.020128375 
#>         175         176         177         178         179         180 
#> 0.092297050 0.171977757 0.171977757 0.085223130 0.257520191 0.119094956 
#>         181         182         183         184         185         186 
#> 0.107649422 0.047419768 0.206890137 0.047419768 0.197697268 0.013598314 
#>         187         188         189         190         191         192 
#> 0.039106009 0.028101197 0.125203217 0.020128375 0.076749210 0.031386464 
#>         193         194         195         196         197         198 
#> 0.069054266 0.138226011 0.138226011 0.062078948 0.055766015 0.050060792 
#>         199         200         201         202         203         204 
#> 0.047419768 0.021283377 0.102297408 0.097182503 0.040269394 0.038124197 
#>         205         206         207         208         209         210 
#> 0.078940461 0.025910296 0.072808118 0.138226011 0.026586258 0.052840747 
#>         211         212         213         214         215         216 
#> 0.023791095 0.044911488 0.087633353 0.035041958 0.076749210 0.031386464 
#>         217         218         219         220         221         222 
#> 0.017999611 0.069054266 0.138226011 0.044911488 0.152367420 0.050060792 
#>         223         224         225         226         227         228 
#> 0.072808118 0.069054266 0.069054266 0.031386464 0.065480287 0.257520191 
#>         229         230         231         232         233         234 
#> 0.017999611 0.040269394 0.028947057 0.028947057 0.011486344 0.085223130 
#>         235         236         237         238         239         240 
#> 0.085223130 0.028101197 0.216395177 0.206890137 0.016092286 0.078940461 
#>         241         242         243         244         245         246 
#> 0.074895975 0.080885022 0.080885022 0.131577971 0.062078948 0.083183715 
#>         247         248         249         250         251         252 
#> 0.055766015 0.036088980 0.072808118 0.069054266 0.138226011 0.065480287 
#>         253         254         255         256         257         258 
#> 0.246777638 0.055766015 0.074895975 0.279902149 0.145153802 0.065480287 
#>         259         260         261         262         263         264 
#> 0.028101197 0.062078948 0.131577971 0.236340569 0.020128375 0.019034834 
#>         265         266         267         268         269         270 
#> 0.017999611 0.076749210 0.031386464 0.119094956 0.052840747 0.062078948 
#>         271         272         273         274         275         276 
#> 0.050060792 0.036088980 0.072808118 0.119094956 0.078940461 0.145153802 
#>         277         278         279         280         281         282 
#> 0.031386464 0.028101197 0.055766015 0.015214614 0.012854886 0.008663632 
#>         283         284         285         286         287         288 
#> 0.069054266 0.014384110 0.138226011 0.131577971 0.125203217 0.021283377 
#>         289         290         291         292         293         294 
#> 0.067373279 0.023791095 0.020128375 0.069054266 0.062078948 0.131577971 
#>         295         296         297         298         299         300 
#> 0.047419768 0.102297408 0.092297050 0.083183715 0.033165610 0.055766015 
#>         301         302         303         304         305         306 
#> 0.050060792 0.036088980 0.039106009 0.159872483 0.033165610 0.062078948 
#>         307         308         309         310         311         312 
#> 0.102297408 0.044911488 0.080885022 0.145153802 0.029699828 0.065480287 
#>         313         314         315         316         317         318 
#> 0.028101197 0.026586258 0.058843165 0.050060792 0.197697268 0.015214614 
#>         319         320         321         322         323         324 
#> 0.141902698 0.069054266 0.131577971 0.062078948 0.055766015 0.197697268 
#>         325         326         327         328         329         330 
#> 0.037020397 0.159872483 0.033165610 0.062078948 0.062078948 0.025150875 
#>         331         332         333         334         335         336 
#> 0.119094956 0.022503133 0.040269394 0.013598314 0.037020397 0.076749210 
#>         337         338         339         340         341         342 
#> 0.069054266 0.131577971 0.026586258 0.125203217 0.246777638 0.102297408 
#>         343         344         345         346         347         348 
#> 0.007312391 0.085223130 0.175776758 0.080885022 0.035041958 0.035041958 
#>         349         350         351         352         353         354 
#> 0.076749210 0.031386464 0.069054266 0.025150875 0.025150875 0.023791095 
#>         355         356         357         358         359         360 
#> 0.107649422 0.020128375 0.011486344 0.080885022 0.076749210 0.076749210 
#>         361         362         363         364         365         366 
#> 0.055766015 0.159872483 0.072808118 0.069054266 0.065480287 0.131577971 
#>         367         368         369         370         371         372 
#> 0.125203217 0.058843165 0.246777638 0.055766015 0.055766015 0.040269394 
#>         373         374         375         376         377         378 
#> 0.072808118 0.052840747 0.072808118 0.031386464 0.069054266 0.145153802 
#>         379         380         381         382         383         384 
#> 0.055766015 0.119094956 0.052840747 0.050060792 0.107649422 0.020128375 
#>         385         386         387         388         389         390 
#> 0.019613298 0.076749210 0.076749210 0.138226011 0.029699828 0.131577971 
#>         391         392         393         394         395         396 
#> 0.131577971 0.058843165 0.055766015 0.025150875 0.119094956 0.055766015 
#>         397         398         399         400         401         402 
#> 0.023791095 0.052840747 0.052840747 0.044911488 0.042529960 0.042529960 
#>         403         404         405         406         407         408 
#> 0.087633353 0.040269394 0.074895975 0.025910296 0.080885022 0.031386464 
#>         409         410         411         412         413         414 
#> 0.023791095 0.069054266 0.065480287 0.023791095 0.052840747 0.038124197 
#>         415         416         417         418         419         420 
#> 0.038124197 0.036088980 0.131577971 0.107649422 0.044911488 0.067373279 
#>         421         422         423         424         425         426 
#> 0.052840747 0.092297050 0.080885022 0.076749210 0.076749210 0.131577971 
#>         427         428 
#> 0.040269394 0.071042725 
#> 
#> $se.fit
#>          1          2          3          4          5          6          7 
#> 0.15984908 0.04699153 0.03001251 0.04065602 0.02797819 0.02718358 0.02445754 
#>          8          9         10         11         12         13         14 
#> 0.03260362 0.04065602 0.03520904 0.03526794 0.03180879 0.03180879 0.02347583 
#>         15         16         17         18         19         20         21 
#> 0.03282705 0.03783163 0.04092455 0.01954432 0.03260362 0.01906271 0.05953477 
#>         22         23         24         25         26         27         28 
#> 0.02493291 0.05539944 0.04015325 0.04015325 0.04015325 0.03250006 0.03180879 
#>         29         30         31         32         33         34         35 
#> 0.02347583 0.15984908 0.04015325 0.03526794 0.02445754 0.04699153 0.04699153 
#>         36         37         38         39         40         41         42 
#> 0.03520904 0.02797819 0.02322033 0.02146313 0.03500625 0.04092455 0.02793108 
#>         43         44         45         46         47         48         49 
#> 0.03257234 0.03526776 0.10203645 0.04065602 0.02797819 0.02493291 0.05539944 
#>         50         51         52         53         54         55         56 
#> 0.02469664 0.02322033 0.03526794 0.02280751 0.02146313 0.04065602 0.07639468 
#>         57         58         59         60         61         62         63 
#> 0.02493291 0.02493291 0.02322033 0.02322033 0.02044203 0.02012444 0.03520904 
#>         64         65         66         67         68         69         70 
#> 0.02493291 0.02718358 0.02265100 0.02146313 0.02044203 0.07639468 0.02469664 
#>         71         72         73         74         75         76         77 
#> 0.02469664 0.03250006 0.02146313 0.04695296 0.04015325 0.02265100 0.03180879 
#>         78         79         80         81         82         83         84 
#> 0.03282705 0.02559575 0.04092455 0.05953477 0.06150545 0.03069993 0.03526794 
#>         85         86         87         88         89         90         91 
#> 0.03250006 0.02559575 0.04065602 0.04980428 0.03243272 0.04065602 0.03520904 
#>         92         93         94         95         96         97         98 
#> 0.03069993 0.02718358 0.02265100 0.02265100 0.02280751 0.03282705 0.03500625 
#>         99        100        101        102        103        104        105 
#> 0.02677849 0.02677849 0.04403970 0.05815646 0.03001251 0.02718358 0.04015325 
#>        106        107        108        109        110        111        112 
#> 0.02146313 0.03180879 0.02445754 0.03783163 0.02900627 0.01954432 0.01819911 
#>        113        114        115        116        117        118        119 
#> 0.02718358 0.15984908 0.02347583 0.03500625 0.03783163 0.05418085 0.03250006 
#>        120        121        122        123        124        125        126 
#> 0.02347583 0.02012444 0.08867502 0.02629712 0.06526695 0.02718358 0.02280751 
#>        127        128        129        130        131        132        133 
#> 0.12997937 0.03282705 0.03282705 0.02445754 0.04403970 0.05418085 0.04699153 
#>        134        135        136        137        138        139        140 
#> 0.03520904 0.02718358 0.02280751 0.03180879 0.02117362 0.02559575 0.02044203 
#>        141        142        143        144        145        146        147 
#> 0.01850452 0.05539944 0.02059614 0.02044203 0.02718358 0.05539944 0.03526794 
#>        148        149        150        151        152        153        154 
#> 0.02265100 0.13840242 0.03250006 0.11789285 0.20273748 0.04065602 0.02493291 
#>        155        156        157        158        159        160        161 
#> 0.02322033 0.02265100 0.03250006 0.03180879 0.03180879 0.04092455 0.04092455 
#>        162        163        164        165        166        167        168 
#> 0.04702832 0.02012444 0.05232060 0.05232060 0.05649921 0.03243272 0.03520904 
#>        169        170        171        172        173        174        175 
#> 0.02493291 0.02322033 0.03282705 0.02677849 0.02075978 0.02075978 0.04092455 
#>        176        177        178        179        180        181        182 
#> 0.11091255 0.11091255 0.04699153 0.14838308 0.03250006 0.03282705 0.02559575 
#>        183        184        185        186        187        188        189 
#> 0.11413926 0.02559575 0.11175584 0.01954432 0.03243272 0.02299568 0.03526794 
#>        190        191        192        193        194        195        196 
#> 0.02075978 0.03520904 0.02493291 0.02718358 0.04695296 0.04695296 0.02322033 
#>        197        198        199        200        201        202        203 
#> 0.02280751 0.02445754 0.02559575 0.02094646 0.03500625 0.03783163 0.02900627 
#>        204        205        206        207        208        209        210 
#> 0.02997587 0.04980428 0.03330039 0.03069993 0.04695296 0.02234057 0.02347583 
#>        211        212        213        214        215        216        217 
#> 0.02146313 0.02677849 0.04403970 0.02797819 0.03520904 0.02493291 0.02044203 
#>        218        219        220        221        222        223        224 
#> 0.02718358 0.04695296 0.02677849 0.06526695 0.02445754 0.03069993 0.02718358 
#>        225        226        227        228        229        230        231 
#> 0.02718358 0.02493291 0.02469664 0.14838308 0.02044203 0.02900627 0.03294827 
#>        232        233        234        235        236        237        238 
#> 0.03294827 0.01879279 0.04699153 0.04699153 0.02299568 0.11789285 0.11413926 
#>        239        240        241        242        243        244        245 
#> 0.02012444 0.04980428 0.05232060 0.04065602 0.04065602 0.04015325 0.02322033 
#>        246        247        248        249        250        251        252 
#> 0.04702832 0.02280751 0.03082455 0.03069993 0.02718358 0.04695296 0.02469664 
#>        253        254        255        256        257        258        259 
#> 0.13840242 0.02280751 0.05232060 0.17270293 0.05539944 0.02469664 0.02299568 
#>        260        261        262        263        264        265        266 
#> 0.02322033 0.04015325 0.12997937 0.02075978 0.02059614 0.02044203 0.03520904 
#>        267        268        269        270        271        272        273 
#> 0.02493291 0.03250006 0.02347583 0.02322033 0.02445754 0.03082455 0.03069993 
#>        274        275        276        277        278        279        280 
#> 0.03250006 0.04980428 0.05539944 0.02493291 0.02299568 0.02280751 0.01994838 
#>        281        282        283        284        285        286        287 
#> 0.01931338 0.01719483 0.02718358 0.01975567 0.04695296 0.04015325 0.03526794 
#>        288        289        290        291        292        293        294 
#> 0.02094646 0.05649921 0.02146313 0.02075978 0.02718358 0.02322033 0.04015325 
#>        295        296        297        298        299        300        301 
#> 0.02559575 0.03500625 0.04092455 0.04702832 0.02629712 0.02280751 0.02445754 
#>        302        303        304        305        306        307        308 
#> 0.03082455 0.03243272 0.07639468 0.02629712 0.02322033 0.03500625 0.02677849 
#>        309        310        311        312        313        314        315 
#> 0.04065602 0.05539944 0.02384613 0.02469664 0.02299568 0.02234057 0.02265100 
#>        316        317        318        319        320        321        322 
#> 0.02445754 0.11175584 0.01994838 0.11696804 0.02718358 0.04015325 0.02322033 
#>        323        324        325        326        327        328        329 
#> 0.02280751 0.11175584 0.03001251 0.07639468 0.02629712 0.02322033 0.02322033 
#>        330        331        332        333        334        335        336 
#> 0.02184164 0.03250006 0.02117362 0.02900627 0.01954432 0.03001251 0.03520904 
#>        337        338        339        340        341        342        343 
#> 0.02718358 0.04015325 0.02234057 0.03526794 0.13840242 0.03500625 0.01609214 
#>        344        345        346        347        348        349        350 
#> 0.04699153 0.10203645 0.04065602 0.02797819 0.02797819 0.03520904 0.02493291 
#>        351        352        353        354        355        356        357 
#> 0.02718358 0.02184164 0.02184164 0.02146313 0.03282705 0.02075978 0.01879279 
#>        358        359        360        361        362        363        364 
#> 0.04065602 0.03520904 0.03520904 0.02280751 0.07639468 0.03069993 0.02718358 
#>        365        366        367        368        369        370        371 
#> 0.02469664 0.04015325 0.03526794 0.02265100 0.13840242 0.02280751 0.02280751 
#>        372        373        374        375        376        377        378 
#> 0.02900627 0.03069993 0.02347583 0.03069993 0.02493291 0.02718358 0.05539944 
#>        379        380        381        382        383        384        385 
#> 0.02280751 0.03250006 0.02347583 0.02445754 0.03282705 0.02075978 0.03257234 
#>        386        387        388        389        390        391        392 
#> 0.03520904 0.03520904 0.04695296 0.02384613 0.04015325 0.04015325 0.02265100 
#>        393        394        395        396        397        398        399 
#> 0.02280751 0.02184164 0.03250006 0.02280751 0.02146313 0.02347583 0.02347583 
#>        400        401        402        403        404        405        406 
#> 0.02677849 0.02793108 0.02793108 0.04403970 0.02900627 0.05232060 0.03330039 
#>        407        408        409        410        411        412        413 
#> 0.04065602 0.02493291 0.02146313 0.02718358 0.02469664 0.02146313 0.02347583 
#>        414        415        416        417        418        419        420 
#> 0.02997587 0.02997587 0.03082455 0.04015325 0.03282705 0.02677849 0.05649921 
#>        421        422        423        424        425        426        427 
#> 0.02347583 0.04092455 0.04065602 0.03520904 0.03520904 0.04015325 0.02900627 
#>        428 
#> 0.05455469 
#> 
#> $lower
#>            1            2            3            4            5            6 
#> 0.0878312490 0.0334117474 0.0095339815 0.0345592500 0.0092255221 0.0356796393 
#>            7            8            9           10           11           12 
#> 0.0221154851 0.0051466247 0.0345592500 0.0354113349 0.0777227238 0.0705011636 
#>           13           14           15           16           17           18 
#> 0.0705011636 0.0251427675 0.0643396161 0.0502985492 0.0435468933 0.0012532115 
#>           19           20           21           22           23           24 
#> 0.0051466247 0.0009018518 0.0113958953 0.0083387001 0.0753340755 0.0783422947 
#>           25           26           27           28           29           30 
#> 0.0783422947 0.0783422947 0.0751197167 0.0705011636 0.0251427675 0.0878312490 
#>           31           32           33           34           35           36 
#> 0.0783422947 0.0777227238 0.0221154851 0.0334117474 0.0334117474 0.0354113349 
#>           37           38           39           40           41           42 
#> 0.0092255221 0.0332067821 0.0053015852 0.0573800574 0.0435468933 0.0141706812 
#>           43           44           45           46           47           48 
#> 0.0012318059 0.0098577425 0.0627569082 0.0345592500 0.0092255221 0.0083387001 
#>           49           50           51           52           53           54 
#> 0.0753340755 0.0348198331 0.0332067821 0.0777227238 0.0281497924 0.0053015852 
#>           55           56           57           58           59           60 
#> 0.0345592500 0.0694779275 0.0083387001 0.0083387001 0.0332067821 0.0332067821 
#>           61           62           63           64           65           66 
#> 0.0027277942 0.0020175921 0.0354113349 0.0083387001 0.0356796393 0.0309186620 
#>           67           68           69           70           71           72 
#> 0.0053015852 0.0027277942 0.0694779275 0.0348198331 0.0348198331 0.0751197167 
#>           73           74           75           76           77           78 
#> 0.0053015852 0.0773828404 0.0783422947 0.0309186620 0.0705011636 0.0643396161 
#>           79           80           81           82           83           84 
#> 0.0192241931 0.0435468933 0.0113958953 0.0079808664 0.0358343411 0.0777227238 
#>           85           86           87           88           89           90 
#> 0.0751197167 0.0192241931 0.0345592500 0.0270268355 0.0097443846 0.0345592500 
#>           91           92           93           94           95           96 
#> 0.0354113349 0.0358343411 0.0356796393 0.0309186620 0.0309186620 0.0281497924 
#>           97           98           99          100          101          102 
#> 0.0643396161 0.0573800574 0.0165615246 0.0165615246 0.0373668001 0.0135946892 
#>          103          104          105          106          107          108 
#> 0.0095339815 0.0356796393 0.0783422947 0.0053015852 0.0705011636 0.0221154851 
#>          109          110          111          112          113          114 
#> 0.0502985492 0.0120618678 0.0012532115 0.0005438439 0.0356796393 0.0878312490 
#>          115          116          117          118          119          120 
#> 0.0251427675 0.0573800574 0.0502985492 0.0320748546 0.0751197167 0.0251427675 
#>          121          122          123          124          125          126 
#> 0.0020175921 0.0661538155 0.0088236041 0.0726025803 0.0356796393 0.0281497924 
#>          127          128          129          130          131          132 
#> 0.0864773115 0.0643396161 0.0643396161 0.0221154851 0.0373668001 0.0320748546 
#>          133          134          135          136          137          138 
#> 0.0334117474 0.0354113349 0.0356796393 0.0281497924 0.0705011636 0.0047039242 
#>          139          140          141          142          143          144 
#> 0.0192241931 0.0027277942 0.0006446052 0.0753340755 0.0031519667 0.0027277942 
#>          145          146          147          148          149          150 
#> 0.0356796393 0.0753340755 0.0777227238 0.0309186620 0.0878152564 0.0751197167 
#>          151          152          153          154          155          156 
#> 0.0808846350 0.0438836411 0.0345592500 0.0083387001 0.0332067821 0.0309186620 
#>          157          158          159          160          161          162 
#> 0.0751197167 0.0705011636 0.0705011636 0.0435468933 0.0435468933 0.0318560091 
#>          163          164          165          166          167          168 
#> 0.0020175921 0.0228455175 0.0228455175 0.0161941766 0.0097443846 0.0354113349 
#>          169          170          171          172          173          174 
#> 0.0083387001 0.0332067821 0.0643396161 0.0165615246 0.0036238431 0.0036238431 
#>          175          176          177          178          179          180 
#> 0.0435468933 0.0545364406 0.0545364406 0.0334117474 0.0882346832 0.0751197167 
#>          181          182          183          184          185          186 
#> 0.0643396161 0.0192241931 0.0767079538 0.0192241931 0.0717786833 0.0012532115 
#>          187          188          189          190          191          192 
#> 0.0097443846 0.0071862815 0.0777227238 0.0036238431 0.0354113349 0.0083387001 
#>          193          194          195          196          197          198 
#> 0.0356796393 0.0773828404 0.0773828404 0.0332067821 0.0281497924 0.0221154851 
#>          199          200          201          202          203          204 
#> 0.0192241931 0.0041423791 0.0573800574 0.0502985492 0.0120618678 0.0102254902 
#>          205          206          207          208          209          210 
#> 0.0270268355 0.0030271849 0.0358343411 0.0773828404 0.0065589500 0.0251427675 
#>          211          212          213          214          215          216 
#> 0.0053015852 0.0165615246 0.0373668001 0.0092255221 0.0354113349 0.0083387001 
#>          217          218          219          220          221          222 
#> 0.0027277942 0.0356796393 0.0773828404 0.0165615246 0.0726025803 0.0221154851 
#>          223          224          225          226          227          228 
#> 0.0358343411 0.0356796393 0.0356796393 0.0083387001 0.0348198331 0.0882346832 
#>          229          230          231          232          233          234 
#> 0.0027277942 0.0120618678 0.0043167034 0.0043167034 0.0007630250 0.0334117474 
#>          235          236          237          238          239          240 
#> 0.0334117474 0.0071862815 0.0808846350 0.0767079538 0.0020175921 0.0270268355 
#>          241          242          243          244          245          246 
#> 0.0228455175 0.0345592500 0.0345592500 0.0783422947 0.0332067821 0.0318560091 
#>          247          248          249          250          251          252 
#> 0.0281497924 0.0086412928 0.0358343411 0.0356796393 0.0773828404 0.0348198331 
#>          253          254          255          256          257          258 
#> 0.0878152564 0.0281497924 0.0228455175 0.0867221628 0.0753340755 0.0348198331 
#>          259          260          261          262          263          264 
#> 0.0071862815 0.0332067821 0.0783422947 0.0864773115 0.0036238431 0.0031519667 
#>          265          266          267          268          269          270 
#> 0.0027277942 0.0354113349 0.0083387001 0.0751197167 0.0251427675 0.0332067821 
#>          271          272          273          274          275          276 
#> 0.0221154851 0.0086412928 0.0358343411 0.0751197167 0.0270268355 0.0753340755 
#>          277          278          279          280          281          282 
#> 0.0083387001 0.0071862815 0.0281497924 0.0017261954 0.0010641357 0.0003244630 
#>          283          284          285          286          287          288 
#> 0.0356796393 0.0014726539 0.0773828404 0.0783422947 0.0777227238 0.0041423791 
#>          289          290          291          292          293          294 
#> 0.0161941766 0.0053015852 0.0036238431 0.0356796393 0.0332067821 0.0783422947 
#>          295          296          297          298          299          300 
#> 0.0192241931 0.0573800574 0.0435468933 0.0318560091 0.0088236041 0.0281497924 
#>          301          302          303          304          305          306 
#> 0.0221154851 0.0086412928 0.0097443846 0.0694779275 0.0088236041 0.0332067821 
#>          307          308          309          310          311          312 
#> 0.0573800574 0.0165615246 0.0345592500 0.0753340755 0.0077864859 0.0348198331 
#>          313          314          315          316          317          318 
#> 0.0071862815 0.0065589500 0.0309186620 0.0221154851 0.0717786833 0.0017261954 
#>          319          320          321          322          323          324 
#> 0.0329387188 0.0356796393 0.0783422947 0.0332067821 0.0281497924 0.0717786833 
#>          325          326          327          328          329          330 
#> 0.0095339815 0.0694779275 0.0088236041 0.0332067821 0.0332067821 0.0059247490 
#>          331          332          333          334          335          336 
#> 0.0751197167 0.0047039242 0.0120618678 0.0012532115 0.0095339815 0.0354113349 
#>          337          338          339          340          341          342 
#> 0.0356796393 0.0783422947 0.0065589500 0.0777227238 0.0878152564 0.0573800574 
#>          343          344          345          346          347          348 
#> 0.0001921035 0.0334117474 0.0627569082 0.0345592500 0.0092255221 0.0092255221 
#>          349          350          351          352          353          354 
#> 0.0354113349 0.0083387001 0.0356796393 0.0059247490 0.0059247490 0.0053015852 
#>          355          356          357          358          359          360 
#> 0.0643396161 0.0036238431 0.0007630250 0.0345592500 0.0354113349 0.0354113349 
#>          361          362          363          364          365          366 
#> 0.0281497924 0.0694779275 0.0358343411 0.0356796393 0.0348198331 0.0783422947 
#>          367          368          369          370          371          372 
#> 0.0777227238 0.0309186620 0.0878152564 0.0281497924 0.0281497924 0.0120618678 
#>          373          374          375          376          377          378 
#> 0.0358343411 0.0251427675 0.0358343411 0.0083387001 0.0356796393 0.0753340755 
#>          379          380          381          382          383          384 
#> 0.0281497924 0.0751197167 0.0251427675 0.0221154851 0.0643396161 0.0036238431 
#>          385          386          387          388          389          390 
#> 0.0012318059 0.0354113349 0.0354113349 0.0773828404 0.0077864859 0.0783422947 
#>          391          392          393          394          395          396 
#> 0.0783422947 0.0309186620 0.0281497924 0.0059247490 0.0751197167 0.0281497924 
#>          397          398          399          400          401          402 
#> 0.0053015852 0.0251427675 0.0251427675 0.0165615246 0.0141706812 0.0141706812 
#>          403          404          405          406          407          408 
#> 0.0373668001 0.0120618678 0.0228455175 0.0030271849 0.0345592500 0.0083387001 
#>          409          410          411          412          413          414 
#> 0.0053015852 0.0356796393 0.0348198331 0.0053015852 0.0251427675 0.0102254902 
#>          415          416          417          418          419          420 
#> 0.0102254902 0.0086412928 0.0783422947 0.0643396161 0.0165615246 0.0161941766 
#>          421          422          423          424          425          426 
#> 0.0251427675 0.0435468933 0.0345592500 0.0354113349 0.0354113349 0.0783422947 
#>          427          428 
#> 0.0120618678 0.0192560889 
#> 
#> $upper
#>         1         2         3         4         5         6         7         8 
#> 0.5833554 0.2006961 0.1331011 0.1778685 0.1240566 0.1294564 0.1093687 0.1614309 
#>         9        10        11        12        13        14        15        16 
#> 0.1778685 0.1584181 0.1955396 0.1769730 0.1769730 0.1076812 0.1746705 0.1795075 
#>        17        18        19        20        21        22        23        24 
#> 0.1850627 0.1315362 0.1614309 0.1435665 0.2649547 0.1110063 0.2613902 0.2126425 
#>        25        26        27        28        29        30        31        32 
#> 0.2126425 0.2126425 0.1837003 0.1769730 0.1076812 0.5833554 0.2126425 0.1955396 
#>        33        34        35        36        37        38        39        40 
#> 0.1093687 0.2006961 0.2006961 0.1584181 0.1240566 0.1131173 0.1002638 0.1758179 
#>        41        42        43        44        45        46        47        48 
#> 0.1850627 0.1206952 0.2450041 0.1571439 0.4044929 0.1778685 0.1240566 0.1110063 
#>        49        50        51        52        53        54        55        56 
#> 0.2613902 0.1197880 0.1131173 0.1955396 0.1074788 0.1002638 0.1778685 0.3265969 
#>        57        58        59        60        61        62        63        64 
#> 0.1110063 0.1110063 0.1131173 0.1131173 0.1093934 0.1168552 0.1584181 0.1110063 
#>        65        66        67        68        69        70        71        72 
#> 0.1294564 0.1091474 0.1002638 0.1093934 0.3265969 0.1197880 0.1197880 0.1837003 
#>        73        74        75        76        77        78        79        80 
#> 0.1002638 0.2347365 0.2126425 0.1091474 0.1769730 0.1746705 0.1122366 0.1850627 
#>        81        82        83        84        85        86        87        88 
#> 0.2649547 0.2913947 0.1423009 0.1955396 0.1837003 0.1122366 0.1778685 0.2091370 
#>        89        90        91        92        93        94        95        96 
#> 0.1440682 0.1778685 0.1584181 0.1423009 0.1294564 0.1091474 0.1091474 0.1074788 
#>        97        98        99       100       101       102       103       104 
#> 0.1746705 0.1758179 0.1160635 0.1160635 0.1920304 0.2525467 0.1331011 0.1294564 
#>       105       106       107       108       109       110       111       112 
#> 0.2126425 0.1002638 0.1769730 0.1093687 0.1795075 0.1260273 0.1315362 0.1649728 
#>       113       114       115       116       117       118       119       120 
#> 0.1294564 0.5833554 0.1076812 0.1758179 0.1795075 0.2269206 0.1837003 0.1076812 
#>       121       122       123       124       125       126       127       128 
#> 0.1168552 0.3642246 0.1167508 0.2921586 0.1294564 0.1074788 0.5029325 0.1746705 
#>       129       130       131       132       133       134       135       136 
#> 0.1746705 0.1093687 0.1920304 0.2269206 0.2006961 0.1584181 0.1294564 0.1074788 
#>       137       138       139       140       141       142       143       144 
#> 0.1769730 0.1008298 0.1122366 0.1093934 0.1573857 0.2613902 0.1064087 0.1093934 
#>       145       146       147       148       149       150       151       152 
#> 0.1294564 0.2613902 0.1955396 0.1091474 0.5271886 0.1837003 0.4642587 0.6625297 
#>       153       154       155       156       157       158       159       160 
#> 0.1778685 0.1110063 0.1131173 0.1091474 0.1837003 0.1769730 0.1769730 0.1850627 
#>       161       162       163       164       165       166       167       168 
#> 0.1850627 0.2001181 0.1168552 0.2189626 0.2189626 0.2407200 0.1440682 0.1584181 
#>       169       170       171       172       173       174       175       176 
#> 0.1110063 0.1131173 0.1746705 0.1160635 0.1039587 0.1039587 0.1850627 0.4278708 
#>       177       178       179       180       181       182       183       184 
#> 0.4278708 0.2006961 0.5541832 0.1837003 0.1746705 0.1122366 0.4502632 0.1122366 
#>       185       186       187       188       189       190       191       192 
#> 0.4398390 0.1315362 0.1440682 0.1035387 0.1955396 0.1039587 0.1584181 0.1110063 
#>       193       194       195       196       197       198       199       200 
#> 0.1294564 0.2347365 0.2347365 0.1131173 0.1074788 0.1093687 0.1122366 0.1020824 
#>       201       202       203       204       205       206       207       208 
#> 0.1758179 0.1795075 0.1260273 0.1319898 0.2091370 0.1889824 0.1423009 0.2347365 
#>       209       210       211       212       213       214       215       216 
#> 0.1015167 0.1076812 0.1002638 0.1160635 0.1920304 0.1240566 0.1584181 0.1110063 
#>       217       218       219       220       221       222       223       224 
#> 0.1093934 0.1294564 0.2347365 0.1160635 0.2921586 0.1093687 0.1423009 0.1294564 
#>       225       226       227       228       229       230       231       232 
#> 0.1294564 0.1110063 0.1197880 0.5541832 0.1093934 0.1260273 0.1701044 0.1701044 
#>       233       234       235       236       237       238       239       240 
#> 0.1502514 0.2006961 0.2006961 0.1035387 0.4642587 0.4502632 0.1168552 0.2091370 
#>       241       242       243       244       245       246       247       248 
#> 0.2189626 0.1778685 0.1778685 0.2126425 0.1131173 0.2001181 0.1074788 0.1385365 
#>       249       250       251       252       253       254       255       256 
#> 0.1423009 0.1294564 0.2347365 0.1197880 0.5271886 0.1074788 0.2189626 0.6140661 
#>       257       258       259       260       261       262       263       264 
#> 0.2613902 0.1197880 0.1035387 0.1131173 0.2126425 0.5029325 0.1039587 0.1064087 
#>       265       266       267       268       269       270       271       272 
#> 0.1093934 0.1584181 0.1110063 0.1837003 0.1076812 0.1131173 0.1093687 0.1385365 
#>       273       274       275       276       277       278       279       280 
#> 0.1423009 0.1837003 0.2091370 0.2613902 0.1110063 0.1035387 0.1074788 0.1212947 
#>       281       282       283       284       285       286       287       288 
#> 0.1373284 0.1904906 0.1294564 0.1261905 0.2347365 0.2126425 0.1955396 0.1020824 
#>       289       290       291       292       293       294       295       296 
#> 0.2407200 0.1002638 0.1039587 0.1294564 0.1131173 0.2126425 0.1122366 0.1758179 
#>       297       298       299       300       301       302       303       304 
#> 0.1850627 0.2001181 0.1167508 0.1074788 0.1093687 0.1385365 0.1440682 0.3265969 
#>       305       306       307       308       309       310       311       312 
#> 0.1167508 0.1131173 0.1758179 0.1160635 0.1778685 0.2613902 0.1066544 0.1197880 
#>       313       314       315       316       317       318       319       320 
#> 0.1035387 0.1015167 0.1091474 0.1093687 0.4398390 0.1212947 0.4453347 0.1294564 
#>       321       322       323       324       325       326       327       328 
#> 0.2126425 0.1131173 0.1074788 0.4398390 0.1331011 0.3265969 0.1167508 0.1131173 
#>       329       330       331       332       333       334       335       336 
#> 0.1131173 0.1004617 0.1837003 0.1008298 0.1260273 0.1315362 0.1331011 0.1584181 
#>       337       338       339       340       341       342       343       344 
#> 0.1294564 0.2126425 0.1015167 0.1955396 0.5271886 0.1758179 0.2202162 0.2006961 
#>       345       346       347       348       349       350       351       352 
#> 0.4044929 0.1778685 0.1240566 0.1240566 0.1584181 0.1110063 0.1294564 0.1004617 
#>       353       354       355       356       357       358       359       360 
#> 0.1004617 0.1002638 0.1746705 0.1039587 0.1502514 0.1778685 0.1584181 0.1584181 
#>       361       362       363       364       365       366       367       368 
#> 0.1074788 0.3265969 0.1423009 0.1294564 0.1197880 0.2126425 0.1955396 0.1091474 
#>       369       370       371       372       373       374       375       376 
#> 0.5271886 0.1074788 0.1074788 0.1260273 0.1423009 0.1076812 0.1423009 0.1110063 
#>       377       378       379       380       381       382       383       384 
#> 0.1294564 0.2613902 0.1074788 0.1837003 0.1076812 0.1093687 0.1746705 0.1039587 
#>       385       386       387       388       389       390       391       392 
#> 0.2450041 0.1584181 0.1584181 0.2347365 0.1066544 0.2126425 0.2126425 0.1091474 
#>       393       394       395       396       397       398       399       400 
#> 0.1074788 0.1004617 0.1837003 0.1074788 0.1002638 0.1076812 0.1076812 0.1160635 
#>       401       402       403       404       405       406       407       408 
#> 0.1206952 0.1206952 0.1920304 0.1260273 0.2189626 0.1889824 0.1778685 0.1110063 
#>       409       410       411       412       413       414       415       416 
#> 0.1002638 0.1294564 0.1197880 0.1002638 0.1076812 0.1319898 0.1319898 0.1385365 
#>       417       418       419       420       421       422       423       424 
#> 0.2126425 0.1746705 0.1160635 0.2407200 0.1076812 0.1850627 0.1778685 0.1584181 
#>       425       426       427       428 
#> 0.1584181 0.2126425 0.1260273 0.2295103 
#>