Iterate Group Size for a One-Parameter Group Testing Problem

Increasing of group size s for a fixed number of groups n in a binomial group testing design, until a pre-specified power is achieved. Control of bias during iteration. A hypothetical threshold proportion p.hyp and the absolute difference delta to be detected have to be specified.

sDesign(n, smax, delta, p.hyp, conf.level = 0.95,
 power = 0.8, alternative = "two.sided", method = "CP", biasrest = 0.05)

Arguments

n

integer, fixed sample size (number of assays)

smax

integer, maximal group size allowed in planning of the design

delta

absolute difference between the threshold and the true proportion which shall be detectable with specified power

p.hyp

threshold proportion to test against in the hypothesis, specify as a value between 0 and 1

conf.level

confidence level of the decision, default is 0.95

power

level of power to be achieved to be specified as a probability between 0 and 1

alternative

character string, defining the alternative hypothesis, either 'two.sided', 'less' or 'greater' where 'less' calculates the probability that p.hyp is excluded by an upper confidence limit for a true proportion p.hyp - delta, 'greater' calculates the probability that p.hyp is excluded by a lower confidence limit for a true proportion of p.hyp + delta . 'two.sided' calculates min(power(p.hyp - delta, p.hyp + delta)) for a two.sided CI, thus can result in much lower power. Note that coverage probability and power are not necessarily symmetric for upper and lower bound of binomial CI.

method

character string specifying the CImethod to be used for evaluation, see argument method in bgtCI

biasrest

value between 0 and 1 specifying the absolute bias maximally allowed

Details

The power of a confidence interval here is defined as the probability that a confidence interval or limit excludes the threshold parameter (p.hyp) of the hypothesis. This function increases size of groups (number of units in a bulk sample) until a pre-specified power is reached. Since the power does not increase monotone with increasing s for binomial proportions but oscillates between local maxima and minima, the simple iteration given here will generally result in selecting those s, for which the given CI method shows a local minimum of coverage if the null hypothesis is true. Since the positive bias of the estimator in group testing increases with increasing group size, it is checked whether bias is smaller than a pre-specified value (bias.rest). If bias violates this restriction for a given combination n, s, delta, s will not be further increased the actual power of the last acceptable group size s is returned.

Value

A list containing:

sout

the group size necessary to met the conditions

powerout

the exact power for the specified parameters and the group size

biasout

the bias for the specified parameters and the iterated group size

and a number of values specified in the function call or produced in the iteration.

References

Schaarschmidt F (2007) Experimental design for one-sided confidence intervals or hypothesis tests in binomial group testing. Communications in Biometry and Crop Science 2 (1), 32-40. http://agrobiol.sggw.waw.pl/cbcs/

Swallow WH (1985) Group testing for estimating infection rates and probabilities of disease transmission. Phytopathology 75 (8), 882-889.

See also

Examples

## Assume that objective is to show that a proportion
## is smaller than 0.005 (i.e. 0.5%) with a 
## power of 0.80 (i.e. 80%) if the unknown proportion
## in the population is 0.003 (i.e. 0.3%), thus a 
## delta = 0.002 shall be detected. A 95-per-cent 
## Clopper-Pearson CI (corresponding to an exact test)
## shall be used. The maximal number of groups might 
## be 30 where the assay sensitivity is not limited 
## until groupsize s = 100.

sDesign(smax=100, n=30, delta=0.002, p.hyp=0.005,
 alternative="less", method="CP", power=0.8)
#> Power was not reached in the range of s = 2 , 100 
#> Maximal power was reached for s = 100 with power = 0.4662 
#> and bias = 5.957802e-05 
#> alternative hypothesis: true proportion is less than 0.005 
#> assumed true proportion = 0.003 

## One might accept to detect delta=0.004,
## i.e. accepting to reject H0: p>=0.005 with 
## power 80 per cent when the true proportion is 0.001:

sDesign(smax=100, n=30, delta=0.004, p.hyp=0.005,
 alternative="less", method="CP", power=0.8)
#> Power was reached without violating bias restriction 
#> for s = 44 with power = 0.8626 
#> and bias = 1.704e-05 
#> alternative hypothesis: true proportion is less than 0.005 
#> assumed true proportion = 0.001 

sDesign(smax=100, n=30, delta=0.004, p.hyp=0.005,
 alternative="less", method="Score", power=0.8)
#> Power was reached without violating bias restriction 
#> for s = 41 with power = 0.8819 
#> and bias = 1.698e-05 
#> alternative hypothesis: true proportion is less than 0.005 
#> assumed true proportion = 0.001