Summary Method for Group Testing Model (Simple Pooling) Fits

Produce a summary list for objects of class "gt" returned by gtreg.

# S3 method for gt
summary(object, ...)

Arguments

object

a fitted object of class "gt".

...

currently not used.

Details

print.summary.gt is the print function that formats the coefficients, standard errors, etc. and additionally gives 'significance stars' if signif.stars is TRUE. The coefficients component of the result gives the estimated coefficients and their estimated standard errors, together with their ratio. This third column is labeled z ratio using Wald tests. A fourth column gives the two-tailed p-value corresponding to the z ratio based on a Wald test. (It is possible that there are no residual degrees of freedom from which to estimate it. In that case the estimate is NaN.)

Value

summary.gt returns an object of class "summary.gt", a list with components

call

the component from object.

link

the component from object.

deviance

the component from object.

aic

the component from object.

df.residual

the component from object.

null.deviance

the component from object.

df.null

the component from object.

deviance.resid

the deviance residuals: see residuals.gt

coefficients

the matrix of coefficients, standard errors, z-values and p-values. Aliased coefficients are omitted.

counts

the component from object.

method

the component from object.

cov.mat

the estimated covariance matrix of the estimated coefficients.

Author

Boan Zhang

See also

Examples

## --- Continuing the Example from '?gtreg':

data(hivsurv)

fit1 <- gtreg(formula = groupres ~ AGE + EDUC., data = hivsurv,
  groupn = gnum, sens = 0.9, spec = 0.9, method = "Xie")
#> 
#>  Number of minutes running: 0 
#>  

summary(fit1)
#> 
#> Call:
#> gtreg(formula = groupres ~ AGE + EDUC., data = hivsurv, groupn = gnum, 
#>     sens = 0.9, spec = 0.9, method = "Xie")
#> 
#> Deviance Residuals: 
#>     Min       1Q   Median       3Q      Max  
#> -1.1729  -0.9406  -0.8281   1.3386   1.6497  
#> 
#> Coefficients:
#>             Estimate Std. Error z value Pr(>|z|)  
#> (Intercept) -3.11976    1.84816  -1.688   0.0914 .
#> AGE         -0.05692    0.07777  -0.732   0.4642  
#> EDUC.        0.82833    0.50717   1.633   0.1024  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#>     Null deviance: 154.3  on 85  degrees of freedom
#> Residual deviance: 109.5  on 83  degrees of freedom
#> AIC: 115.5
#> 
#> Number of iterations in EM: 43
#>